Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0856541282 -0.0282458232  0.0059014528  0.1612296433  0.0212870235 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -0.0571930538 -0.0688728487 -0.0257254405  0.0079739470 -0.0768037550 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.1517724702  0.0986401766  0.0224800426  0.0224008967 -0.0065489848 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  0.0364533877  0.0110455309  0.0290139812  0.0602774141 -0.0044784104 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -0.0066695842  0.0688443916 -0.0362829582  0.0691883623 -0.0393869483 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.0306002358  0.0229356613 -0.0819306683  0.1651549155  0.0090878465 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0033990885 -0.0752326329  0.0004686981  0.0055265124  0.0298729255 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.0845818404 -0.0261584796 -0.0911624830 -0.0267976077 -0.2122473296 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.0253418825  0.0003668032  0.0126998608  0.0913804226  0.0108073309 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  0.0625776070  0.0150425014  0.0084742504  0.0013065863  0.0781675848 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0382189681  0.0970425519 -0.0166821335  0.0356019605  0.1485612171 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.0102169826  0.0384133200 -0.1010547887  0.0320877615  0.0106611990 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0277882237 -0.0588467441  0.1572564979 -0.1561784425  0.0393500533 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0681660886  0.0207031082 -0.0942367920  0.1202567038 -0.0092417918 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -0.1329568002  0.0280052415  0.0327779002 -0.0160096209  0.0193100559 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0857407819  0.0325844091  0.0669097207  0.0277385234  0.1012716670 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0313283160  0.1569874030  0.0380099210 -0.0069232483  0.0194802951 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0084526285 -0.0217494759  0.0473383917  0.0041403429  0.0688831916 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0691881516 -0.0309992865  0.0488938151  0.0443501650 -0.0195082789 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0222498996 -0.0394362456 -0.1150263065 -0.0005246536 -0.0169341514 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.1501074726 -0.0313985703 -0.0430547222  0.0061573649 -0.1376049465 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -0.0086860634 -0.0420417801 -0.0235433134 -0.0828186175  0.0389524439 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.0665304028 -0.0306578006  0.1485894298  0.0573876447  0.0196766692 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  0.0228628607  0.0401829352  0.0306238173  0.0051630053 -0.0952641749 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.0328559135  0.0421804493  0.0134883342 -0.0722304112 -0.0032154511 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -0.0600695216  0.0145378469  0.1702970393 -0.0480618141  0.0931200840 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  0.0144863240  0.0320566506 -0.1239785540  0.0301841006 -0.1388792236 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  0.0530713545 -0.0157994178 -0.0616617059  0.0385979592 -0.0317973103 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0189543096 -0.1061722735  0.0202747064  0.1066297425  0.0244935856 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.0274614409 -0.0105515044  0.0178203742  0.0656188605  0.1607629447 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0073543613 -0.0842832149  0.0414909802  0.0234927009  0.0372372797 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.0537496482  0.0261909547 -0.0675536708 -0.1045014050 -0.1248846621 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  0.0081372890  0.1057107614 -0.0309551266  0.0569928322  0.1497956244 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -0.0865287204  0.0938306091  0.0683829738  0.0114810487 -0.0656828171 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0058709216 -0.0680911674  0.0488638952 -0.0194903805  0.0436752468 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  0.0555096673  0.0158690882  0.0533176586 -0.0404262386  0.0827040783 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -0.0185487262 -0.0085985278 -0.1367518577  0.0270975865  0.0121450452 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.0271344410 -0.0283872786  0.0334638082  0.1187472350 -0.0557723941 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.1427884028  0.0202691743 -0.0035992102 -0.0059375735 -0.0461755852 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0779844709 -0.0348039607 -0.0206980703 -0.0281395599  0.0870075051 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.0629824125 -0.0266193022 -0.0461516548 -0.0040848236 -0.0718739860 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  0.0518104939 -0.1416240523 -0.0080680759 -0.0397871933 -0.0328185610 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -0.0424765713  0.0471173569  0.0462795584  0.0469532431 -0.1561303388 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0143455102 -0.0391288055  0.0192841911  0.0379632116  0.0313335203 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.0814098060  0.0614714863  0.0040413902  0.0170784420  0.0085510442 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  0.0066907252  0.0099614003 -0.0708253878 -0.1024563403 -0.0254493918 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  0.1398055329  0.0355567319  0.0671078092  0.0717489166  0.0059005223 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -0.0861921215 -0.0798468021 -0.0469474845 -0.1663714362 -0.0131219519 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.0614855424  0.0318806349  0.0474411889  0.1017942749 -0.0939485222 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  0.0444639801  0.0011584914  0.0058521496 -0.0216690502  0.1190603907 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.1245867789 -0.0876414021  0.0994956330 -0.0128365041  0.0478221661 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.1324579547  0.0532560187 -0.0730065747  0.0946076034 -0.0122107528 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.0447234834 -0.0420314168  0.0300221407  0.0626452660 -0.0613733417 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  0.1489677702  0.1299859679  0.0377436888  0.2219335203  0.0193397344 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -0.0745364833  0.0310048651 -0.0845488786  0.0963105027  0.0995191148 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -0.0413982590  0.0175460786 -0.0561144790 -0.1174290887  0.0148156931 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -0.0615850935  0.0562167862  0.0497197306  0.0621864925  0.0161822317 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -0.0090555649  0.0311071058 -0.0347205156  0.0038594133 -0.0017498450 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0152667065  0.0502314779 -0.0627903979  0.0497059613  0.0432858030 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -0.0400382444 -0.0465932573 -0.0870961644  0.0360912627  0.0401019940 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -0.0267987709  0.0336030118  0.0001784414  0.1395963915 -0.0137639979 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0482275784  0.0412059885  0.1501252948  0.0997921980 -0.1195451908 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0888780189  0.0157205003 -0.0279648552  0.0476167050  0.0470205480 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0131475874 -0.0158350076  0.0411584284  0.0995649637 -0.0109957779 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  0.0218770021  0.0582459393  0.0680473460  0.0947158132 -0.1403101360 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -0.0558955403 -0.0827555779 -0.0628411151 -0.1323319077  0.0376140831 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0466602681 -0.1092521394  0.0065765796  0.0109767722  0.1061273479 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0649172679  0.0573182835  0.0788790013 -0.0590231824 -0.0529206697 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.1369480662  0.0050927221 -0.1518758147  0.0280674742 -0.0248992333 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.1390660840 -0.1221243168 -0.0473450919  0.0839896439 -0.0329071473 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0172787006  0.0386377002  0.0665913352 -0.0692187000  0.0013252535 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  0.1861088799 -0.0620175517  0.0881190900 -0.1016652722  0.0174818969 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.0159993542  0.0349538415  0.0580239020 -0.0279774939  0.0877765740 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.1377943121 -0.0557207288  0.0439155035 -0.0242057292 -0.0482799103 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0645833871 -0.0268496427 -0.0255471796  0.0664034900 -0.0831939251 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.0777870609  0.0637846825  0.0008221596  0.1070404627 -0.0200575355 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -0.0527352282  0.2155385288 -0.0764492402 -0.0417952136 -0.0753773137 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0154330935 -0.0689363156  0.0080487656  0.0026470357  0.0435785242 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -0.1333931892 -0.0441843592  0.0339633112  0.0291019221 -0.0530976542 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  0.0111349774 -0.0691654032  0.1464018175  0.0636567670  0.0301656082 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -0.0055650170 -0.0527159115  0.0127635895 -0.0969873840  0.0163999319 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0449499758  0.0435055610  0.0179475224  0.0341196733 -0.0502867053 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.0846215102 -0.0564492606  0.0153720903 -0.1304884377 -0.0322175746 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0139234083 -0.0207641624  0.0449617456 -0.0148414349 -0.0126247425 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  0.0515701162  0.0809657376  0.0692363107  0.0006593688  0.0480085287 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.0195584924 -0.1644283326 -0.1830700264  0.0533969259  0.0374555095 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.0535583804  0.0928616546 -0.0375891228  0.0907380199 -0.0083852031 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.0103884606  0.0203260212  0.0376519653 -0.1019316237 -0.0228504910 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0431578186 -0.0653448074 -0.1018565266 -0.0664610221 -0.0436878769 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0536697869  0.0420169693  0.0204149686