Protein Sequence Similarity Calculation based on Gene Ontology (GO) Similarity

Protein Sequence Similarity Calculation based on Gene Ontology (GO) Similarity

calcParProtGOSim(golist, type = c("go", "gene"), ont = "MF",
  organism = "human", measure = "Resnik", combine = "BMA")

Arguments

golist
A character vector, each component contains a character vector of GO terms or one Entrez Gene ID.

type
Input type of golist, 'go' for GO Terms, 'gene' for gene ID.

ont
Default is 'MF', could be one of 'MF', 'BP', or 'CC' subontologies.

organism
Default is 'human', could be one of 'anopheles', 'arabidopsis', 'bovine', 'canine', 'chicken', 'chimp', 'coelicolor', 'ecolik12', 'ecsakai', 'fly', 'human', 'malaria', 'mouse', 'pig', 'rat', 'rhesus', 'worm', 'xenopus', 'yeast' or 'zebrafish'.

measure
Default is 'Resnik', could be one of 'Resnik', 'Lin', 'Rel', 'Jiang' or 'Wang'.

combine
Default is 'BMA', could be one of 'max', 'average', 'rcmax' or 'BMA' for combining semantic similarity scores of multiple GO terms associated with protein.

Value

A n x n similarity matrix.

Details

This function calculates protein sequence similarity based on Gene Ontology (GO) similarity.

See also

See calcTwoProtGOSim for calculating the GO semantic similarity between two groups of GO terms or two Entrez gene IDs. See calcParProtSeqSim for paralleled protein similarity calculation based on sequence alignment.

Examples

# NOT RUN {
# by GO Terms
go1 = c('GO:0005215', 'GO:0005488', 'GO:0005515', 'GO:0005625', 'GO:0005802', 'GO:0005905')  # AP4B1
go2 = c('GO:0005515', 'GO:0005634', 'GO:0005681', 'GO:0008380', 'GO:0031202')  # BCAS2
go3 = c('GO:0003735', 'GO:0005622', 'GO:0005840', 'GO:0006412')  # PDE4DIP
glist = list(go1, go2, go3)
# }# NOT RUN {
gsimmat1 = calcParProtGOSim(glist, type = 'go', ont = 'CC')
print(gsimmat1)
# }# NOT RUN {
# by Entrez gene id
genelist = list(c('150', '151', '152', '1814', '1815', '1816'))
# }# NOT RUN {
gsimmat2 = calcParProtGOSim(genelist, type = 'gene')
print(gsimmat2)
# }