Calculate the Kier and Hall Chi Path Cluster Indices of Orders 4, 5 and 6

Calculate the Kier and Hall Chi Path Cluster Indices of Orders 4, 5 and 6

extractDrugChiPathCluster(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 6 columns named SPC.4, SPC.5, SPC.6, VPC.4, VPC.5, VPC.6:

SPC.4 - Simple path cluster, order 4

SPC.5 - Simple path cluster, order 5

SPC.6 - Simple path cluster, order 6

VPC.4 - Valence path cluster, order 4

VPC.5 - Valence path cluster, order 5

VPC.6 - Valence path cluster, order 6

Details

Evaluates chi path cluster descriptors. The code currently evluates the simple and valence chi chain descriptors of orders 4, 5 and 6. It utilizes the graph isomorphism code of the CDK to find fragments matching SMILES strings representing the fragments corresponding to each type of chain.

Note

These descriptors are calculated using graph isomorphism to identify the various fragments. As a result calculations may be slow. In addition, recent versions of Molconn-Z use simplified fragment definitions (i.e., rings without branches etc.) whereas these descriptors use the older more complex fragment definitions.

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugChiPathCluster(mol)
head(dat)# }

#>                                                             SPC.4    SPC.5
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 4.4708723 7.284328
#> c1(ccc(cc1)O)NC(=O)C                                    1.0504768 1.384471
#> c1ccccc1NC(=O)C                                         0.6422285 1.041888
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              2.4588951 3.016039
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                3.7782569 5.378380
#> CC(=O)NO                                                0.4082483 0.000000
#>                                                             SPC.6      VPC.4
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 10.643874 1.47443479
#> c1(ccc(cc1)O)NC(=O)C                                     1.685481 0.27498904
#> c1ccccc1NC(=O)C                                          1.093413 0.18892275
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                               3.282758 0.96800577
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                 6.011966 1.21157118
#> CC(=O)NO                                                 0.000000 0.04564355
#>                                                             VPC.5     VPC.6
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 2.0335609 2.5367704
#> c1(ccc(cc1)O)NC(=O)C                                    0.3094837 0.2911208
#> c1ccccc1NC(=O)C                                         0.2490484 0.2107132
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              0.9637957 0.8062145
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                1.3492987 1.2971379
#> CC(=O)NO                                                0.0000000 0.0000000