Calculate the Graph Molecular Fingerprints (in Compact Format)

Calculate the Graph Molecular Fingerprints (in Compact Format)

extractDrugGraph(molecules, depth = 6, size = 1024, silent = TRUE)

Arguments

molecules
Parsed molucule object.

depth
The search depth. Default is 6.

size
The length of the fingerprint bit string. Default is 1024.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A list, each component represents one of the molecules, each element in the component represents the index of which element in the fingerprint is 1. Each component's name is the length of the fingerprints.

Details

Calculate the graph molecular fingerprints. Similar to the standard type by simply considers connectivity. This is hashed fingerprints, with a default length of 1024.

See also

Author

Nan Xiao <https://nanx.me>

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
fp  = extractDrugGraph(mol)
head(fp)# }

#> $`1024`
#>  [1]   31   96  140  186  188  190  195  230  243  245  256  286  287  294  307
#> [16]  317  319  321  323  334  339  348  358  359  367  397  485  486  504  542
#> [31]  563  596  716  742  743  756  760  884  928  943  947  965  967  970 1003
#> 
#> $`1024`
#>  [1]  114  120  186  228  243  245  255  256  294  309  317  321  334  339  356
#> [16]  367  388  422  504  544  573  588  596  678  737  742  743  756  785  884
#> [31]  928  967  970 1003
#> 
#> $`1024`
#>  [1]  114  120  186  228  243  245  294  309  334  339  388  422  504  544  573
#> [16]  588  596  678  737  742  743  756  785  884  928  967  970 1003
#> 
#> $`1024`
#>  [1]   41   45  113  114  129  138  163  173  177  179  196  200  219  224  230
#> [16]  231  243  245  256  290  308  309  315  398  487  544  547  554  587  594
#> [31]  611  621  680  699  707  730  735  736  738  740  742  743  745  746  747
#> [46]  779  856  869  884  885  886  895  905  917  926  928  936  946  948  954
#> [61]  960  963  964  966  967  970  974  978  979  983 1002 1003 1004 1014 1017
#> 
#> $`1024`
#>  [1]   10   14   24   26   58  114  120  130  131  155  175  180  186  196  199
#> [16]  200  213  228  243  245  256  272  286  294  296  309  316  317  321  334
#> [31]  339  367  393  396  422  444  457  486  504  544  553  583  588  596  630
#> [46]  678  685  692  699  737  742  743  745  756  785  789  808  839  852  884
#> [61]  885  886  890  893  899  926  928  935  960  967  970  983 1003 1004 1007
#> [76] 1012 1015
#> 
#> $`1024`
#>  [1]  188  243  245  486  505  742  743  928  967  970 1003
#>