Descriptor that Calculates Kier and Hall Kappa Molecular Shape Indices

Descriptor that Calculates Kier and Hall Kappa Molecular Shape Indices

extractDrugKappaShapeIndices(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 3 columns named Kier1, Kier2 and Kier3:

Kier1 - First kappa shape index

Kier2 - Second kappa shape index

Kier3 - Third kappa shape index

Details

Kier and Hall Kappa molecular shape indices compare the molecular graph with minimal and maximal molecular graphs; see http://www.chemcomp.com/Journal_of_CCG/Features/descr.htm#KH for details: "they are intended to capture different aspects of molecular shape. Note that hydrogens are ignored. In the following description, n denotes the number of atoms in the hydrogen suppressed graph, m is the number of bonds in the hydrogen suppressed graph. Also, let p2 denote the number of paths of length 2 and let p3 denote the number of paths of length 3".

Author

Nan Xiao <https://nanx.me>

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugKappaShapeIndices(mol)
head(dat)# }

#>                                                             Kier1    Kier2
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 20.727041 9.000000
#> c1(ccc(cc1)O)NC(=O)C                                     9.090909 4.132653
#> c1ccccc1NC(=O)C                                          8.100000 4.000000
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              11.076923 4.022161
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                18.340265 8.203125
#> CC(=O)NO                                                 5.000000 2.250000
#>                                                            Kier3
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 4.716269
#> c1(ccc(cc1)O)NC(=O)C                                    3.265306
#> c1ccccc1NC(=O)C                                         3.111111
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              3.324100
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                5.551497
#> CC(=O)NO                                                4.000000