Calculate Molecular Distance Edge (MDE) Descriptors for C, N and O

Calculate Molecular Distance Edge (MDE) Descriptors for C, N and O

extractDrugMDE(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns one column named nAtomLAC.

Details

This descriptor calculates the 10 molecular distance edge (MDE) descriptor described in Liu, S., Cao, C., & Li, Z, and in addition it calculates variants where O and N are considered.

References

Liu, S., Cao, C., & Li, Z. (1998). Approach to estimation and prediction for normal boiling point (NBP) of alkanes based on a novel molecular distance-edge (MDE) vector, lambda. Journal of chemical information and computer sciences, 38(3), 387-394.

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')

mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugMDE(mol)
head(dat)

#>                                                         MDEC.11   MDEC.12
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0 1.5557226
#> c1(ccc(cc1)O)NC(=O)C                                          0 0.8944272
#> c1ccccc1NC(=O)C                                               0 1.0542115
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0 0.0000000
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0 1.3164802
#> CC(=O)NO                                                      0 0.0000000
#>                                                          MDEC.13 MDEC.14
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 2.176960       0
#> c1(ccc(cc1)O)NC(=O)C                                    1.144714       0
#> c1ccccc1NC(=O)C                                         1.154701       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              1.216440       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                1.313264       0
#> CC(=O)NO                                                1.000000       0
#>                                                          MDEC.22   MDEC.23
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 8.900408 22.978873
#> c1(ccc(cc1)O)NC(=O)C                                    3.301927  6.294690
#> c1ccccc1NC(=O)C                                         6.083643  4.039695
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              0.000000  0.000000
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                8.788908 11.415712
#> CC(=O)NO                                                0.000000  0.000000
#>                                                         MDEC.24    MDEC.33
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0 12.7788281
#> c1(ccc(cc1)O)NC(=O)C                                          0  0.9654894
#> c1ccccc1NC(=O)C                                               0  0.5000000
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0  1.1905508
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0  2.4906435
#> CC(=O)NO                                                      0  0.0000000
#>                                                         MDEC.34 MDEC.44
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0       0
#> c1(ccc(cc1)O)NC(=O)C                                          0       0
#> c1ccccc1NC(=O)C                                               0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0       0
#> CC(=O)NO                                                      0       0
#>                                                           MDEO.11   MDEO.12
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 1.5188513 0.9572486
#> c1(ccc(cc1)O)NC(=O)C                                    0.1428571 0.0000000
#> c1ccccc1NC(=O)C                                         0.0000000 0.0000000
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              0.6506976 0.0000000
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                1.2205432 0.0000000
#> CC(=O)NO                                                0.3333333 0.0000000
#>                                                         MDEO.22 MDEN.11
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0       0
#> c1(ccc(cc1)O)NC(=O)C                                          0       0
#> c1ccccc1NC(=O)C                                               0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0       0
#> CC(=O)NO                                                      0       0
#>                                                           MDEN.12 MDEN.13
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 0.0000000       0
#> c1(ccc(cc1)O)NC(=O)C                                    0.0000000       0
#> c1ccccc1NC(=O)C                                         0.0000000       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              0.7663094       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                0.0000000       0
#> CC(=O)NO                                                0.0000000       0
#>                                                          MDEN.22 MDEN.23
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 0.000000       0
#> c1(ccc(cc1)O)NC(=O)C                                    0.000000       0
#> c1ccccc1NC(=O)C                                         0.000000       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              1.650964       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                0.500000       0
#> CC(=O)NO                                                0.000000       0
#>                                                         MDEN.33
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0
#> c1(ccc(cc1)O)NC(=O)C                                          0
#> c1ccccc1NC(=O)C                                               0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0
#> CC(=O)NO                                                      0