Descriptor that Calculates the Principal Moments of Inertia and Ratios of the Principal Moments

Descriptor that Calculates the Principal Moments of Inertia and Ratios of the Principal Moments

Usage

extractDrugMomentOfInertia(molecules, silent = TRUE)

Arguments

molecules

Parsed molucule object.

silent

Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 7 columns named MOMI.X, MOMI.Y, MOMI.Z, MOMI.XY, MOMI.XZ, MOMI.YZ, MOMI.R:

MOMI.X - MI along X axis

MOMI.Y - MI along Y axis

MOMI.Z - MI along Z axis

MOMI.XY - X/Y

MOMI.XZ - X/Z

MOMI.YZ - Y/Z

MOMI.R - Radius of gyration

One important aspect of the algorithm is that if the eigenvalues of the MI tensor are below 1e-3, then the ratio's are set to a default of 1000.

Details

A descriptor that calculates the moment of inertia and radius of gyration. Moment of inertia (MI) values characterize the mass distribution of a molecule. Related to the MI values, ratios of the MI values along the three principal axes are also well know modeling variables. This descriptor calculates the MI values along the X, Y and Z axes as well as the ratio's X/Y, X/Z and Y/Z. Finally it also calculates the radius of gyration of the molecule.

Examples

sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')
# \donttest{
mol = readMolFromSDF(sdf)
#> Error in loadMolecules(normalizePath(sdffile)): The package "rcdk" is required to load molecular structures
dat = extractDrugMomentOfInertia(mol)
#> Error in evaluateDescriptor(molecules, type = "MomentOfInertiaDescriptor",     silent = silent): The package "rcdk" is required to compute molecular descriptors
head(dat)# }
#> Error in eval(expr, envir, enclos): object 'dat' not found