Descriptor that Calculates the Number Failures of the Lipinski's Rule Of Five

Descriptor that Calculates the Number Failures of the Lipinski's Rule Of Five

extractDrugRuleOfFive(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns one column named LipinskiFailures.

Details

This descriptor calculates the number failures of the Lipinski's Rule Of Five: http://en.wikipedia.org/wiki/Lipinski%27s_Rule_of_Five.

Author

Nan Xiao <https://nanx.me>

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugRuleOfFive(mol)
head(dat)# }

#>                                                         LipinskiFailures
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O                0
#> c1(ccc(cc1)O)NC(=O)C                                                   0
#> c1ccccc1NC(=O)C                                                        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                             0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                               0
#> CC(=O)NO                                                               0