Descriptor that Calculates the Vertex Adjacency Information of A Molecule

Descriptor that Calculates the Vertex Adjacency Information of A Molecule

extractDrugVAdjMa(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns one column named VAdjMat.

Details

Vertex adjacency information (magnitude): \(1 + \log_2^m\) where \(m\) is the number of heavy-heavy bonds. If \(m\) is zero, then 0 is returned.

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugVAdjMa(mol)
head(dat)# }

#>                                                          VAdjMat
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 5.807355
#> c1(ccc(cc1)O)NC(=O)C                                    4.459432
#> c1ccccc1NC(=O)C                                         4.321928
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              4.700440
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                5.523562
#> CC(=O)NO                                                3.000000