CTD Descriptors - Transition

CTD Descriptors - Transition

Usage

extractProtCTDT(x)

Arguments

x

A character vector, as the input protein sequence.

Value

A length 21 named vector

Details

This function calculates the Transition descriptor of the CTD descriptors (Dim: 21).

References

Inna Dubchak, Ilya Muchink, Stephen R. Holbrook and Sung-Hou Kim. Prediction of protein folding class using global description of amino acid sequence. Proceedings of the National Academy of Sciences. USA, 1995, 92, 8700-8704.

Inna Dubchak, Ilya Muchink, Christopher Mayor, Igor Dralyuk and Sung-Hou Kim. Recognition of a Protein Fold in the Context of the SCOP classification. Proteins: Structure, Function and Genetics, 1999, 35, 401-407.

See also

Examples

x = readFASTA(system.file('protseq/P00750.fasta', package = 'Rcpi'))[[1]]
extractProtCTDT(x)
#> prop1.Tr1221 prop1.Tr1331 prop1.Tr2332 prop2.Tr1221 prop2.Tr1331 prop2.Tr2332 
#>   0.27094474   0.16042781   0.23351159   0.26737968   0.22638146   0.17112299 
#> prop3.Tr1221 prop3.Tr1331 prop3.Tr2332 prop4.Tr1221 prop4.Tr1331 prop4.Tr2332 
#>   0.21033868   0.20499109   0.23707665   0.27272727   0.15151515   0.24598930 
#> prop5.Tr1221 prop5.Tr1331 prop5.Tr2332 prop6.Tr1221 prop6.Tr1331 prop6.Tr2332 
#>   0.18181818   0.02139037   0.15686275   0.21925134   0.22816399   0.15864528 
#> prop7.Tr1221 prop7.Tr1331 prop7.Tr2332 
#>   0.25133690   0.21568627   0.18003565