Profile-based protein representation derived by PSSM (Position-Specific Scoring Matrix) and auto cross covariance

Profile-based protein representation derived by PSSM (Position-Specific Scoring Matrix) and auto cross covariance

extractProtPSSMAcc(pssmmat, lag)

Arguments

pssmmat
The PSSM computed by extractProtPSSM.

lag
The lag parameter. Must be less than the number of amino acids in the sequence (i.e. the number of columns in the PSSM matrix).

Value

A length lag * 20^2 named numeric vector, the element names are derived by the amino acid name abbreviation (crossed amino acid name abbreviation) and lag index.

Details

This function calculates the feature vector based on the PSSM by running PSI-Blast and auto cross covariance tranformation.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

See also

extractProtPSSM extractProtPSSMFeature

Examples

x = readFASTA(system.file('protseq/P00750.fasta', package = 'Rcpi'))[[1]]

dbpath = tempfile('tempdb', fileext = '.fasta')
invisible(file.copy(from = system.file('protseq/Plasminogen.fasta', package = 'Rcpi'), to = dbpath))
pssmmat = extractProtPSSM(seq = x, database.path = dbpath)
pssmacc = extractProtPSSMAcc(pssmmat, lag = 3)
tail(pssmacc)