Quasi-Sequence-Order (QSO) Descriptor

Quasi-Sequence-Order (QSO) Descriptor

extractProtQSO(x, nlag = 30, w = 0.1)

Arguments

x
A character vector, as the input protein sequence.

nlag
The maximum lag, defualt is 30.

w
The weighting factor, default is 0.1.

Value

A length 20 + 20 + (2 * nlag) named vector

Details

This function calculates the Quasi-Sequence-Order (QSO) descriptor (Dim: 20 + 20 + (2 * nlag), default is 100).

References

Kuo-Chen Chou. Prediction of Protein Subcellar Locations by Incorporating Quasi-Sequence-Order Effect. Biochemical and Biophysical Research Communications, 2000, 278, 477-483.

Kuo-Chen Chou and Yu-Dong Cai. Prediction of Protein Sucellular Locations by GO-FunD-PseAA Predictor. Biochemical and Biophysical Research Communications, 2004, 320, 1236-1239.

Gisbert Schneider and Paul Wrede. The Rational Design of Amino Acid Sequences by Artifical Neural Networks and Simulated Molecular Evolution: Do Novo Design of an Idealized Leader Cleavge Site. Biophys Journal, 1994, 66, 335-344.

See also

Examples

# NOT RUN {
x = readFASTA(system.file('protseq/P00750.fasta', package = 'Rcpi'))[[1]]
extractProtQSO(x)

# }