Sequence-Order-Coupling Numbers

Sequence-Order-Coupling Numbers

extractProtSOCN(x, nlag = 30)

Arguments

x
A character vector, as the input protein sequence.

nlag
The maximum lag, defualt is 30.

Value

A length nlag * 2 named vector

Details

This function calculates the Sequence-Order-Coupling Numbers (Dim: nlag * 2, default is 60).

References

Kuo-Chen Chou. Prediction of Protein Subcellar Locations by Incorporating Quasi-Sequence-Order Effect. Biochemical and Biophysical Research Communications, 2000, 278, 477-483.

Kuo-Chen Chou and Yu-Dong Cai. Prediction of Protein Sucellular Locations by GO-FunD-PseAA Predictor. Biochemical and Biophysical Research Communications, 2004, 320, 1236-1239.

Gisbert Schneider and Paul Wrede. The Rational Design of Amino Acid Sequences by Artifical Neural Networks and Simulated Molecular Evolution: Do Novo Design of an Idealized Leader Cleavge Site. Biophys Journal, 1994, 66, 335-344.

See also

See extractProtQSO for quasi-sequence-order descriptors.

Examples

# NOT RUN {
x = readFASTA(system.file('protseq/P00750.fasta', package = 'Rcpi'))[[1]]
extractProtSOCN(x)

# }