Retrieve Drug Molecules in MOL Format from the KEGG Database

Retrieve Drug Molecules in MOL Format from the KEGG Database

Usage

getMolFromKEGG(id, parallel = 5)

Arguments

id

A character vector, as the KEGG drug ID.

parallel

An integer, the parallel parameter, indicates how many process the user would like to use for retrieving the data (using RCurl), default is 5. For regular cases, we recommend a number less than 20.

Value

A length of id character vector, each element containing the corresponding drug molecule.

Details

This function retrieves drug molecules in MOL format from the KEGG database.

See also

Examples

id = 'D00496'  # Penicillamine
# \donttest{
getMolFromKEGG(id)# }
#> Error in httr2::resp_body_string(resp[[i]]): `resp` must be an HTTP response object, not a <httr2_failure> object.