Retrieve Protein Sequence from the KEGG Database

Retrieve Protein Sequence from the KEGG Database

getSeqFromKEGG(id, parallel = 5)

Arguments

id

A character vector, as the protein ID.

parallel

An integer, the parallel parameter, indicates how many process the user would like to use for retrieving the data (using RCurl), default is 5. For regular cases, we recommend a number less than 20.

Value

A list, each component contains one of the protein represented by amino acid sequence(s).

Details

This function retrieves protein represented by amino acid sequence from the KEGG database.

See also

See getFASTAFromKEGG for retrieving protein sequence in FASTA format from the KEGG database.

Examples

id = c('hsa:10161', 'hsa:10162')
getSeqFromKEGG(id)