Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -6.843244e-02 -2.606057e-03 -8.732498e-02  7.900168e-02 -1.172942e-01 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  1.356244e-01 -6.376097e-02 -6.477919e-02 -4.391569e-02 -2.475580e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -5.691450e-02 -5.856633e-02  7.409402e-03  7.282555e-02 -9.438531e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  7.344828e-02 -3.647935e-02 -2.147873e-02  5.341191e-03 -2.472176e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -1.397539e-02 -5.461784e-02  6.704564e-02 -1.062856e-01  1.556299e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  6.185173e-02 -1.271240e-01 -4.090093e-02 -3.467538e-02 -9.397981e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  3.687236e-02  4.878642e-02 -5.056394e-02  8.737931e-02 -3.854040e-03 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -1.124589e-01  1.145954e-01 -3.537049e-02  5.353729e-02 -1.160663e-01 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  7.151735e-02 -6.823814e-02  1.439899e-01  7.460382e-03 -1.015022e-01 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  3.109044e-02  1.121863e-02  1.183277e-01  3.340156e-02 -1.155693e-03 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -6.924658e-03  2.727459e-02  9.438111e-02  1.396347e-01  2.230012e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  8.036373e-02  1.053033e-01 -4.122578e-02 -2.358240e-02 -6.773859e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  3.721082e-02 -6.013786e-03  6.266343e-03 -2.972109e-02 -3.236019e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  1.186111e-02 -2.941893e-02 -5.407092e-02 -9.823740e-02  4.373967e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  1.003466e-02  1.234251e-01 -2.439374e-02  5.935826e-02  2.677826e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -3.655870e-02 -2.630365e-03 -2.689889e-02 -1.036157e-01 -3.960114e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  5.383792e-02  2.531718e-02 -1.209715e-02  4.080340e-02  2.558904e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  9.579778e-03 -5.007592e-02 -9.370463e-02  6.177114e-03 -1.673539e-04 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -3.968150e-02 -6.073377e-02 -7.388420e-02 -1.466155e-02  9.104597e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  1.347018e-02  5.781660e-03 -4.343109e-02 -4.160519e-02  9.706082e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  3.233900e-02 -6.347688e-02  4.537141e-02 -8.019531e-03  9.824511e-03 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -1.195392e-02 -8.196771e-02 -2.523844e-05 -9.011319e-02  6.355815e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  1.164529e-02  9.694197e-02  9.089498e-02 -9.865767e-03  3.636572e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -1.995236e-02  8.125083e-02  3.162442e-02  8.259737e-02  1.540263e-01 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  2.360853e-02  9.498774e-03  4.825332e-02  9.731102e-02 -5.535672e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -6.617964e-02 -1.863297e-02 -1.058370e-02  4.694639e-03 -7.109267e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -2.663100e-02 -7.672972e-02 -9.468806e-02  4.544368e-03 -9.814700e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  2.236930e-02 -4.546994e-02  1.398088e-02 -6.468276e-02  5.127358e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  4.974927e-02 -2.905461e-02 -8.229138e-02  4.261630e-02  1.241420e-01 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  1.006855e-02  8.463543e-02 -1.329427e-03 -2.898604e-02 -1.147375e-01 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  2.784938e-02 -1.693889e-03 -5.202310e-02  1.296705e-02  2.075323e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -1.449565e-02  2.643142e-02 -4.201194e-02 -3.431193e-03  1.220439e-01 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  2.678508e-03 -1.362194e-01  7.118783e-02  6.192362e-02 -7.765990e-03 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  1.676778e-02  9.075696e-02 -1.773648e-02  8.679736e-02  3.780002e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -7.054457e-02 -6.119572e-03  1.630452e-01  8.030017e-02 -9.390979e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -4.491815e-02 -4.982047e-02  6.573202e-03  1.379876e-01  9.998273e-03 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  3.844262e-02  2.709702e-02  3.850235e-03 -7.097479e-02 -2.769567e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  8.761119e-03 -1.118787e-02 -1.388390e-01  1.970166e-02 -8.977161e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  1.536219e-01 -4.163116e-02  1.287905e-01 -1.871902e-02 -8.995828e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  1.173331e-01 -6.362253e-02  8.848089e-03 -4.256217e-02 -2.499713e-01 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -2.155720e-02  8.454458e-02  3.028458e-02  1.150004e-01 -3.862429e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -5.065237e-02  1.185053e-01  2.336979e-02 -4.419610e-02  5.944595e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  7.658305e-02 -1.532512e-01 -5.646407e-02 -1.540859e-01 -8.833132e-03 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  1.249034e-01  3.404951e-02  1.685262e-02 -3.340143e-02  1.226669e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  1.106792e-01  4.327117e-02 -5.095971e-02  2.242383e-02 -2.712515e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -4.382535e-02  6.744914e-02  3.707349e-02  2.909951e-02  8.478382e-03 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -1.266045e-02 -1.015590e-04  3.104591e-02  8.974574e-02 -7.164340e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -7.609473e-03  3.673556e-02 -4.301696e-02 -1.288082e-02 -1.307419e-01 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  1.095455e-01 -1.355986e-01  2.079635e-02 -4.654715e-02 -4.023039e-02 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  5.315412e-02 -1.738030e-02 -6.991788e-02  2.040991e-02 -7.279317e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -2.610034e-02 -1.878516e-02 -1.392655e-02  1.197111e-02  8.681867e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -3.873235e-02  3.753877e-02  9.564215e-02  7.190867e-02  1.220146e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -1.195475e-01 -3.698076e-02  7.452699e-02 -6.486970e-02 -1.956609e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -5.853264e-02 -1.020574e-01 -9.448455e-02  3.605638e-02  5.965319e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  1.793624e-02 -6.294881e-03 -2.487944e-02  1.387888e-01  1.013335e-01 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -8.171092e-02  4.557479e-02 -2.390296e-02  2.469787e-02  1.011543e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -8.281824e-02  7.993656e-02  3.707588e-02 -6.381425e-02 -3.588466e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -6.081293e-02 -4.866470e-02 -4.706369e-02 -3.825574e-02 -3.979447e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  1.235743e-01  5.852131e-02 -4.434194e-02 -7.351457e-02 -4.564041e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  1.164813e-01 -9.721720e-02 -1.341768e-01 -2.156153e-02  6.842946e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  4.412250e-02 -3.486009e-02 -4.383986e-03 -6.966515e-02 -5.136369e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -2.876023e-02  3.904676e-02  1.534440e-02 -7.537831e-02  2.372889e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -6.274982e-02  8.628438e-02  1.534145e-03  6.507263e-02 -6.961647e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  8.632926e-02 -7.392260e-02 -4.856684e-02  2.098635e-02 -1.022401e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  5.415526e-02 -2.643430e-02 -6.341402e-02 -6.920198e-02 -8.561301e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -1.892743e-01  8.195292e-03 -8.691214e-02  2.269055e-02 -6.810970e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -7.104235e-02  3.907129e-03 -4.963645e-02  6.642689e-02  2.424596e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -6.090825e-02  7.374347e-02  4.139379e-02  7.240483e-02  3.920931e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -6.442726e-03 -1.142033e-01 -5.754708e-02  5.607204e-02 -4.489676e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  6.152527e-02 -9.488408e-02  2.868713e-02  1.673272e-02 -5.173730e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  6.224833e-02 -2.538669e-02 -3.659246e-02 -5.487092e-02 -2.982472e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  3.094794e-02 -9.231878e-02  6.730512e-02 -5.178080e-02  3.097740e-03 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -2.320214e-02  5.278432e-02  3.881785e-03  3.141784e-02  5.645275e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -8.943470e-02 -2.571739e-02  7.776271e-03 -1.796533e-02  2.881407e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -1.163967e-01 -1.136849e-01 -5.768017e-02  3.458655e-02  8.519654e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  4.884978e-02  7.197819e-02  6.313616e-02 -6.705657e-02 -9.933832e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -4.553151e-02  9.431880e-02  7.866352e-02 -1.415741e-02 -8.546099e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  5.465382e-02 -4.127475e-02 -1.058307e-02 -9.477922e-02  2.648272e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -1.346839e-03 -3.242897e-02  1.193180e-01 -1.000090e-01  1.557702e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  2.430141e-02 -1.148585e-01 -4.241712e-02 -1.822795e-02 -1.822633e-03 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -8.134667e-02  1.410968e-01  1.414557e-01 -2.967527e-03  1.065835e-01 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  8.822246e-02  1.214030e-01 -1.333603e-02  5.787974e-03  5.907212e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  1.014565e-01 -5.713895e-02  5.367979e-03  1.325341e-01 -1.156019e-01 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -5.764029e-02  2.520996e-02 -2.401834e-02  3.330835e-02  6.892281e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  5.282434e-02  1.283148e-01 -1.050478e-01 -9.404283e-02  2.533477e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -2.637363e-02  4.294132e-02 -3.720768e-02  6.218853e-02  5.289615e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -1.519822e-01  5.540188e-02 -1.480728e-02  4.236101e-02  5.715580e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  1.561311e-01 -5.468263e-04 -1.703053e-02  9.663321e-02  6.127766e-03 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  8.832215e-02  4.620867e-02  1.014463e-02  6.574550e-02  7.226905e-03 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  4.947262e-02  3.750484e-02 -6.868360e-03