Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  0.0142647066  0.0795535009 -0.0198436333 -0.0456847750  0.0732032819 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0819119659  0.0182382414 -0.1452372502 -0.0087427471  0.0312285777 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  0.0141017760 -0.0293640860 -0.0909832535 -0.0050596963  0.0055721810 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  0.0325825255  0.0124466488 -0.0977803951  0.0501108592  0.0357653098 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -0.0867722091 -0.0022262632 -0.0582654038 -0.0084021348  0.0615875944 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.0306853647  0.0317619067 -0.0897808810  0.0405568797  0.0125395531 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0043457305 -0.0200529314 -0.0046746195 -0.0286434341  0.1832428047 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.0898933629 -0.0156850669 -0.0199998778 -0.0805584422  0.0468940078 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -0.0684397715  0.0486822575 -0.0649440351 -0.0208260277  0.0147262294 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0605361659  0.0148414261  0.0371483891  0.1137671892  0.0494243618 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0982939891 -0.1096351272 -0.0966628383 -0.1046323357  0.0070509414 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -0.0221376408  0.0365088594 -0.1102960550  0.0366503701  0.0080220748 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0547234855 -0.1719235544  0.0415413435  0.0171836359  0.0241745596 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0970748884 -0.0471265648  0.1549971641 -0.0969198720 -0.0167705149 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.0181367265  0.0182367430  0.0325886898 -0.1240254021  0.0949766296 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  0.0189897904 -0.0763076649  0.0391071708  0.0937604026  0.0411887098 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -0.0714337974  0.0518889638 -0.0482958481  0.0834223497 -0.0311892721 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0368690705  0.0521012263 -0.0879693843  0.0912932091  0.0885044554 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -0.0910148921  0.0555508166  0.0010972781  0.0368595777  0.0270853351 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  0.0036221308 -0.0782247161  0.1056674456 -0.0888071994  0.0607512761 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0585976951 -0.0441119232  0.0249378794 -0.0473944636 -0.0564610752 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.0236356800  0.0699407422 -0.1470725261  0.0706126938 -0.0516732245 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.0879121394  0.0796481661  0.0027620458  0.1901334008 -0.0290577943 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  0.0053332172 -0.0843263077 -0.1137658416 -0.0435766483 -0.0147876269 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.0137679606 -0.0492148273 -0.0857311278 -0.1383931852  0.1601341116 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0713884511  0.0524115950  0.0236580147 -0.0454588101 -0.0531484551 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  0.0383286060 -0.0836422159 -0.1256570313 -0.0504472347  0.0813621372 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  0.0318022689  0.0614782683 -0.0439829017 -0.0529227454 -0.0669970720 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -0.0851690116 -0.0246678007 -0.1191330046  0.0456292501 -0.0331293044 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.1171310173  0.0166984080 -0.0181310437  0.0801357870  0.0638008636 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  0.0216175651 -0.0669149686  0.0373926022  0.0415595995 -0.0520433519 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.0490543061 -0.0374819571  0.0251361394 -0.0275796441  0.0748239803 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0288965598  0.0171972281 -0.0204544267 -0.0897601329  0.0510611703 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -0.0068491665 -0.0012512405  0.0197143222  0.0425298637 -0.0246942775 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -0.0161762943 -0.0151257247 -0.0848883287  0.0961391823  0.0078498313 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0590443914  0.0256839796  0.0463825682 -0.0552368366 -0.0874118877 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -0.0534170860 -0.0239246850  0.0262132916 -0.0399123338 -0.0476516478 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  0.0420842826  0.0025246189  0.0235164674  0.0889878548  0.0392191716 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -0.0956120527  0.0250316955  0.0996459816 -0.1162661839 -0.0812555430 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  0.0892664392 -0.1098540675  0.1134376646  0.0853008019  0.1305980011 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  0.0446119721  0.0611766798 -0.1272418538  0.0227169106  0.0373390322 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.0056361423 -0.1054486271  0.1015918522 -0.0406398893  0.0220923675 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -0.0765949163 -0.0644954122  0.0711791449 -0.0848134579  0.1254797239 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0623921366 -0.1017787865  0.0159094298 -0.0849288535 -0.0082579139 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.0271607872  0.0137295007  0.1090182291 -0.0222929797 -0.1079884743 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.1231412640  0.0319032991 -0.0182024068 -0.1092809226 -0.0069915212 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0349514008  0.1274923118  0.0651016623  0.0038179214  0.0003874893 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0420741053 -0.0222300618 -0.0451408167 -0.0332003916  0.0229670590 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  0.0633193747  0.0045076870  0.0009430915  0.0091546330  0.1204140784 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  0.0678371314  0.0426812052  0.0757489937  0.0815721618 -0.1570800176 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  0.0239199447 -0.1642773207 -0.0083171612 -0.0535190397  0.0434539483 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -0.0787493524  0.0157446855  0.1288076022  0.0178140580 -0.0046969206 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.0459238249  0.0244612484 -0.0070986760  0.0070199981  0.0164925670 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.0540960886  0.2035281016 -0.0409801254 -0.0638227822 -0.0266143203 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0878536436 -0.0144358760  0.0015956575 -0.0270694153  0.0352372692 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0262864861 -0.0048130429  0.0310347927 -0.0616939062 -0.0393043664 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.1366083015  0.0413175773 -0.0066436958  0.0920099023 -0.0568228953 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -0.0293020914  0.0697047685 -0.0802229685 -0.0580766311 -0.0036814808 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0355902136  0.0816683317 -0.0171410772 -0.0156746606  0.0506432398 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -0.0314361328 -0.0176051137  0.0765415602 -0.0240251085 -0.1262215706 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0479772579 -0.0152409131  0.0146391596 -0.0321425108  0.0261972733 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  0.0127213419 -0.0618540926  0.1292410316  0.0333182868  0.0427971703 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -0.1228544413 -0.0466679435 -0.0112918856 -0.0808016604 -0.0658428964 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0992174856  0.0087627074  0.0542323333 -0.0042493586 -0.0255749558 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  0.1280991802  0.1108520866  0.0459669187  0.0604044734 -0.0654792669 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -0.0028830338  0.0181770535  0.0528880947  0.0821532777  0.0840562863 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  0.0991022743  0.0103188661 -0.1253441528  0.0448539549  0.0343815735 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0817603576 -0.0723685526  0.0585736485 -0.0416337608 -0.0252979175 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.0197565747  0.1068568984 -0.1248343821 -0.0963833294 -0.0457188515 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0450886929 -0.0412841532  0.0347011286  0.0456255401  0.1159934533 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  0.0120621208 -0.0332300420  0.0226949644  0.0345237946 -0.2142725980 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  0.0213980101  0.1056257237  0.0771801092 -0.0285545265  0.0026431427 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  0.0202769841  0.0454718918  0.0115811914 -0.1249149642 -0.0437988735 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  0.0396809467 -0.0095061545 -0.0496757620  0.0476072729 -0.0658305649 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0119035525 -0.0050833087  0.0397408739  0.0694495397  0.1210650926 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.1045273295 -0.1051175298  0.1106163217 -0.0529109379 -0.0864675413 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -0.0561271564  0.0281279067 -0.1137747451 -0.0612931895 -0.0004068257 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0611702460  0.0738268126  0.0006028406  0.0591191914 -0.0192153844 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -0.0615899412  0.0688747856  0.0064513241  0.0522711571  0.0086963870 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0558565642  0.0826995646 -0.0484996048  0.0234329435 -0.1454669368 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -0.0651161354 -0.0032779321  0.0219690492 -0.0448027921  0.0470711699 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0432552657 -0.0198332799  0.0706068043  0.0336780454  0.0933531046 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -0.0179830830 -0.0585494721 -0.0086904368 -0.0121062823 -0.1109821509 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0434547033 -0.0291977141  0.0199833910 -0.0525253103 -0.0056963140 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  0.0460695842  0.0266973571  0.0313554942 -0.0326991171  0.0089669994 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.0028856050 -0.0964011924  0.0172567927 -0.0863183753  0.0121526941 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -0.0221062783 -0.1329967097 -0.0946594853  0.0827276183 -0.0936703620 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -0.0172135902  0.0514813784 -0.0272561744 -0.0557312217 -0.0043832418 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  0.0286941605  0.0925289459  0.0055584039  0.0073243867  0.1056909983 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0360662266  0.0029400096 -0.0549765697