Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0142483468 -0.0228854836  0.0019009147  0.0194756196  0.0941402537 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -0.1446088465 -0.0301451417 -0.0408943847 -0.0140775451  0.0530623127 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0399536421 -0.0300567583  0.0918200173  0.0746043691  0.0347804076 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.0159981658  0.0658656865 -0.0401802284  0.0238192894  0.0130324270 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -0.0438030682  0.0362708201 -0.0879933522 -0.0397434862 -0.0091072855 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -0.0054979966  0.0371761627  0.0499004705 -0.0432344069 -0.0182846716 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0496601115  0.0393458285 -0.0021909859 -0.0926422724 -0.1399298927 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -0.0166293629  0.0257698015 -0.0419803459 -0.0179804614  0.0592064306 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -0.0500771014 -0.0072409934 -0.0118821051 -0.1257892331  0.0924985855 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0890284908  0.0275021008  0.0941315431  0.0423022652  0.0661326376 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  0.0642559744  0.0258132921  0.0190030482  0.0692118641  0.1220790290 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -0.0893716465 -0.0620888167 -0.0201451644  0.0047039757 -0.0348316027 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0724195613  0.0703788648 -0.0247157680  0.0731198348 -0.0335669471 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  0.0003757571 -0.1022022128 -0.0363055617 -0.0335130922 -0.0181760261 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.0803084186 -0.0169379512 -0.0632965790 -0.1404247098  0.1142332720 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.1148466068 -0.1617564207  0.0335599540 -0.0519187971 -0.0079913267 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0050491768 -0.0247749630 -0.1406221933 -0.0712306984 -0.2311706061 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0217940867  0.0245435011 -0.0148134404  0.0155918227 -0.1154098354 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -0.0040353397  0.0283052171 -0.0537843638  0.0161096064  0.0029976183 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  0.0718100534  0.0481977525 -0.0174677366  0.0266826246  0.1024785653 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.1249885262 -0.0105163034  0.0094816710  0.0390743754  0.0722120838 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.0493579345 -0.0700971892  0.0571853473  0.0057200316  0.0992664857 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -0.0661634046 -0.0057049834  0.0131309258  0.1735096713 -0.0124800486 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  0.0546513731  0.0049522528 -0.0753888126  0.0461212574 -0.0658665574 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  0.0374225393 -0.0038401067  0.0589777299 -0.0959650587  0.0645493908 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0040919926  0.0475967092  0.1039778976  0.1147917643  0.0048106499 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0012570995  0.0904332242 -0.0399920676 -0.0111813441 -0.1051968291 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  0.0131320980 -0.0527928328 -0.0917592690  0.0209724913 -0.0569965256 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0060045709 -0.1132632316  0.1647424181  0.0715362527  0.0466073737 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  0.0451586439  0.0324825458 -0.0087800589 -0.0165862800  0.1427627413 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  0.1169520019 -0.0732658932 -0.0960332227  0.0414855813 -0.0736486950 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -0.0195384297  0.0221047440  0.0528736507  0.1269791051 -0.0926249501 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0305864769 -0.0354190417  0.0470469156  0.0381743757 -0.0682808648 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -0.0963907303  0.0097994139 -0.0621865769  0.0704578010 -0.0188644740 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -0.0216560462  0.0142713528  0.0544771235  0.0631691659  0.0533380396 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0446712837  0.0227816634  0.0155746579  0.0448551265  0.0024510576 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0246751676  0.0976773889  0.0918029868  0.0173966566 -0.0670840113 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  0.0830381358  0.0225812965  0.0189989638  0.0974790908 -0.0904453566 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.0341175625  0.0833977279  0.1182158319  0.0049362962 -0.0375101193 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0464702738  0.0091899425 -0.0496240698 -0.0496681978  0.0687367532 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.1569409409 -0.0045315069 -0.1369471468 -0.0466432051 -0.0660472335 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.0960832882 -0.0111447341 -0.0123024637 -0.0680386632 -0.0965652283 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.0028057177 -0.0140232822 -0.0566248447 -0.1398162739  0.0440041248 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -0.0861750158 -0.0410188371 -0.1861684223 -0.0648108657  0.0026677470 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  0.0847757074  0.2165324900 -0.0055871570  0.0389112073 -0.0226509766 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.0815951368 -0.0556208757 -0.1072222027 -0.0912809486  0.0324986312 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0827547699 -0.0433420277  0.0976548881 -0.0275943166 -0.0086679817 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -0.1218786262 -0.0034025861 -0.0107744807  0.0784203737 -0.1222432456 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.0115097517  0.0899996672 -0.1024832569  0.1012282201  0.0720476892 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  0.0450479385  0.0402772747 -0.0505952022  0.0166855336  0.0769714323 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  0.1233843396  0.0629771289  0.0703196087  0.1106607504  0.0288580757 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.0496699131  0.0205957890 -0.0363306505 -0.0821920577  0.0904469661 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.0810780422 -0.0031132832  0.0020964418  0.0210473177 -0.0717215518 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  0.0561286759  0.1100113278  0.0531393600 -0.0182794663  0.0491097736 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0352437674  0.1757839594  0.1497493917 -0.1039095774  0.0008783436 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -0.0773653911 -0.1347085615 -0.1310782217  0.0478837861  0.0108453221 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.0072986942 -0.0966980764 -0.0104125679  0.0283241475 -0.0511110236 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -0.0239166179  0.1097945286  0.0143309748 -0.1701740608  0.0459544751 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  0.0155679513 -0.0909368542  0.0532719335  0.0033233439 -0.0367014694 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -0.0057873036 -0.1314006770 -0.0429648671 -0.0288433616  0.0407255354 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0874817049  0.0034303451  0.0639155917  0.0784477418  0.0086209764 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.1616979066 -0.0931260015 -0.0115739275 -0.0347181813 -0.0434562155 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0166304170  0.0763394479 -0.0049361792  0.0828962704  0.0250244587 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -0.0111691280 -0.0675812723 -0.0207172298 -0.0107164852  0.2245153312 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.0242778809  0.0617362259  0.2182219013 -0.1542478649  0.0904850000 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.0363029449 -0.0238283810  0.0277823072  0.1184827332  0.0195329459 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  0.0368593554  0.0166331624 -0.0649579270 -0.0181200481 -0.1247208390 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.1448148822  0.0150811900 -0.1000951770  0.0288478501 -0.0036597114 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.1878333142  0.0291575879  0.0254473643 -0.0013840428 -0.0369667738 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  0.0709670203 -0.0203179157  0.1343578219 -0.0633269041  0.0564580377 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.1346646357 -0.0108010947  0.0156204347  0.1580418556 -0.0830505319 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  0.0493745506  0.0597141649 -0.0643059993  0.0308311642 -0.0875644572 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.1230269261  0.0209915942  0.0179452261 -0.0377001942 -0.0248257044 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.1705001735 -0.0743687851  0.0289575866  0.0341944778 -0.0582157613 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0386687083 -0.0506393108  0.0456270505  0.0583550306 -0.0367611190 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.1692912206 -0.1488598109 -0.0017474304  0.0701823372 -0.0939188538 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0294783233 -0.0423058758  0.0259082401 -0.0801025085 -0.0279413548 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.1623783728  0.0179895868 -0.1357500011 -0.0855314259  0.0062542447 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -0.0300542734 -0.0047633293  0.0306165366 -0.0015262146 -0.0095872077 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.1278996949 -0.0243942857 -0.1507691217 -0.0499392640  0.0587360232 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.0212609671  0.0399253598 -0.0364056036  0.0772479664 -0.0952777633 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0730496802  0.1362252457  0.0184334550 -0.0711597151 -0.0894383796 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.1742042663  0.0113261164 -0.0599178030 -0.0185082447  0.0520365608 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  0.0355918722  0.0428105669 -0.1227042192 -0.0837020544 -0.1044191958 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.0357888961  0.0565378814 -0.0367343480 -0.1362871831 -0.1147027794 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.1015329868  0.0655471868  0.0729478134  0.0265827122 -0.0413653097 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -0.0042251085  0.0725796283 -0.0466662525 -0.0478346315  0.0494098465 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -0.1128697139  0.0373057712 -0.0205566252 -0.0632938379 -0.0859268011 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  0.0370761739  0.0500980459  0.0324925457  0.0539457485  0.1139779653 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0609141962  0.0349750398 -0.0054437258