Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -6.929884e-02 -3.717374e-02 -4.621823e-02  1.445832e-01  2.209574e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  4.448730e-02 -2.026914e-01  3.651747e-02 -3.876794e-02  3.626264e-03 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -4.136088e-03  7.479292e-02 -1.031863e-02  1.253297e-01 -9.215867e-03 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  5.496281e-02 -5.098185e-03  3.723593e-02 -8.060431e-03  5.611063e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -7.359603e-02 -5.241508e-02 -7.445928e-02  8.033706e-02  1.226398e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -8.948681e-03 -1.152444e-02  5.275923e-03 -7.445444e-02 -1.161179e-01 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  8.317657e-02  3.887004e-02 -9.398719e-02  3.088018e-02 -1.275307e-01 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  2.811150e-02 -1.497121e-04 -1.016191e-02 -7.961390e-02  5.504735e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  4.419574e-02 -1.355195e-02  1.307488e-01 -1.149630e-02 -4.531522e-03 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -5.615506e-02  8.780833e-02  1.538436e-02 -6.159850e-03  6.403900e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  1.146940e-02 -4.691385e-02  1.997858e-02  2.570830e-02 -4.480980e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -4.040574e-02 -5.916187e-02  2.676501e-02  2.901186e-02 -8.063681e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  1.542087e-02  1.804245e-02  4.462985e-02  6.024899e-02 -6.066773e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  1.156957e-02  1.735958e-01  9.836988e-02 -3.448603e-02 -1.242445e-01 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -7.443046e-02 -5.732850e-02 -9.567449e-02 -7.634457e-02  5.273120e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -1.190176e-02 -3.607808e-02  3.555092e-02  1.324801e-01 -1.650958e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  3.892377e-02 -4.144886e-02  5.817585e-02  1.480044e-02  5.786700e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -7.567301e-02 -1.026571e-01 -4.751226e-02  1.395062e-01 -2.797566e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -2.401531e-02 -2.205053e-02 -2.425548e-02 -1.992338e-02 -4.247195e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -5.978038e-02 -4.617769e-02 -3.593461e-02  5.375293e-02 -1.687136e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -7.540747e-02 -3.906373e-02 -6.319428e-02 -5.765851e-02 -2.207596e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  6.097700e-02  6.429030e-02  2.351460e-02  9.211949e-03 -3.296695e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  8.502069e-02 -4.952301e-02  5.421065e-02 -4.078161e-02  1.650668e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -7.253120e-02  4.724998e-03 -1.191356e-01  1.084895e-01 -6.520702e-03 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  1.876740e-02  5.719254e-02 -4.945698e-02 -4.405615e-02  2.833312e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -9.208642e-02 -4.596378e-02 -2.361868e-02  4.514267e-02  1.046060e-01 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  2.697281e-02  1.123879e-01 -7.389491e-02  9.289691e-02  1.137517e-01 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  2.587163e-02  8.062337e-03 -1.594773e-02 -1.013803e-01  1.827183e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  7.004499e-03  1.201188e-01 -8.691229e-02  1.412879e-01 -7.319626e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  1.496120e-01 -5.474106e-02  2.754495e-02 -4.618853e-02 -7.479132e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -5.218099e-02 -6.139016e-03 -1.717045e-01 -3.315078e-02  3.604435e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -5.617302e-02  4.036391e-02  6.598594e-02 -4.784849e-04 -1.732638e-01 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  4.720236e-02  2.014227e-03 -1.017073e-01  4.894717e-02 -5.185701e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -2.604316e-02  6.860153e-02  1.091318e-01 -8.439524e-04 -1.084961e-01 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  7.060218e-02  6.397490e-02 -8.163196e-02  2.316699e-04  4.187970e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  3.061644e-02 -3.607728e-02 -5.378169e-02  2.305786e-02 -3.558390e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -3.492567e-02 -9.365090e-03  8.230517e-02  1.116076e-02 -5.383498e-03 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -8.757107e-02  7.511732e-02  1.978316e-02  8.338895e-02  1.186791e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -3.493076e-02  1.228579e-02 -2.714108e-02 -6.075692e-02  1.121759e-01 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  5.727886e-02 -4.626150e-02 -1.715191e-01  1.496847e-02  5.426425e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  2.235771e-02 -1.328474e-01  2.592065e-02  7.164085e-02 -4.822317e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -1.276994e-01 -1.172594e-01 -2.615006e-02 -3.208486e-02 -1.021345e-01 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  1.106746e-01  1.023157e-01  8.260322e-03 -1.878816e-01 -6.835349e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  2.763549e-03  8.857638e-02  2.319089e-02 -2.827058e-02  1.448334e-01 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  7.781340e-02  6.181292e-02 -2.347024e-02  4.990365e-02  6.307328e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  2.058315e-02  6.493804e-02 -5.286505e-03  4.822093e-02  1.242162e-01 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  5.994126e-03  8.296318e-02  9.438056e-03  7.534674e-02  3.483205e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  1.265368e-02  6.633311e-02 -1.711198e-02 -4.057126e-02 -2.116067e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -3.307328e-02  7.113201e-02  6.247400e-02  1.214051e-01  7.811288e-03 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  1.152139e-01 -6.514062e-02 -8.360374e-02 -6.834612e-02 -1.268447e-01 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  2.627784e-02 -4.733546e-02  4.341392e-03  2.019916e-03  8.480302e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -6.470751e-02  2.306710e-02 -2.769580e-02  8.351060e-03  5.768352e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -2.035327e-02  6.739502e-02 -1.830153e-02 -4.521414e-02 -7.051543e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -1.127989e-02 -4.705194e-02 -1.118709e-01  5.979946e-02 -4.423434e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  1.157487e-01  3.482251e-02 -5.848807e-02  2.060998e-02 -8.937582e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -2.377373e-02  1.230705e-02  7.404983e-02  9.727138e-02 -4.890544e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -6.951011e-02  8.402585e-02 -1.095019e-01  3.388375e-02  6.771848e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -7.073525e-03 -2.049869e-05  2.595272e-02 -4.142824e-02 -4.765434e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  4.357869e-02 -8.868458e-02 -8.939018e-02 -1.490041e-02  4.764428e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -1.358102e-01  9.894747e-02 -1.174572e-02 -4.673052e-02  9.025232e-03 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -5.190225e-02 -2.388938e-02 -1.932692e-02  4.656408e-03  9.227540e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  5.166757e-02 -5.772034e-02  6.132485e-02  2.601073e-02  1.230264e-01 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  8.199757e-02 -6.389965e-02  1.650026e-02  6.912731e-02 -5.189651e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -3.132191e-02 -3.676156e-02 -4.617473e-02  2.147498e-02  1.095673e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -7.757700e-02 -1.099882e-01  4.602722e-02  5.187100e-02 -5.792100e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -5.325690e-02 -4.115701e-02  4.116600e-02 -4.227784e-02  6.596993e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -8.816228e-02 -3.404124e-02 -3.718351e-02  1.127672e-02  1.459081e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  3.382756e-03 -4.441025e-02 -6.043526e-02 -3.506280e-02  5.163297e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  4.380698e-02 -6.530649e-02  2.090758e-02 -1.269123e-02  3.318947e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  4.182937e-02 -1.570913e-02  3.831724e-02  4.782060e-03  2.665275e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -1.187559e-02  9.874083e-03  3.525746e-02 -6.234715e-02 -1.028828e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -3.619518e-02 -6.213993e-02  8.951547e-02  6.310114e-03 -1.027439e-01 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  2.635121e-02 -1.050894e-01 -8.992158e-02 -9.333446e-02  1.253906e-01 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -1.154031e-02 -9.025262e-02 -4.702630e-02 -9.012746e-02  1.412673e-01 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -5.054096e-02 -4.232981e-03 -5.646043e-02  7.013396e-02  4.948350e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  2.301508e-02 -8.285049e-02 -1.529284e-01 -2.804309e-02  5.255738e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  4.354098e-02 -4.804204e-02 -1.496844e-02 -4.459730e-03  1.080009e-01 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -3.500608e-03 -6.847686e-02 -4.073517e-02 -9.105316e-03 -3.386681e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -1.588069e-02  3.024741e-04  6.220081e-02  1.622656e-02 -1.072412e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  4.420626e-02  1.004765e-01  2.574681e-02 -4.212458e-02  1.725132e-03 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  5.163789e-02  2.931653e-02  3.987861e-02 -8.197035e-02  7.148988e-03 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -6.815264e-03  2.730554e-02  3.262793e-04 -5.657343e-02  7.551571e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  3.496833e-02  1.852224e-02  1.693436e-01 -5.097558e-02  6.190395e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  6.934990e-02 -8.224537e-02 -6.181090e-02 -6.464137e-02 -6.525569e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -6.675064e-03  2.178006e-02 -2.968864e-02  1.360390e-02 -5.671068e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -2.992213e-02 -1.572081e-02 -7.356469e-03 -6.704087e-04  2.878934e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -5.491946e-02 -1.036085e-01 -6.573575e-02 -1.016083e-01 -7.718375e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -2.441703e-02  1.409572e-02 -3.165856e-02 -5.981478e-02  5.123485e-03 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  4.246558e-02  2.783104e-02 -4.952403e-02 -9.740617e-02  5.314593e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -1.394757e-02  6.684187e-03  1.457321e-02