Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0189919898  0.0439221887 -0.0960508174 -0.0120599768  0.0482875846 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -0.0903950980  0.0081512305  0.1429366795  0.0383841497  0.0177996431 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  0.0675908719  0.0060115637  0.0432014160  0.0531071245 -0.0992079038 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  0.1188474912  0.0206347918  0.0286100281  0.1115176873  0.0801247333 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0116959841 -0.0337813782 -0.0292778116  0.0944635193  0.0573380229 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -0.0325377455  0.0233484393 -0.1615698757 -0.0952722472  0.0852722101 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0788382252  0.0902970291 -0.1226789479  0.0325053583  0.0245332581 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -0.0218898429  0.0081316561 -0.0774798525  0.1108366950  0.0847299120 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.0597189953 -0.0589041476  0.0645001760  0.0665893094  0.0132882868 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0284369499  0.0813607392  0.1365893489  0.0025093525 -0.0634547754 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  0.0443394241  0.0052472666 -0.0117494478 -0.0750922521  0.0581890208 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.1108956968 -0.0741937727  0.0914303025 -0.1007988440  0.0654784455 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -0.0546390511 -0.0136296141  0.0612400108  0.0899777801  0.0417207572 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#>  0.0770100941 -0.0166874516  0.0606455053  0.0233613292 -0.0053976463 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.0535242133 -0.0214417847  0.0537600377 -0.0038423284 -0.0975133943 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  0.0489603569 -0.1456475308  0.0187837432  0.0341426591 -0.0350797412 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0389447979 -0.0218167080 -0.0061610852  0.0354178333  0.0400851460 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0145791986 -0.0194559695 -0.0167933216  0.1151959254 -0.1547023788 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -0.0345738145  0.0576454361  0.0520393725 -0.0102608692  0.1458926617 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0270332507 -0.0150777628 -0.1099055958  0.0734172921  0.0737914811 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -0.0070629719  0.0082498018  0.1081262784 -0.0608963510  0.0307359322 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.1058285964 -0.1491986726  0.0358153899 -0.0181953824 -0.0143351219 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.0311658393 -0.0566014959  0.0114634649 -0.0803342986 -0.0040754810 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  0.0506996783  0.0587556380  0.0519858364  0.1155747754  0.0512599794 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.0157585826  0.0582422347  0.1087312403 -0.0305885970  0.0080519956 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0431970694  0.0204607970 -0.0377325920  0.0452442821 -0.0176885764 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  0.0506135158  0.1205692067 -0.0999339743  0.2183759118 -0.1006006515 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  0.2141886521 -0.0773026314  0.0513668129 -0.0603459151  0.0545357535 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0722460485  0.0117301285 -0.0356943431  0.0611340005  0.0542134602 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  0.0109746301  0.0207069025  0.0651396210 -0.0789593669  0.0289653292 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  0.0449529437 -0.0413463907  0.0309370932  0.0905689129 -0.1159667777 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.0470141382 -0.1115012941 -0.0989304701 -0.1564810276 -0.0198986590 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  0.1176492810 -0.0803717469  0.0176966806 -0.0304328337  0.1004081124 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0657800319 -0.1161344535  0.0711587432 -0.0210727634 -0.0042984381 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0392923791  0.0180238638  0.0699260921  0.0299155677 -0.0242090971 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  0.0064946055  0.0574803548  0.1580746239  0.0011786182  0.0316304601 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0690168609  0.1698372277 -0.0522838813  0.0951760809  0.0222372541 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.0317726891 -0.0852484774 -0.0198200228 -0.1555676826 -0.0001153137 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -0.1150838253  0.1013964496 -0.0195516502  0.0106254028 -0.0195907952 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0562169557 -0.0338879677 -0.0199118866  0.0991479686 -0.0330619666 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.0288519669  0.0716683464 -0.0217467836  0.0210602081 -0.0200373974 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.1092908156  0.1783288479 -0.0295594591 -0.0372177555 -0.0249809237 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.0112784341  0.0676477328  0.0448350364 -0.0265386800 -0.0209127514 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -0.0786300512  0.0970197614 -0.0505707822  0.1228287764  0.0599125956 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  0.0564091286  0.0996099879  0.0808377026 -0.0264931690  0.0430432020 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  0.1008234690 -0.1414975116 -0.0081518347 -0.0245824649 -0.0321138367 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  0.0227527558  0.0710996760  0.0605133802 -0.0232352911 -0.0655343644 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0287402886  0.0021051016 -0.0663178082  0.0531982870 -0.0162169127 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  0.0623673472  0.0135490390 -0.0101506185 -0.1001193055  0.0773734148 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  0.0538607841 -0.0186083291  0.0321028298 -0.0888496590 -0.0069232941 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.0458715339  0.0264271481 -0.1036488062 -0.0938820305  0.1893644635 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -0.0178990580 -0.0441390678 -0.0725098943  0.1160187879  0.0470766252 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -0.0330175437 -0.0480701433 -0.0532440501 -0.0799026755  0.0173362456 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  0.0553415121 -0.1429727255 -0.1880283168  0.1456643376 -0.0730415393 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0597726516  0.0704411590 -0.1573652948 -0.0754960640  0.0315858792 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0313634418  0.0226799797  0.0463535372 -0.0331322920 -0.0180272147 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -0.0888567426  0.0611447474  0.1031679072  0.0741523306  0.0595252062 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0512267747 -0.1559710601  0.0263960022  0.0376134482  0.0885938740 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0859375320  0.0056419342 -0.0327584427  0.0850877427 -0.0045720267 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0420778080  0.0433374047  0.0251603088  0.0622450910  0.0273598584 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0958548316  0.0956078345  0.0254809768  0.1371951353 -0.0510624730 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  0.0297450243 -0.1145772700  0.0351555846 -0.0717497431 -0.0391827337 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -0.1143856578  0.0399622817 -0.0431909519  0.0216187719  0.0343192904 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0264170497  0.0187544998 -0.0273834823 -0.0652878925 -0.0314685514 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.0169236596 -0.0240774405 -0.0989886779  0.0653132871  0.0185248441 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -0.0863793692 -0.0007852133 -0.0853329406 -0.0863178832 -0.1035374550 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  0.0214191669  0.1398993367  0.0892235789  0.0097766022 -0.1696417302 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0501158616  0.0924712519 -0.1480968383 -0.1336922619 -0.0448422238 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.1323563064  0.0729048396 -0.1888672079  0.0315273122 -0.0240715234 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  0.0343207311 -0.0471633454  0.1093097014 -0.0239699841  0.0126635874 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0784799244 -0.0386831364 -0.0770199602 -0.0379029071 -0.0267343096 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.0272083293 -0.0780775490  0.0608662817 -0.0143262128  0.0254302436 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  0.0698555382 -0.0149869716  0.0268578665  0.0386250539 -0.0717305897 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.0784333791 -0.1004621229  0.0164405004 -0.0942140759  0.0629049398 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0127657806  0.0984882063 -0.0412842805 -0.0698315514 -0.0436226329 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  0.1116071916  0.0395822203 -0.1089688934  0.0548997024 -0.0083961937 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0672177302  0.0325472406  0.0210346395 -0.0396819126  0.0178513811 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0918992006  0.1258190459  0.1592891655  0.0814045604  0.0022236607 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0274691923 -0.0832876087  0.0031363497  0.0010773344 -0.1628472645 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0067804247 -0.0020191089 -0.0815891210  0.0025469571  0.2113715028 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.0531138264 -0.0059271110  0.0088924009  0.0612088196  0.0850636548 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0717254457  0.0309036780 -0.0035489663 -0.0065880627 -0.0173063933 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -0.0044071256  0.0348588748 -0.0331733195 -0.0182018917  0.0294427027 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  0.0398366529  0.0124934056  0.1024757691 -0.0915954966  0.0218367904 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.0215401947  0.0963471182  0.0823245166 -0.0214350084 -0.1835630824 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.0562010318  0.0455176831 -0.1118817133 -0.0680685660 -0.2193268484 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.0717311928  0.0147678280 -0.0862228419  0.0422895786 -0.1102890576 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.0388018486 -0.0728137828 -0.0710223144 -0.0100764535  0.0685747681 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  0.0421413654 -0.0363159857 -0.1190974001 -0.1098012725 -0.0328992492 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -0.1018754556  0.0386847419 -0.0197292112