Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -5.628332e-02  1.830608e-02  2.978597e-02 -6.873691e-02 -2.343379e-01 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -4.072131e-02 -7.335293e-02 -2.696131e-02  2.336727e-02 -1.262010e-01 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -5.539347e-02  9.986982e-02 -7.579707e-02  2.106864e-02  3.334214e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -9.756151e-02 -5.568617e-02  6.829618e-02 -1.509087e-01  2.927785e-03 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  9.705299e-02 -4.522173e-02  4.362624e-03 -1.328329e-01 -5.169819e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -4.161898e-02 -7.628460e-02 -4.091153e-02 -2.075344e-02 -9.250890e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -1.534069e-03 -2.771455e-02 -5.717296e-02 -1.379781e-01  1.177682e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -2.295209e-02 -5.330497e-02 -1.639971e-02 -4.883490e-02  6.229140e-03 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  2.626691e-02 -1.311237e-02  6.687855e-02  5.605011e-03  3.215984e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -2.755058e-01  8.003058e-02  1.983915e-01 -8.312283e-02  1.275440e-01 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  2.326148e-03 -2.283297e-02 -3.229932e-03  6.290645e-02 -8.838667e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -6.414177e-02  3.848165e-02  4.422136e-02 -1.226182e-02 -3.490276e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -1.237664e-02  4.770382e-02 -5.336146e-03 -7.889072e-02  6.552212e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -1.045373e-01  4.802778e-02  1.047122e-01  3.621665e-02  1.703125e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -5.216119e-02  1.705553e-02  4.704835e-02  6.708595e-02 -2.352729e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -5.085355e-02  8.921503e-02 -1.469648e-01 -3.161446e-02  2.646022e-03 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -8.186667e-02 -5.062380e-02 -6.220492e-02  1.075025e-01  1.037682e-01 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -1.075260e-01 -2.763290e-02 -1.932706e-01  3.421511e-02  7.997302e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -1.252674e-01  1.353382e-01 -1.004880e-02 -1.116092e-01 -4.442791e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -1.686108e-02 -9.074319e-02 -6.656485e-02  8.443300e-02 -4.143151e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -1.021228e-01 -6.668838e-02 -3.430822e-02 -7.734551e-03 -1.092514e-01 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -6.397263e-02  1.000115e-02 -8.013414e-03 -8.928399e-02 -1.899522e-01 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  4.431300e-02  4.006123e-02  5.925669e-02  3.595327e-02  2.373059e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  5.672731e-02 -1.870243e-01  7.279764e-02 -1.514610e-02  1.708086e-01 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -6.026250e-02 -1.109763e-01 -9.132521e-02 -4.637701e-03  4.772218e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -5.558645e-03 -1.428317e-01 -1.982913e-02  6.040807e-02  1.246713e-03 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  5.849263e-02 -4.188384e-03  1.750811e-02 -4.968076e-02  6.607376e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  5.888710e-03  1.308684e-01  7.873647e-03  5.810019e-02  1.258327e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  2.130954e-02  1.805292e-02 -4.797598e-02  1.447271e-01 -1.437544e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  3.475587e-03  3.170226e-02 -7.628049e-02  8.761231e-02 -1.153526e-01 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  1.199951e-01 -6.468574e-02 -1.298329e-02 -2.785052e-02  3.488567e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -3.523770e-02  1.355756e-03 -8.121433e-02  1.562192e-01  1.092137e-01 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  3.825129e-02  1.412857e-02 -1.054159e-02  6.752457e-02  3.240694e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -4.284277e-02 -5.801670e-02  5.057622e-02  4.009173e-03  7.874026e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  2.367247e-02 -3.308895e-02 -2.242578e-02  1.271500e-02  9.820035e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  9.136323e-02  1.267161e-01  1.060580e-01 -5.793072e-02 -3.967578e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  1.500677e-02 -1.628496e-01 -3.616831e-02  3.424884e-02  7.926469e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -4.712129e-02 -1.292995e-01  8.873414e-02  6.704874e-02  1.836749e-03 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -1.952401e-03 -6.285590e-02  1.150011e-02  7.460416e-02 -2.718956e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  2.720991e-02 -1.261982e-01 -5.983103e-02 -3.221131e-02 -1.049444e-01 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  1.009139e-02 -4.944443e-02  7.590064e-02 -1.097890e-01 -2.087199e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  4.271925e-02 -7.171489e-02 -6.233365e-02  2.437082e-02 -1.844760e-01 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -1.922821e-02 -8.499780e-02 -4.673129e-02 -1.881141e-02 -1.232406e-01 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  1.347347e-03  1.040240e-02  7.390190e-03 -3.611585e-02  3.677182e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  3.490675e-02 -8.561196e-02  2.608466e-02  1.916799e-02 -6.419025e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  2.105380e-02  6.810676e-02 -8.072462e-02  1.239093e-02 -3.238712e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  8.505826e-02  5.917425e-02 -4.131245e-02  6.140427e-02  1.513958e-01 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  1.218866e-01 -1.291466e-02 -1.955075e-01  8.382905e-02  2.330506e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -1.251268e-01 -8.213228e-02  1.094164e-01 -5.668046e-02 -1.247069e-01 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -1.618849e-02  5.200788e-02  4.612272e-02 -7.498924e-03  3.382596e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  2.818308e-02  3.057900e-02 -8.124745e-02  6.449510e-02  8.845461e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  5.309869e-03  1.198812e-01  1.001899e-01 -2.755025e-02  1.596488e-01 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -6.635829e-02  2.249339e-02  2.665828e-02  7.457609e-02  4.063103e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -6.243223e-02 -1.371650e-01  1.222190e-01  6.978385e-02 -1.283745e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  2.775811e-02 -2.859045e-02  1.068915e-01 -4.062702e-02 -5.521558e-03 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  2.848551e-02 -1.069463e-02 -5.721263e-02  1.480668e-03 -2.905785e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  8.803480e-03 -3.395250e-02  7.932125e-03 -5.404630e-02  9.090711e-03 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -4.361016e-02 -5.457380e-02  3.819553e-02 -7.553856e-02 -2.131601e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -9.918738e-02 -1.914488e-02 -2.124567e-02 -5.265123e-02  5.425966e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -2.059219e-02 -9.281088e-02 -3.537939e-02 -1.250158e-02 -2.383942e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  1.918551e-03 -3.316654e-02  1.929991e-02  8.563158e-03  7.162588e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -6.241857e-02 -2.242147e-02 -2.642555e-02 -2.499884e-02 -9.259446e-03 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -3.447451e-02  1.483339e-01 -3.279722e-02 -1.062264e-01  5.324039e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  2.616714e-02  4.980985e-02  6.643570e-03 -5.388736e-03 -9.970484e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  5.760156e-02  5.906402e-02  8.387656e-02 -4.842300e-02 -1.187421e-03 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  4.870974e-02 -8.828630e-02  4.769466e-04 -4.618684e-02 -4.335139e-03 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  7.420953e-02  1.073992e-01 -5.493736e-02  7.050396e-02 -6.272656e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  1.794038e-02 -5.817494e-02 -4.943095e-02  1.271982e-01  4.930586e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  9.926234e-03 -3.581290e-02 -6.670277e-05 -1.077980e-01  7.754926e-03 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -6.624261e-02 -2.899300e-02 -2.542324e-02 -2.311927e-02 -7.658047e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -3.734712e-02 -1.498585e-01 -7.658458e-02  1.901376e-01  3.252914e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  6.531584e-02  4.763479e-02 -6.892240e-03  1.907439e-02  5.327332e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -1.281520e-02 -4.555447e-02 -1.472889e-01 -9.935705e-02 -8.345483e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  4.105933e-02 -1.882572e-02  1.051976e-01  1.671768e-03 -6.692205e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  1.098218e-01 -2.474998e-03  5.714150e-03 -4.755353e-02 -1.156561e-01 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -7.623830e-02  9.854946e-02  8.435020e-02  9.950238e-03  1.872244e-01 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  1.239096e-02 -9.064211e-02 -6.790957e-02  8.134676e-02  1.595936e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  2.408976e-02 -5.059398e-02 -3.337354e-02 -1.349196e-03 -5.300562e-02 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  2.422127e-02 -5.166149e-02  8.997362e-04 -7.963047e-02  9.214080e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -6.293512e-02  3.681634e-02 -2.101938e-02 -3.267078e-03 -1.360831e-01 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -7.265757e-02  1.473878e-01 -4.256894e-02  3.563749e-02 -1.925780e-01 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  9.271486e-02 -1.167858e-01  3.122402e-03 -5.483505e-02 -3.644026e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  2.900048e-03 -8.263594e-02  1.262217e-01  6.744369e-02 -1.096513e-01 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -1.651030e-02 -5.616157e-02 -4.927749e-02 -6.314463e-02  1.526216e-01 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -4.084462e-03 -1.026227e-01 -3.879689e-03 -1.051665e-02 -3.624268e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -4.323364e-02 -2.081922e-02 -5.775778e-02  3.221338e-02  1.770640e-01 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -3.714036e-02  9.782434e-02 -5.825144e-02  1.169673e-01 -4.570832e-03 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -7.267307e-02 -1.993444e-01 -2.416920e-02 -4.334937e-02  3.832709e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  4.580913e-02  5.151436e-02 -5.913498e-02  1.577324e-02 -4.995892e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  4.792851e-02 -6.677311e-02  9.128425e-03