Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0996411399  0.0787877333  0.0132055664 -0.0428235213 -0.0572208336 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0322889058 -0.0367061037 -0.0245918437  0.0616093914  0.0114761084 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  0.0605438059 -0.0900209117  0.0716243513  0.0241918375  0.0093824537 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.0850140098 -0.0779420384  0.0145415696  0.0519236119 -0.0479597595 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0708809839 -0.1229831095 -0.0848360237 -0.0602542507  0.0267781642 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.1386394079  0.0225999603 -0.0682121712  0.0294496368 -0.0820352628 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0125350436  0.0741472258  0.0628503961  0.1406719974  0.0356245272 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.1054211215 -0.0713882462 -0.0216514056 -0.0721521119  0.1172091660 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.0584401332 -0.0423114671 -0.0855573127 -0.0715482968  0.0272049748 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  0.0851909463  0.0701927239 -0.0605694593  0.0100930391  0.0244259176 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0599804639  0.0016989559 -0.0608267715 -0.0094812227 -0.1199005697 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.0958237978 -0.0287932777  0.0437719395  0.1323771036  0.0328200812 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -0.1453387947  0.0114135373 -0.0253234718  0.0281708018 -0.0370476093 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0405431870 -0.0429056740  0.0369908362  0.0148135281  0.0076434743 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.0017617248 -0.0741798668  0.0463098981  0.0028861731  0.0200016268 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0278705054  0.0335994298 -0.0465299012  0.0210428242 -0.0134759796 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -0.0018433964  0.0853971638  0.0033822685  0.1017523837  0.0739315828 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0089131124  0.0417689081 -0.1276895741  0.1210555486  0.0983761801 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0439094079 -0.0179864630 -0.0753936269 -0.0198105660  0.0403302963 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0928301591  0.0154962613  0.0120811679  0.0656722995 -0.1829365311 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0764287540  0.0392128086  0.0904602228 -0.0560364938  0.0417306451 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.0035401677 -0.1225670046 -0.0081751280 -0.0312174608  0.0667366864 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.0069323578  0.0410386346  0.0934281246  0.1279485919  0.0289951688 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  0.0463713944 -0.0990214110  0.1470866351 -0.0084726617  0.1139218565 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  0.0648681843 -0.0163630842 -0.0385856858 -0.0259802342 -0.0053053750 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0971385202 -0.0735595321  0.0122855998  0.1041836702 -0.0048335267 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  0.0190115054  0.0385032354 -0.0245062565  0.1393426996 -0.0389780093 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  0.0050017739  0.0021370116  0.0062955520  0.0192960881 -0.1103564439 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0499804269 -0.1129597550  0.0025218425  0.0932856073  0.0791058493 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.1373636509  0.0394316981 -0.0464146045 -0.0639444385 -0.0965082737 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  0.0615173017  0.0668727757  0.0181209332 -0.0197834106 -0.0228655009 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.0261450093  0.0315505602 -0.0733248418 -0.2261561436 -0.0391528025 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  0.1076130991 -0.0178067157  0.0557569256  0.1133974121 -0.0465578881 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -0.1624752155  0.0064449544 -0.0843438754 -0.0627239643  0.0026427430 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -0.0210479978  0.0080565586  0.1015739710  0.0238423397  0.0303629415 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0026730546  0.0791199833 -0.0702965939  0.0376895614  0.0469901025 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0687358888  0.0393255496 -0.0355466900  0.0660077487 -0.0148328863 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.0703912188 -0.1116721007 -0.0167583792  0.0141852037 -0.0619108650 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -0.0056460005 -0.0506364239 -0.0350568806  0.1138893712  0.0033502681 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0537186866 -0.0486880956  0.0741610709 -0.0000982989 -0.0930985669 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.0035435977  0.0100010436  0.0502669934 -0.0602345564  0.0512499058 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.0510780930  0.0336490281  0.0319092643 -0.0764259570 -0.0079883820 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -0.0868728515 -0.0015358846 -0.0778405456  0.0882236115 -0.0143230338 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0473395110 -0.0442767233  0.0451051485 -0.0560892344 -0.0136116138 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.0325012120  0.0431291178 -0.0180617624  0.0762692123  0.1368902393 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  0.0925245228  0.0305158017 -0.0528123646 -0.0344150942  0.0436883923 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0013431676  0.0227060423 -0.0507072557  0.0247222457 -0.0236425858 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -0.0540650452  0.0702195915 -0.0121670788  0.0059833968 -0.0447484606 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.0320583238 -0.0068431383 -0.0499488776  0.0860451208  0.0279232376 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.0722993771  0.0390112348  0.0487262306 -0.0376982309 -0.0258073255 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  0.0806736985 -0.0015597325 -0.0802557482 -0.0358863159 -0.0099357914 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.0147510903 -0.1187673502 -0.1002090461  0.0152493080 -0.0427680923 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -0.0631700671  0.0643622819  0.1381054108 -0.1218893855 -0.0427068689 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  0.0773533789 -0.0199320214  0.0361057754  0.1318237045 -0.0192023009 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -0.0440214618 -0.0878400630  0.0211161230  0.0608575777  0.1571572112 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -0.1672491014  0.1317249589  0.0132462688 -0.0590850644  0.0056287100 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -0.0935210943  0.1045112684 -0.0696276040 -0.0042761060  0.2139628771 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -0.0434841351  0.0034339546  0.0948561947 -0.0359499266 -0.0771774245 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  0.0342096408 -0.0820489172  0.0594682514 -0.0046154277  0.0344331404 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -0.0456492487 -0.0344560082 -0.1099033281  0.0375089014  0.0923453774 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -0.1749477300  0.0334631963  0.0328372003  0.0797249626  0.1744440451 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  0.0970433612  0.0953961268 -0.0092195330  0.0526321101 -0.1458610521 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0020206992  0.0107442958 -0.0118923967  0.0264432402 -0.0602604910 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -0.0880725824  0.0026704588 -0.0917313079  0.0289781327 -0.0841054750 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.1994842154 -0.0023855245 -0.0053683619  0.1733836437 -0.1394589917 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.1212334534  0.0109786361  0.0626689114 -0.0603094933  0.0408830595 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0249647839  0.1276344264 -0.0166636346  0.0183442337 -0.0540375570 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -0.0606000174  0.0476597520 -0.0208553195  0.0363353776 -0.0984197222 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.0627304418  0.0886925296 -0.0267737920  0.0327872133 -0.0277056693 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  0.0686994882 -0.0197981929  0.0410141689  0.1221284400 -0.0571689570 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0434734967  0.0169403499 -0.0908197411  0.0096615404  0.0482352219 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.0417858991  0.0150505523 -0.0001564481 -0.0720720979  0.1134398403 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.0336997480  0.0568701950 -0.0494494683  0.0403311242  0.1316208719 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.0752096619 -0.0135269965  0.1247768033 -0.0137953429 -0.1341630341 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -0.0034806548  0.0716091535  0.0862939529  0.0572995419  0.0809043382 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  0.1778064096 -0.1179013131 -0.0416796564 -0.0628454767 -0.0904761490 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0483667504  0.0506831481 -0.0355517066 -0.0875200970 -0.0345568166 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0146763078 -0.0398686136 -0.1625007000 -0.0386026040  0.1482980360 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0952094930  0.0520573886  0.0526289575 -0.0188158278 -0.0741099567 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0288364041 -0.0233978075 -0.0586720890  0.0075644697  0.0998868487 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.1507697228  0.0226875513  0.0558470194 -0.0539736257 -0.0151979208 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  0.0037410758  0.0508209819  0.0051257852 -0.0121751516 -0.0133297228 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -0.0375520856 -0.0307467566 -0.0618643562  0.0280701486 -0.0703777052 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  0.0100790834 -0.0958346043  0.0234769540  0.0803581506  0.0014819792 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.1207275266  0.1069731653  0.0520054739 -0.1093002077  0.0618131897 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.1661330559 -0.1617366283  0.0221429196 -0.0615521293 -0.1407578064 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#> -0.1335669275  0.1330697050  0.0592565927  0.0491458997 -0.1033872640 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.1066791165 -0.0939415608 -0.1249851280  0.0565755825  0.0587492567 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0317532085  0.0048631277 -0.0034006560  0.0154745069 -0.0469609587 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.1243860962  0.0008069626  0.0293691647