Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  0.0845798815 -0.0098958841  0.0792443746 -0.0600752141  0.1096522060 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0494713705  0.0027703224  0.0343988804  0.0485523302 -0.0278193582 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0216326781  0.0796350673  0.2445242942 -0.0151599458  0.0351119385 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.0181655935 -0.0152628755  0.0887383658  0.0667127943  0.1137940027 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0452385769 -0.0008785879  0.1025971739  0.0779692654  0.0153753333 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.1095998296 -0.1130477936 -0.0882977416  0.0967281976 -0.0251553548 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0670370319  0.0741832624 -0.0917649529 -0.1141166087  0.0236387294 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.0064355312  0.1073775654  0.0111431854 -0.0886695629  0.0628651141 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -0.1340316955  0.0694097201  0.0736988002 -0.1066842621  0.0835322468 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0186786404  0.0343859649 -0.0582931200 -0.0980549848  0.1338914192 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0047399041  0.1123969126 -0.0872354787 -0.0977305641  0.1464926574 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.0485347936 -0.0255183976  0.0448360021  0.0375627846 -0.0496696637 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0395235617 -0.0178006283 -0.0191831428 -0.0512117665 -0.1106685806 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0589646651  0.0456752314  0.1091156278 -0.0455543769 -0.0465177705 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -0.0012985196 -0.0557660227 -0.0324459351 -0.0092665721  0.0088205663 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0185333032  0.0437354443 -0.0576522169  0.0159385139  0.0146147772 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -0.1082827000 -0.0321303380 -0.0935200357 -0.0937675784  0.0190457459 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0325883721 -0.0075621561 -0.0280856613  0.0291666940  0.1122955606 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -0.0662878280 -0.0526373531 -0.0738243955  0.0926477872 -0.0883352651 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.0016239202  0.0054512078 -0.0500433339  0.0023288306 -0.0081039845 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -0.0536048714  0.0424004495 -0.0347267539  0.0255912968  0.0788078328 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -0.0190191818 -0.0529145790 -0.0424126846 -0.0136488479 -0.1131115206 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.0212646050  0.1035417618 -0.0085780053  0.0324968447  0.0954711780 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.0540099369 -0.0151921038 -0.0148451441  0.0721519046 -0.1241933967 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  0.0279125869 -0.0476961170 -0.1228835428  0.0187776299 -0.0609433445 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0922065908  0.1939528325  0.0665428657  0.1076117377 -0.0915060741 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0069622970 -0.1287355241 -0.0345855289 -0.1224066050 -0.0473193497 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0779963299 -0.0121740727 -0.1055324399 -0.0183057766  0.0661787240 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0494154850  0.0576067017 -0.0321281093  0.0711705281  0.0785391805 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.0176832257 -0.0061957353 -0.0855747386 -0.0376913449 -0.0046831731 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  0.0944544846  0.0080598938  0.0270914113  0.0230125087 -0.0633640568 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.0483797994 -0.0215271987  0.0684462578  0.0414797699  0.0127385595 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0389676961 -0.0408605450  0.1220162881  0.0788432048  0.0506604616 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0496536914 -0.1519299540  0.0483663058  0.0073824975 -0.0117391613 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0765438760 -0.0648885425  0.0542575317  0.0500169381  0.0252462961 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0633464401 -0.0778583852 -0.0950407575 -0.0333797730 -0.0052143832 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0381550529 -0.0425322474  0.0277871801  0.1199522903  0.0262228313 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  0.0119172280  0.0440744476 -0.0065443458 -0.0999249434  0.1480872199 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.0841014216  0.1153056330 -0.1354739589 -0.1623462542 -0.0086417946 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.0547314509  0.0407452592  0.0422437804 -0.1380610254  0.0577697616 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.0434454084  0.0447584721 -0.0516171875 -0.0412916045 -0.0659489140 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.0434210502  0.0103881253  0.0301385763  0.0018247144  0.0098333452 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -0.0570755115  0.0066707884 -0.0245051085  0.0815913625 -0.1941383726 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0207702420  0.1052506880 -0.0653513165 -0.0131488725  0.0899228572 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  0.0306913420 -0.0291128669  0.0202454431  0.0815546773 -0.0395169032 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  0.0192585716  0.0911261655 -0.1111309462 -0.0172794701 -0.0252989646 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0594456699  0.0871222171  0.0542546446 -0.0910762773  0.0164275172 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0459174470  0.0048146723 -0.0216930503 -0.0239284861  0.0344141927 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  0.0093050575 -0.0645127529 -0.0592197751 -0.0361745610  0.2087974147 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.0468222524 -0.0391592927 -0.0323788128 -0.0742831967 -0.0499337222 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  0.0448715836  0.0710389846 -0.0593287384  0.0426044693 -0.0216261959 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.0590811703 -0.0317854381  0.0521658789 -0.0683090741 -0.0855853721 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.0308199430 -0.0462899236  0.0503230810 -0.0889981736  0.0459578029 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.0795329676 -0.0073871009 -0.0531523837  0.0459378071  0.0138100626 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0677497289  0.0957787572 -0.0745792494  0.0462640444 -0.0108781257 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0507438032  0.0005220668 -0.0052591268  0.0553061476 -0.0370959426 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -0.0577660862  0.0332690243  0.0718685253 -0.1422041846 -0.0518546094 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0427522040  0.0677653610  0.0510539189 -0.0206595884 -0.0441769333 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0789836550 -0.0004598660  0.0632975374  0.1894121084  0.0346690013 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0443838826 -0.0356232618  0.0749675126  0.0723435708  0.0758316535 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -0.0241298346 -0.0092487568  0.1096220838  0.1247352332 -0.0190014178 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0168043699  0.0235206530  0.0337743315 -0.0534468292 -0.0652701872 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -0.1299123655 -0.0555564266 -0.0721897664  0.0757710805 -0.1001935190 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#> -0.0152434673 -0.0884162962  0.0786404170 -0.0299396682  0.0738357574 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  0.0234219857 -0.0030303514 -0.0215569043 -0.0385706101  0.0048652581 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.0500702326  0.0221314605 -0.1365293585  0.0264722313 -0.1248121036 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0015724113  0.0733771354  0.1234940587  0.0450756211  0.0312149163 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.1357764141  0.0181959093  0.0236022159 -0.0258154573 -0.0285423045 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.0206293164  0.0303089285 -0.0190086654  0.1047534357 -0.0533936875 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  0.0687369119 -0.0264045826 -0.0005332225  0.0992348164  0.0587355846 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0325450092 -0.0458174456 -0.0227949018 -0.0916832018  0.0679908400 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  0.0030359010 -0.0545159497  0.0883711222  0.0486993066  0.0229248132 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.1501968290  0.1205926359 -0.0529265775  0.0374877980 -0.0642951003 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.0068490607  0.0083966071  0.0283183667 -0.1162346007  0.0784437395 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0199664235 -0.0365220020 -0.0999597354  0.0717529982 -0.0048705805 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.0288508066 -0.1224892427  0.0193275287 -0.0456331278  0.0810641089 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -0.0088813268  0.1132040659  0.0317626266  0.0416257552 -0.0530666158 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0448708546  0.0467133370  0.0238964074  0.0555849113 -0.0345464930 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -0.0881441550  0.0630318801 -0.0517487872 -0.1105225777 -0.0293040021 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0619434569  0.1571434125  0.1827482885  0.1257438775 -0.0406438423 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.0480765383 -0.0321334962  0.1011394607 -0.0463339640  0.1692388226 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0351678973  0.0904051923 -0.2008625657 -0.1050033624 -0.0367487634 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.0073205672 -0.0526093453  0.0375870865  0.0384423278 -0.0668397458 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0122716631 -0.0555274005 -0.0296997897 -0.1157848471 -0.0161776962 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.0402477161  0.0227784890  0.0546311532  0.0128465208 -0.0479126496 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.0398490095  0.0682910584  0.0061129195  0.1343414413  0.0955996025 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.0030114929  0.0851346359  0.0031951391 -0.1151064645 -0.1410927833 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.0331306260  0.0212046460  0.1344999131  0.0121732922  0.0977080381 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0764748099 -0.0636841986  0.0399686321  0.0665956836 -0.0076556817 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0654207056 -0.0146496113  0.0578301331