Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0227034156  0.1206802014  0.1810153037  0.0635391283  0.0016185517 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0250103971  0.1585670593 -0.1374053351 -0.1016240512 -0.0330562508 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0250828063  0.0195483033 -0.0831932423 -0.0307320419  0.0641057190 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.0313062862  0.0702642611  0.0083412378 -0.0395673595 -0.0662989701 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0025637374 -0.0839719185  0.1502017041  0.0736950656  0.0841601311 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.1417135417  0.1072458638 -0.0662899910 -0.0304959465 -0.0276631472 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  0.0134621557 -0.0258238549  0.0136920262 -0.0467704552  0.1331028290 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -0.0316276759 -0.0951639026  0.0277927392 -0.1291798738 -0.0721343744 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -0.0451326112 -0.0088715744 -0.0007339706  0.0857302052  0.1608621513 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0523741322 -0.1052698334  0.0006850443 -0.0086828894  0.0107792764 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.1603889788  0.0089936796  0.0242563888 -0.0288272479  0.0135484671 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -0.0200497779 -0.0244115855 -0.0011604580  0.0453561916 -0.0007226951 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0419666927 -0.0669456121 -0.0819385849  0.0305110795  0.0214678895 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0548004220 -0.0900941072 -0.0592571859 -0.1041093514  0.0149254639 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -0.1160260082  0.0567021904  0.1614509869  0.0316924746 -0.0347413517 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0046380970 -0.1601339344 -0.0193242741  0.1602243446  0.0546079459 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -0.0533685829  0.1430416928 -0.0724983051  0.0321093467  0.1550511282 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.0516795085  0.0174623751 -0.0577190337  0.1515404045  0.0616209273 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0283945007  0.0548418474 -0.0382784978 -0.0793627099 -0.1001763276 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#> -0.1059991139 -0.0329118410 -0.0263492698  0.0387794645  0.0213404566 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.1275876075  0.0720400203  0.0239919979 -0.0906739480 -0.1102926156 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.0705478085  0.0116528270  0.0882388805  0.0261416562  0.0671659382 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -0.0397453167 -0.0451668292  0.0427667488 -0.0371939069  0.0787465086 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.0055443743  0.0104405059 -0.0511074976 -0.0144188722 -0.0269920608 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  0.1582370505 -0.0620599123  0.0910846911  0.0077701710 -0.1243914072 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  0.0065883038 -0.0890076798  0.0146428002  0.1187725262  0.1639452170 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0629308435 -0.1089747295  0.0107787792 -0.0181750124 -0.1317540129 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  0.1249481752 -0.0310636135  0.0140380186  0.0182576057  0.0936407769 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0421684675  0.0074337499 -0.0826855526  0.0494360721 -0.1683040489 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  0.0517213777  0.1136054507 -0.0107059820  0.0193387952 -0.1006169786 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0508629557 -0.1701710761  0.0545288843 -0.0332522592  0.0093192371 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -0.0462697986  0.0360980430  0.0330589545  0.1124142003  0.0907402967 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0364326647  0.0310830265 -0.0157987273  0.0143984848 -0.1002216319 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -0.0095140948 -0.0171517833 -0.0383292323  0.0675738782 -0.0956826885 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#> -0.0736927522 -0.1605924723 -0.0098496227 -0.0004164559  0.0061247678 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#>  0.0443548990 -0.0228466421  0.0537922395 -0.0405818157  0.0335134512 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0791676112  0.0217226942 -0.0625362095  0.0011581880 -0.0119790400 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  0.0332753576 -0.0581262446  0.0273016512 -0.0070957068 -0.0772132673 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -0.0022913243 -0.0374056354 -0.0666261236  0.0524644493  0.0072916557 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  0.0820463688  0.0016813939  0.0902959937  0.0693268651 -0.0659107539 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.0007368610 -0.0072941259  0.0374854205 -0.0677124679 -0.0555686397 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.0508919015 -0.0091872433 -0.1184469555 -0.0339875368  0.0300624388 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.0475504176 -0.0818092012 -0.1047425550 -0.1078621170 -0.0837348123 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0123298586  0.0632918239 -0.0098291702 -0.0038027444 -0.0539209413 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.0573236168 -0.0912709762 -0.1952746424  0.0353612548  0.0385847219 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.0269145951 -0.0662077825  0.0297811433 -0.0195785604 -0.1768626467 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0685509309  0.0505623418  0.0065047357  0.0370574351  0.0093692154 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0281028072  0.1121943367  0.0209663917 -0.0226458192  0.0145366771 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  0.0585858606  0.0028319530  0.1020850556  0.1356490041  0.0347521790 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  0.0312971498 -0.1588208915  0.1100424387  0.0350907026  0.0862477042 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  0.0800419177  0.0083380580  0.0143934094 -0.0266248044 -0.0543833431 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.0130249352 -0.0822605387  0.0004904340 -0.0847563007  0.0675769261 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.0977471735  0.0454493432  0.0748399833  0.0760734344 -0.0745513191 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  0.1093541474  0.0391818791 -0.1808410375 -0.0287628309  0.0407042523 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -0.0549810303 -0.0775978346 -0.0147444763 -0.0526101873  0.0142629611 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0485420869 -0.0010864842  0.0444903363  0.0393275287 -0.0280440894 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.0238546558  0.0851335928  0.0408436728  0.0471053791  0.0166875298 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0550740869 -0.0361375827  0.0341011237  0.0724929667 -0.0588673358 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  0.0827671190 -0.0956538537 -0.0440372863  0.1179242275 -0.0678176396 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.1143419053  0.0401036883  0.0355736206  0.0905450675 -0.0863785572 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0405286816  0.0314210606  0.0828236639  0.0618344637  0.0873010682 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0903820867 -0.0912628723 -0.1001026217  0.0961575829 -0.0020995033 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -0.0703916590 -0.0484201073  0.0873647552 -0.0364266660  0.0149253406 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0832046583  0.0796487195 -0.0717709648 -0.0059737330  0.0272008519 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  0.0392083161  0.0794591100 -0.1037713775 -0.0632125569 -0.0595160174 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -0.0885906362 -0.0021931538  0.0117570936 -0.0488530157 -0.1723601603 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  0.1253629997  0.0599517803 -0.0100083288 -0.1127248450  0.0299163986 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -0.0094159960  0.0764079931 -0.0406084495  0.0150332191 -0.0083584072 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.0569448861  0.0333103881  0.0829520715 -0.0070150506  0.1393740174 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  0.0821649421  0.0438251947  0.0649410867 -0.0157631208 -0.0004094925 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0922734585  0.0811259352 -0.0264705778 -0.0213488787 -0.0902132459 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.0017640324  0.0498729656  0.0319826349  0.0465822999 -0.0937583804 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.0161982304 -0.1337181902 -0.0416714105 -0.1318840672  0.0219082934 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  0.0354113409 -0.0870720777  0.0576355370  0.0155906163  0.0073515053 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0407732539  0.0059270053 -0.0044289873  0.0005476923 -0.0697329673 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  0.0067131688  0.0387358612  0.0779065370  0.0406108615  0.0989436377 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -0.1238098371  0.0044912892  0.0373769961 -0.0078336084  0.1430315417 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0880564560 -0.0142157830 -0.0274915958  0.0524267901  0.0112989786 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0679102546  0.0048154184  0.0214685254  0.1593817318 -0.0898173387 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  0.0253914261 -0.1318013239 -0.1033898872  0.0477634412 -0.0178353485 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.0042654403  0.0016133368 -0.0995115741  0.0986300578  0.0507524698 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0131745073  0.0428080625 -0.0464979891 -0.0113292004 -0.0518618547 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -0.0252222330 -0.1444298893  0.0668732837  0.0295387792 -0.0667206633 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0341089745  0.0036595884  0.0668828281  0.0228282298  0.0083072337 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -0.1196677155 -0.1076005962  0.0318807620  0.0863129780 -0.1083078924 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -0.1033120137  0.1293548608 -0.0276453528  0.0651902472 -0.1071883902 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.1366425463  0.0776052600 -0.0203794452  0.1013835057  0.0359903931 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -0.1148258609 -0.0531496083  0.0584353425  0.0224984195  0.1062457054 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0779694576 -0.0205530123  0.0239790783  0.0850616320  0.0916189518 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  0.0786963252  0.0887566804  0.0724266455