Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -4.549461e-02  2.008471e-02 -1.974745e-02 -3.181312e-02  5.760394e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  3.017448e-02  1.847630e-02  1.227253e-01 -1.249771e-01  4.478875e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  6.996702e-02 -2.197847e-02 -8.003105e-02 -1.381499e-02 -4.919832e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  4.132723e-02 -2.964089e-02  3.943260e-02 -9.474064e-02 -1.719739e-01 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -8.184476e-03  4.089179e-02 -5.674969e-02  6.321560e-02  9.132186e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  6.527389e-02 -5.810114e-02 -1.728337e-01  3.805646e-02  1.050637e-01 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -7.349052e-03 -1.874590e-02  1.003586e-02  4.818738e-02  3.768403e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -6.423740e-02 -5.816186e-02 -4.120585e-02  3.283408e-02  3.669808e-02 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -2.386637e-03  2.291140e-02 -2.345975e-02 -3.291573e-02 -4.971840e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -1.275498e-02 -1.685960e-04  5.573340e-02 -2.683721e-02 -8.059831e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -6.714024e-02  1.106033e-02  2.267647e-02  6.892942e-02  9.907288e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  1.513627e-02 -4.064951e-02 -4.799892e-02  1.527916e-03  5.708088e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -1.015470e-01 -5.392518e-02  6.552763e-02  7.498544e-02  6.015953e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -5.884124e-02  2.105628e-02  9.798342e-03  5.895295e-02 -1.196560e-01 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -8.904138e-03 -1.281238e-01 -4.706939e-02 -1.417176e-01  4.835515e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  1.247566e-03  1.307175e-01  5.203474e-02 -3.162319e-02 -1.121280e-01 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -2.220768e-02 -2.650872e-02 -4.150918e-02 -2.970752e-02  2.470109e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -6.912351e-02 -1.024031e-01  1.092256e-02 -4.156235e-02  1.721297e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  7.191092e-02  1.042133e-01  1.020908e-01  8.796948e-05  1.469563e-01 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  3.876516e-02  9.730745e-02  5.314059e-02  1.235450e-02  5.664613e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  4.753646e-02  8.735467e-02 -9.894140e-03 -4.891726e-02  7.529597e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -1.132246e-01 -6.893104e-02  2.599247e-02 -1.043008e-02  9.906703e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  5.079842e-02  1.053360e-02 -1.326796e-03 -2.498035e-02 -1.730659e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -6.043810e-02 -7.474080e-02 -4.316949e-02 -4.481662e-02  6.710917e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  2.647237e-02 -5.571264e-02 -5.912608e-02 -6.960323e-02 -8.100584e-03 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  5.625077e-02 -2.547967e-02 -3.036097e-02 -3.297676e-02 -1.004572e-01 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  1.242552e-02 -1.468414e-01 -2.275478e-02 -5.125305e-02 -1.985286e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -8.587982e-02 -3.850694e-02 -2.254055e-02  8.930212e-02 -1.288265e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -5.074185e-02  8.554307e-02  6.025139e-03 -1.175364e-01  1.309838e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  4.288632e-02  3.072868e-03 -1.704067e-01 -1.698653e-02 -5.001304e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -7.137179e-02  2.391696e-02  2.887178e-02  1.011104e-01 -2.746604e-03 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -5.236541e-03  1.067459e-01  4.231753e-02 -9.779916e-02 -1.775578e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -1.623418e-02 -9.590314e-02 -1.001899e-01  3.537558e-02  6.671978e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  5.479690e-02 -2.599577e-02 -2.711717e-03  1.202791e-01  6.518706e-03 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  3.326652e-02  8.517849e-02  7.837860e-02  1.406894e-01  7.196888e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -6.421538e-02  1.212382e-02  4.888615e-02  2.588137e-02 -5.933234e-03 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  1.030175e-01  8.838852e-02  5.873947e-02 -6.647252e-02 -6.423293e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -4.461698e-02  9.386762e-02  1.787882e-02 -4.769743e-02 -1.591245e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  6.861414e-02 -2.286639e-02 -1.720727e-02 -8.613956e-02  3.259983e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  7.868649e-02  1.467331e-02 -6.193909e-02 -1.644696e-02  2.908543e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -3.406339e-02  1.138966e-01  1.072941e-02  1.088473e-01 -1.628846e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -1.020726e-01  8.541601e-02  4.693401e-02 -2.160991e-02  1.104130e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -6.310814e-04  8.950007e-03  1.643762e-03 -2.225563e-03 -1.684657e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  1.486643e-02 -3.967950e-02  3.852890e-02 -2.448457e-02 -8.775051e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -2.010792e-02  4.465027e-02 -2.539733e-02 -2.367816e-02  7.297487e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  1.113554e-01 -2.307618e-02  1.230590e-01  2.263477e-01  3.839347e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  2.158944e-02  5.456604e-03  1.595224e-02  1.209458e-02  4.198015e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -6.448061e-03 -7.688940e-04 -1.765692e-02 -3.761533e-02  4.826453e-02 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -2.675031e-02  3.185020e-02  7.934528e-03  7.993727e-02 -8.925502e-02 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -2.661741e-02  4.183756e-03  6.208879e-02  3.072228e-03 -4.318464e-06 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -1.211027e-02 -7.217358e-03  1.132520e-02 -1.776277e-02  1.272227e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -5.365003e-02  8.952166e-05 -1.057169e-01 -1.337740e-02  7.122772e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  9.292454e-02 -1.219511e-02  1.922178e-02  7.597684e-02  9.513945e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  2.652314e-02  4.255903e-02 -1.396097e-02  3.973649e-02 -4.562592e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  4.349372e-02 -7.198210e-03 -8.744847e-02  9.379283e-02  9.911267e-03 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -8.128824e-02  1.786712e-02 -5.271547e-02  3.812542e-03 -2.907765e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  2.685813e-02  1.027077e-01  8.196318e-02 -9.398854e-02 -9.752424e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -1.894673e-02 -6.590985e-02  1.378866e-01 -1.412307e-01  1.484479e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -6.877446e-02 -1.039927e-01  8.238344e-02 -4.519985e-02 -4.493457e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#> -6.532024e-02 -7.741305e-02  8.865810e-02 -1.387004e-02  5.267623e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -4.959345e-02 -1.139778e-02 -1.063744e-01  4.648568e-02  3.911119e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -1.876952e-02 -1.106814e-01 -4.208568e-02 -7.400154e-02  4.070320e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -2.990661e-02 -4.916238e-02  4.691605e-02  6.106009e-02  2.159674e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  5.650951e-02 -4.044846e-02  1.774140e-02  1.007205e-01  6.533358e-02 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  6.286415e-02 -7.078931e-02  8.558723e-02  5.684001e-03 -4.378894e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  5.407915e-02  2.140238e-02  1.026037e-01  5.850584e-02  8.561274e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -1.434584e-01  1.474002e-02 -6.226726e-02 -1.131165e-01  8.150673e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  6.899205e-02 -9.113635e-02  1.517674e-02 -1.874467e-01 -2.641475e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#> -5.756118e-02  8.500987e-03 -1.474097e-02 -9.778151e-02  1.403493e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  5.242823e-02  1.934639e-02 -3.286553e-02  3.636396e-02 -8.819150e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -2.519508e-02  3.178222e-02 -3.921388e-03  1.781137e-01 -1.874695e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -1.376709e-01 -5.030179e-02  6.253160e-02  6.514218e-02 -1.952865e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -5.399372e-02 -6.364468e-03 -3.778701e-02  8.162342e-02 -2.958099e-04 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  6.975612e-02  7.272635e-03 -4.451119e-02 -8.256301e-02  5.905294e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  1.175628e-01 -3.404641e-02 -1.341353e-01 -7.667669e-02 -1.672200e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  1.887510e-02  1.918709e-01  4.179590e-02 -4.159884e-02 -7.909560e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -1.228103e-03  6.329971e-02 -1.012142e-01  3.931147e-02 -1.342398e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  1.984488e-01 -9.632396e-03  1.907207e-02 -2.139337e-02 -3.318279e-03 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -5.700017e-02  6.032010e-02 -3.315795e-02  3.530771e-02  2.113162e-03 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -1.252243e-02 -3.400020e-02  8.213717e-02 -3.706366e-02  1.257267e-01 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  1.709268e-01  8.708859e-02  2.988276e-02 -1.272872e-01  7.457753e-02 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  8.088364e-02  7.972356e-02 -2.534839e-02 -1.079269e-01  1.218903e-01 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -3.791191e-02  1.705223e-02  2.793129e-02  1.287123e-01  2.816503e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -4.084358e-02 -4.530686e-02  4.141050e-02  3.848884e-03 -7.320289e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  5.685366e-02  1.092002e-01  1.000349e-01 -6.752139e-03  6.399317e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  2.321025e-02 -3.763259e-03  6.225306e-02  4.573166e-02 -6.868398e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  3.628085e-02  4.654871e-02  1.318757e-01 -4.662123e-02  4.537769e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -1.298361e-01  1.214735e-02  1.657559e-02 -5.488354e-02  1.496371e-01 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#>  3.021484e-02 -1.058801e-02  1.415246e-01 -4.419638e-02 -2.749757e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -9.159417e-02 -3.157580e-03 -4.781981e-02