Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -5.366801e-02  1.604241e-01 -1.112817e-01  5.301101e-02  7.571621e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -5.029407e-02  7.449981e-02  8.647345e-02 -3.532886e-02 -6.518342e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -3.662939e-02  6.741048e-02 -4.099374e-02 -6.678246e-02  3.537840e-02 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  2.988191e-02  5.625277e-02  1.337820e-02 -3.841141e-02  7.624730e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -1.576572e-02  1.274621e-01 -1.407006e-02 -4.254893e-02 -1.012835e-02 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  2.635571e-02 -2.935979e-02  3.763383e-02  7.433525e-02  5.529973e-02 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -8.318153e-02  7.302757e-02  7.134821e-02  6.028798e-02 -7.184591e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -5.244825e-02 -7.397715e-02  3.419136e-03 -5.606745e-04  1.328117e-01 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -2.003714e-02 -6.681601e-02  4.774458e-02  1.320488e-02  9.799729e-02 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  1.804511e-02  2.469505e-02  5.716385e-02 -2.293213e-02 -6.283779e-02 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -3.994681e-02  7.003294e-02  3.546599e-02  3.110750e-02 -2.348588e-02 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  5.992407e-02  7.953148e-04 -9.600778e-02  6.106545e-03 -1.090427e-01 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  6.951826e-02 -1.599836e-02  2.651191e-02 -2.386743e-02  8.066181e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -4.012454e-02  9.719506e-02 -1.053783e-01 -5.348289e-02  1.519824e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -9.640240e-03  1.003744e-01 -8.582451e-02  8.350354e-02 -7.770051e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -1.102664e-02 -5.094703e-02  5.474522e-02  1.240206e-01 -5.100228e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  1.302322e-01 -6.652905e-02  1.998435e-02 -5.076705e-02 -4.262833e-04 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  3.619694e-02 -2.127940e-02 -3.883513e-02 -3.617565e-02 -5.386475e-02 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -8.953084e-02 -2.197923e-02 -1.480685e-01 -8.922170e-03 -3.604547e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  4.752708e-04  4.279352e-02  5.848371e-02 -4.470023e-02 -2.446537e-02 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#> -3.778646e-02 -2.460887e-02  3.116412e-03 -7.820859e-02  3.343497e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -4.091644e-02  1.353430e-01  5.000884e-02 -4.371166e-02 -3.319073e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  2.776083e-02  5.321210e-03  5.750006e-03 -7.929202e-02  7.468419e-02 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -1.155773e-01  7.152156e-02 -4.027380e-03 -6.621957e-02  2.021902e-03 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -5.608837e-02 -7.577482e-02 -4.578933e-02 -4.568714e-02 -4.242702e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#>  1.848837e-02 -6.006415e-02  5.964501e-03 -2.786749e-02  1.013563e-01 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -4.533005e-02 -5.497400e-02 -8.041296e-02  8.548676e-02  9.115645e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#>  1.632896e-02 -8.936057e-02 -6.940367e-02  3.290999e-02  6.049816e-03 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -3.891763e-02 -5.519582e-02  6.815598e-02  3.150036e-02 -3.919018e-02 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -4.369519e-02  1.610036e-01  2.047218e-03 -1.379083e-01 -1.515689e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  6.198078e-02  4.560637e-02 -9.897892e-03  4.158527e-02  3.460376e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  3.442006e-02  6.260461e-02  2.807619e-02 -1.097107e-01 -6.844590e-02 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  8.535527e-03 -6.595346e-02  8.355050e-02  5.344664e-02 -1.736984e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -3.311423e-02  1.955401e-03  6.262955e-02  7.117168e-02 -4.512629e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  7.771399e-02 -3.334386e-02 -7.078501e-02 -2.734458e-02  6.238558e-02 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -1.279991e-02  5.641707e-02 -9.955965e-02 -5.465200e-02  1.055326e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -6.239723e-02 -4.764158e-02  1.537606e-01 -2.478182e-02 -6.394799e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  2.858787e-02 -2.716335e-02 -3.628118e-02 -3.204849e-02  8.924515e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  7.898017e-02  4.998447e-02  8.081590e-02  6.867763e-02  7.221842e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -1.183659e-02 -2.265991e-02  7.068280e-02  9.734524e-02  7.057982e-02 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -2.248077e-02 -6.574533e-02  4.275986e-02  2.889392e-02  2.520110e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  5.744768e-02  7.681521e-02 -2.415849e-02 -5.846830e-02  3.723611e-02 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  6.889285e-02  2.932430e-02  5.499210e-02 -3.931841e-02  1.693306e-04 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  2.266768e-02 -2.217341e-02 -3.115840e-02 -7.138262e-02  2.777614e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -8.414151e-02  5.920899e-02  5.859404e-02  2.868404e-02 -5.578258e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -5.331462e-02  1.821652e-03 -6.197031e-02 -2.221410e-02 -6.442190e-03 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -1.013745e-01  1.202248e-02  2.095288e-02 -4.640632e-02  1.570038e-01 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  9.283991e-02 -1.071024e-02 -2.985467e-02 -1.277060e-01  1.426997e-01 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  5.276929e-02  8.484832e-02 -1.316018e-01  5.825094e-02  4.513089e-03 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -6.753835e-02  8.047177e-02  1.447776e-02  9.568140e-02  8.345939e-02 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -6.344774e-02  3.248163e-02  1.781590e-01  2.723566e-02  6.302262e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  2.106184e-02  5.126925e-02  8.763064e-02  2.606161e-02 -3.532389e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  1.285518e-01 -1.822247e-02  6.656105e-02  3.145865e-02  2.198760e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -1.161151e-01 -2.659579e-02  4.621891e-02  2.979737e-02  3.022856e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -8.013765e-03  6.537273e-02 -6.874120e-02  4.936601e-02  9.992394e-03 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -8.225861e-02 -1.286619e-02  5.558559e-02  3.638040e-02  1.093074e-01 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#> -6.253488e-02 -1.311094e-02  1.129120e-01  1.855797e-01 -1.076689e-01 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  7.734381e-02 -3.250968e-02 -1.457511e-01  6.848222e-02 -4.845262e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  5.408350e-02  2.057433e-02 -2.498926e-02  6.174129e-04  5.269167e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  4.275414e-02 -5.142461e-02 -1.353135e-01  2.392947e-02 -9.642338e-03 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  1.544266e-02  7.324687e-03  1.413309e-01  1.562339e-01  8.362387e-02 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  1.669286e-02 -4.440702e-02  1.999851e-02  1.994018e-02 -4.407179e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  6.849602e-03  6.081787e-02  8.522197e-02 -1.111563e-02 -4.549693e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  6.872270e-02  3.450653e-02 -2.906642e-03 -4.108915e-02  1.053277e-01 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#>  7.353137e-02 -2.578749e-03  2.461973e-03  3.030721e-02 -4.523417e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  5.262277e-03  3.381628e-02  4.412923e-03 -6.397632e-02  2.271920e-02 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  1.143286e-01  9.424170e-02  3.980488e-02 -5.034264e-02  7.421320e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -2.146216e-03  1.169245e-01 -1.860292e-02  3.934403e-02 -6.803528e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  6.150769e-02 -8.778837e-03  4.567811e-02 -7.421903e-02 -2.301984e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  1.251227e-02 -6.357141e-02 -1.235009e-01 -3.088619e-03  1.252186e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#>  1.409766e-03 -4.793288e-02  6.149309e-02  9.933744e-02  1.648713e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -4.369320e-02  3.164173e-02 -2.956335e-02  1.793004e-02  1.083297e-02 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -5.823922e-03 -3.876520e-02  5.733408e-02  2.999889e-02 -1.067088e-01 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -1.070031e-01  8.259247e-02  4.284089e-02 -5.420570e-02  1.436705e-01 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -1.572628e-01  1.426988e-01 -6.036842e-02  4.772852e-02 -8.146590e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  1.352882e-01 -7.151583e-02 -4.275030e-02  8.342431e-02 -8.013368e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -1.362059e-02 -2.854477e-02 -2.352917e-01  4.536668e-02  3.306512e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -3.252201e-02 -7.119443e-02  3.944841e-02  7.742570e-02  6.476151e-03 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  3.921444e-02 -8.842803e-02 -7.117758e-02  9.736247e-02 -2.041332e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -8.314093e-02  2.440730e-02 -1.423160e-01 -2.011682e-02 -2.676153e-02 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -3.324274e-05  4.370506e-02  2.470042e-02 -9.659714e-02 -1.270058e-01 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  2.673960e-02 -7.098696e-02  3.045050e-02 -1.418066e-01  4.939527e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  4.464054e-02 -7.109137e-02  5.795036e-02 -4.106265e-02  3.854261e-03 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -6.654335e-02  2.184771e-02  1.687558e-01 -3.538475e-02  5.538053e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  1.650151e-02  4.691144e-02 -6.001890e-02  8.369262e-03  2.167716e-02 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  4.166708e-02 -8.644391e-02 -7.096165e-02  9.460582e-02  3.275074e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  1.042235e-01  1.189082e-01  1.918626e-02  7.170951e-02 -2.526679e-02 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  5.829983e-02 -8.058433e-02  2.909849e-02 -6.320501e-02  1.157313e-01 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -1.080561e-01 -6.799182e-02 -7.919794e-02  1.878550e-02  3.608863e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#>  6.898033e-02 -1.732461e-01 -3.459933e-04