Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  6.047907e-02  6.360311e-02  4.056132e-02 -2.244697e-03  1.814470e-02 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -1.185563e-01  4.538206e-02  6.744912e-02 -1.019893e-01 -1.001507e-02 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#>  1.250191e-02 -7.661711e-02 -8.733034e-02 -5.215079e-02 -1.142153e-01 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -8.496595e-02 -4.442943e-02  1.196872e-01 -2.073245e-02  4.887995e-02 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -2.853533e-02 -1.636027e-02 -4.166617e-02  9.162534e-04  1.156888e-01 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -8.741689e-03  3.209783e-02 -7.013873e-02  1.834855e-02 -1.223895e-01 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  8.485714e-02  6.255551e-02  5.354157e-02 -7.879023e-02  5.099464e-02 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  1.607434e-02  2.383587e-02  3.517513e-02  5.079194e-02 -3.171543e-03 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#> -4.610774e-02 -1.456884e-04  7.840682e-02 -1.393407e-02  2.826505e-03 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -1.490490e-02  8.826318e-03  2.390610e-02 -1.490611e-01 -6.297457e-03 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  1.026901e-01  9.577374e-02 -1.389951e-02 -1.305355e-01 -6.224183e-03 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  2.321485e-02 -1.129686e-01  3.961082e-03  5.001792e-02 -2.296536e-02 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  1.489522e-02  1.017058e-01 -2.601423e-02 -1.014605e-02  1.273221e-02 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -6.247968e-02 -6.131206e-04  5.148757e-02  2.525539e-03  3.859070e-02 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  3.186601e-02 -1.867851e-02 -7.173662e-02  3.132582e-02 -3.574470e-02 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  4.024420e-03 -2.148628e-02 -5.606717e-02 -7.516505e-02 -5.915695e-02 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -4.366934e-03  3.191279e-02  5.465991e-02 -3.996534e-02  4.530912e-02 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -6.014751e-03  5.357729e-02 -4.428216e-02 -3.472753e-02  4.482951e-03 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#> -3.743544e-02  2.200519e-02  1.869886e-02 -2.084370e-02  3.399873e-02 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  6.167779e-02  1.318996e-02  1.071112e-01  2.176426e-02  1.124442e-03 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  3.907338e-02 -8.861859e-03  5.656787e-02  9.929886e-02 -7.583594e-02 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  9.915348e-02  3.356925e-02  9.332142e-02 -3.833426e-02  2.891602e-02 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  1.005346e-01  6.371462e-02  7.337549e-02 -6.230428e-02 -8.686394e-03 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  1.264054e-01 -9.876704e-02  7.036221e-02  6.399880e-02  7.215251e-02 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  1.428571e-02 -2.702904e-02 -6.777119e-02  6.463218e-02  3.135894e-02 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -1.115276e-01 -1.055562e-01  7.141302e-02  1.996416e-02 -1.650800e-02 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -4.387452e-02  4.923093e-02  7.495270e-03  3.381839e-03 -4.649553e-02 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -9.891757e-03  6.594794e-02  8.189908e-03 -6.434742e-02  1.321003e-02 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#> -5.288070e-02  7.299946e-02  2.709327e-02  5.140878e-02 -3.475411e-03 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  4.135581e-03 -5.165911e-02 -1.363638e-02  9.875667e-02  3.657347e-02 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#>  1.363589e-01  6.044404e-02  6.487710e-02  5.544094e-02  2.112169e-02 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  2.908174e-03  6.919793e-02 -1.026224e-01  1.024518e-01  1.013131e-01 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  5.658003e-02  3.073450e-02  1.073866e-01  1.155272e-02  3.019293e-02 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -7.705416e-03  1.835928e-03 -5.199762e-03  4.856341e-02  2.678619e-02 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  3.596452e-02 -3.141878e-02  6.436976e-02  5.498476e-02 -5.170511e-03 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -1.138923e-02 -3.264098e-02  2.376319e-02 -4.869186e-02 -6.825954e-02 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  2.518490e-02  6.458258e-02 -1.997898e-02  2.522626e-03 -1.398912e-02 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -9.760351e-02 -1.451608e-01 -7.979212e-02  1.069289e-02 -2.312813e-02 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -9.628682e-02  9.583453e-02 -3.902103e-02  3.133551e-02  1.741421e-02 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  9.785653e-02  5.527591e-02  6.298752e-02  1.215184e-01 -1.020654e-01 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  1.579046e-02  1.002872e-02 -3.655306e-02  3.864259e-02  1.772539e-02 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -2.735700e-02 -4.193594e-02  2.135506e-02  1.706973e-02 -8.902392e-03 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -4.268173e-02  1.089692e-01 -7.659513e-02 -4.679962e-05  1.582021e-02 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  2.622998e-02  5.760253e-02  1.873369e-02 -8.097170e-02  4.969464e-02 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  3.960373e-02 -7.236639e-02  4.297921e-02 -9.992597e-03 -7.916213e-02 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  9.249812e-03 -5.958996e-02  3.813915e-02  6.184222e-02 -3.676258e-02 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#>  8.683524e-03  5.949943e-02 -1.832425e-02 -1.072501e-01  1.048539e-02 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -7.992982e-02 -7.521405e-02  6.203308e-02 -4.352945e-02 -1.645469e-01 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#>  6.916169e-02  3.213697e-04 -4.785327e-02 -5.084179e-02 -1.526391e-02 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -6.004518e-02  1.026787e-01  2.431815e-02 -2.993243e-02  4.175719e-03 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -7.726120e-02  2.888087e-02 -5.012024e-02 -9.939540e-02 -2.994491e-02 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  8.882491e-02  8.499092e-02 -1.977515e-02  1.322733e-01  6.265104e-02 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -5.418810e-03  1.944810e-04  9.530563e-02 -1.266563e-02 -1.795037e-02 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#>  7.686473e-02  9.465615e-02  3.579235e-02 -5.744426e-02  2.800303e-02 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  5.782914e-02  4.331100e-02  1.947195e-02  7.855209e-02 -9.651986e-02 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -3.090081e-02 -7.066425e-02 -2.303511e-02 -1.258169e-02 -9.006102e-02 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  5.729941e-02 -6.536860e-02  8.613209e-02  5.432777e-02 -7.027043e-02 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -1.153761e-02  6.340650e-03 -3.748591e-02 -1.827647e-02  8.630214e-02 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -4.731345e-03 -3.006473e-02  4.108684e-02 -5.363229e-02  1.671752e-02 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  1.152659e-01  1.265806e-01 -1.016016e-01  9.241598e-02 -7.227977e-02 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -4.020042e-02 -4.102888e-02 -6.582642e-02  1.929049e-02  1.005845e-01 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  1.231214e-01 -1.635014e-01  3.595294e-02  2.798769e-02  7.659032e-02 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -5.667839e-04  5.795556e-02  9.644499e-02  8.891979e-02  8.330429e-02 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  3.118180e-02 -1.095758e-02 -1.630351e-02 -3.731151e-02  1.109624e-01 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -9.140412e-02 -1.333838e-02 -1.996877e-03  1.419933e-02 -3.151302e-02 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  5.284060e-02  5.035818e-02 -3.543674e-02  1.750601e-01 -4.992860e-03 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  1.076291e-01 -8.327815e-02  1.331981e-02  4.397557e-03  1.524666e-02 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  5.408126e-02  7.471553e-02 -4.113197e-03  7.743337e-02 -4.251834e-02 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  4.166748e-02  5.577782e-02 -2.941699e-02 -1.320154e-02  5.787675e-02 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#>  7.789386e-02  9.318571e-05  8.722858e-02  1.410733e-02 -5.240068e-02 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -2.977033e-02 -1.162868e-01  6.101445e-02 -6.581586e-02 -6.613853e-02 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  6.889635e-02 -8.422438e-02 -2.775682e-02  3.542233e-02 -9.020745e-03 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  5.386053e-02 -1.861892e-02  3.007875e-02  4.771414e-02  9.108388e-02 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -3.672814e-03  5.708768e-02  1.704608e-01 -2.909365e-02  6.131912e-02 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -8.240916e-02  6.579466e-03  1.077777e-01  1.067444e-02 -4.612255e-02 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  6.133008e-02 -7.415770e-02 -3.260293e-02 -8.970229e-02 -5.254394e-02 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  4.307095e-02 -1.159049e-01  4.498267e-03 -4.431340e-02  1.005870e-02 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -2.662950e-02 -3.163580e-02 -1.343889e-01  4.636889e-03 -4.795793e-03 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -6.309957e-02 -1.527148e-02  2.160258e-03 -6.064426e-02 -8.385168e-02 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -3.771234e-02 -1.349864e-02 -6.332530e-02  8.989308e-04  1.684884e-01 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -6.639696e-02 -7.642161e-02  3.100679e-04 -4.475110e-03 -2.497444e-02 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -5.339271e-02  1.002028e-01 -6.719304e-02 -3.526655e-02 -9.721040e-02 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -7.116052e-02  7.542166e-02  2.065650e-02 -1.430173e-02 -2.880256e-02 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  7.879527e-02 -4.477038e-02 -8.160812e-02  9.661062e-03 -1.731500e-02 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#> -5.808657e-02 -4.368096e-02  6.324978e-02  4.413289e-02  8.595525e-03 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  2.577211e-02 -2.612465e-03  4.189043e-02 -8.684563e-02  4.373112e-02 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  1.232691e-01 -1.085961e-02 -6.526011e-02  8.383189e-02 -1.215782e-01 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -1.589083e-02  1.096467e-01 -3.445329e-02 -1.302256e-03  5.601184e-02 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -8.902023e-02  4.016040e-02 -5.754593e-02  6.579792e-02 -1.159829e-02 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -3.830111e-02  3.599590e-02  4.730539e-02