Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#>  0.0227736515  0.0774448827  0.0843250702 -0.0525377155 -0.0533167938 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -0.0024552939  0.0105530719  0.0966180313  0.1012443415 -0.0132536279 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0744229355  0.0801887186 -0.0540080968  0.0066185149 -0.1089787357 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  0.0038806839  0.0365884910  0.0899865961 -0.0844453329  0.0586591065 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#> -0.1241605654 -0.1276535797 -0.0741455188  0.0033733116  0.2136235871 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#> -0.0138076954 -0.0184644120 -0.0510677216  0.1042102973 -0.1006426383 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0156808602  0.0566630871  0.0464216557  0.0018999345  0.0274789650 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#>  0.0084879185 -0.0139134910  0.2219790710  0.0617899755  0.0462625606 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.0390190414 -0.0439504521 -0.0133317737 -0.0516823212  0.0112276819 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#> -0.0349908689 -0.0997229096 -0.0195147829 -0.0052691437  0.0007982061 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#> -0.0224833654  0.0237313766  0.0283070876  0.0824971702  0.0308603402 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#> -0.0245479241  0.0602846337 -0.0861316554  0.0815219002 -0.1916179840 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.1358999901 -0.0531306373  0.1350671696  0.0800887014 -0.0366435479 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0248932768  0.0028205782  0.0589770578 -0.0410128370 -0.0931022486 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#>  0.0020383410 -0.0030713280 -0.0323934595  0.1043103236 -0.0702134266 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#> -0.0011835622  0.0051196010  0.1735797554  0.0769403475  0.1249610933 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#> -0.0179734756  0.0329173803 -0.0053855059 -0.0805583224  0.0436954394 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#> -0.1101344662  0.1428307688  0.0420916472  0.0542008471  0.0321549538 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.1072184904 -0.0130752519 -0.0181472163 -0.1050374015 -0.1302884375 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  0.0566774647  0.0809448980  0.0150886175  0.0120223886  0.0030219526 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0352939324  0.0085965739  0.0246719218 -0.0595073188  0.1045637600 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.1013761021  0.0781534398 -0.0488952923 -0.0354322768 -0.0238744742 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -0.0255993979  0.0709862032  0.0033015197 -0.1392061386 -0.0308210865 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#>  0.0641652978 -0.0096287929  0.0172604187 -0.0930050478  0.0274867902 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  0.1253793562 -0.1009399461 -0.0481127757 -0.0585751902 -0.0449262860 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -0.0451158818 -0.0557961663 -0.1305278208  0.0185600556 -0.0437200050 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0361031933 -0.0078881698 -0.1036792348 -0.0042330287  0.0533539303 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0404099215  0.0043251197 -0.0752410965  0.0729564629  0.0785742366 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0253399285 -0.0524356070  0.0939831128  0.2179131226  0.0360795646 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.0413957347  0.0192510442  0.0546132480 -0.0031266696  0.0218902118 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0177411873 -0.0012966780 -0.0723547584 -0.1463462604  0.0541547261 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#> -0.0386148041  0.0111850560 -0.0815515623  0.1488175201  0.1199225423 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.1012884351  0.0226458290 -0.0806562326 -0.1140469155 -0.0136821539 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#> -0.0124651803  0.0407142605 -0.0523819583  0.0090227128 -0.1683107192 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0238290124  0.0042234751  0.1298809564 -0.0081874704  0.0467607161 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0589809128 -0.0022074312  0.0001442314  0.0897728093  0.0090799539 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#>  0.0098887734  0.0981738225 -0.0157263487 -0.0974657685  0.0389564930 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#> -0.0260778039  0.0577645306  0.0469568397 -0.0598365384  0.0400279144 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#> -0.0846911422 -0.0748472344  0.0885234803 -0.0257983168  0.0445630173 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#> -0.1760189490  0.0432621723 -0.0607847149 -0.0050028963 -0.0372588746 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  0.0429885936 -0.0494053872  0.0195251550  0.0107167093 -0.0517360709 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#>  0.0773432814 -0.0813424310  0.0469485839 -0.1401245515 -0.0816183285 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#> -0.0914787843  0.0181172525  0.0087229548 -0.0795764552  0.0509114077 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -0.0311414006  0.0457624460  0.0383817851 -0.0716041997  0.1074691116 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.0079118952  0.1038279682 -0.0597742981  0.0601420189 -0.1202684341 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.0022689237  0.0427398003  0.1157726412  0.0495763630  0.0704905299 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0599511343 -0.0215144140 -0.0967047301  0.0428041483 -0.1593290393 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0575482036 -0.0008813439 -0.0145508055 -0.0625647729 -0.0008895905 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.1130611100  0.0179506537  0.1033932818 -0.0003142048 -0.0011623415 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.1144319676  0.0600280948  0.1037886632 -0.0483290964  0.0159228076 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.1065062382  0.0548189469 -0.0312745161 -0.0109370208  0.1756116383 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.1308576001  0.1033530797 -0.1117629169  0.0399254776  0.0146143935 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -0.0208873232 -0.0228864886  0.0309967637  0.0244454220  0.0113191313 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.0364648462  0.0196710464  0.0322969947  0.0415290601  0.0095138783 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0193278492 -0.1352025543 -0.0248682231  0.0402440168 -0.0710656079 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0849652537  0.0688692102 -0.0311333170  0.0545667309  0.0440974553 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.0784526733  0.0571014450 -0.0248481885 -0.0395857091 -0.0700571914 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#> -0.0409780312  0.0170769924  0.0231924574  0.1110324165  0.0108849169 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0121961000 -0.0161112499 -0.0045538080 -0.0019375808 -0.0074174124 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0315054636 -0.0151790617  0.0702342038  0.1094668100  0.0176062603 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0900921133 -0.0351055236  0.1063301234  0.0042317596 -0.0200992325 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0361279102  0.0015969216 -0.0237082870 -0.0808109719  0.0553408347 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0729339659  0.0088886156 -0.0765452913 -0.0039255864 -0.0090074428 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0409439433  0.0401700840  0.0703560553 -0.1385976005  0.0203882187 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.0562146747 -0.0479565280 -0.1527463864 -0.0366665324  0.0147048732 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#>  0.0281403304 -0.0145130079 -0.1342206786 -0.0308050616 -0.0317955628 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#>  0.0138126340 -0.0271166708  0.0704536590 -0.0701556426 -0.0048589933 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.1165545732 -0.0397420027  0.0555246687 -0.1284437423  0.0135902983 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.1502928615  0.0202661685 -0.1679457908  0.0546148667 -0.1006462267 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0698083362  0.0512973324 -0.0619369407 -0.1325891161 -0.0479158283 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0570009100  0.0275070920 -0.0055445094 -0.0288438353  0.0353197799 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.0461649020  0.0272753601 -0.1029305014  0.0175691275  0.0812830948 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#>  0.0183979246  0.0250120195 -0.0713530119 -0.0221652683  0.1445877698 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.0958413898  0.0094162143  0.1158415418 -0.0336222632 -0.0270757227 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#>  0.0466155446 -0.0351133224 -0.0786254091  0.0108030284 -0.0666191292 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  0.0119939121  0.1114393453 -0.0471500794 -0.0663739825 -0.0550945032 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0545494142 -0.1137623645 -0.0236951056 -0.1099287950 -0.0012731174 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  0.1519109355  0.0044313247 -0.0108881602  0.0158393047 -0.0029127670 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0560929616 -0.1117502311 -0.0339744342 -0.0227105559 -0.0716624413 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0727784562 -0.0741860143  0.0081385349 -0.0913442745  0.0100445034 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -0.0222005808  0.0922160045  0.0313588220 -0.0114666228 -0.0719038618 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0087816150 -0.0188148809 -0.0556692308  0.0614249252 -0.0245982537 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -0.0635151473  0.0813265586  0.0624502597  0.0527336849 -0.0346038839 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.0404375255 -0.0088326729 -0.0530576604  0.0532982466 -0.0858220235 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  0.0184565441 -0.0874802628  0.0228477380 -0.0986035238  0.0207966182 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.1352236240 -0.0435537915  0.1136865653  0.1043457690 -0.0013675822 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.0187595359  0.0062149105  0.0274370267 -0.0147228407 -0.0231967041 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#> -0.0098735897 -0.0872288399  0.0001544451  0.0111357228 -0.0401313018 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0327924068  0.0118027802  0.0547991141  0.0407869418  0.0174959133 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -0.0044065277  0.0011351710 -0.0392244464