Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0945106887  0.0559064066 -0.0265694487  0.0527723909  0.0113187755 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#>  0.0920362550 -0.0465970794 -0.0692291083 -0.0929664518 -0.0730025679 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0607083914  0.0465185325 -0.0581957080  0.0851752316  0.0452943698 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#>  0.0825265769  0.0527311376 -0.0700073897 -0.0479772275  0.0572327686 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0527468569 -0.0457348034  0.0974026580 -0.0202354401 -0.1318655226 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.0556508855 -0.0599443989  0.0891387675  0.0341247198  0.0349195149 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#>  0.0694999434  0.0168082698  0.0328831194  0.0361686302 -0.0711338418 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -0.0321294646 -0.0318677608 -0.0026108592 -0.0387993359 -0.0252699158 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.0417623202  0.0293512967 -0.0288436507  0.0831354526  0.0387686148 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  0.0288937616 -0.0985110997  0.0380390646  0.1018567012  0.0126167377 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  0.0687850475  0.0306399523 -0.0056618374 -0.0450763658  0.1500884002 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.0092750161  0.0182522290  0.1289247748  0.0106070748 -0.0543094909 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#>  0.0793229171  0.1246940854 -0.0729281458 -0.0093183531 -0.0172905934 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0548884831 -0.0208609553 -0.0634842904  0.0013920241  0.0149943652 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -0.0951851428 -0.1043712455  0.0082497331 -0.0244408867  0.0570178700 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  0.0155219294  0.0512180880  0.1469832800 -0.0054064390 -0.0352872111 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0222225007  0.0019986450  0.0520497275  0.0461427245 -0.0298406854 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  0.0131203634  0.0510943498 -0.0429495271  0.0717888005 -0.1564402000 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0529935873  0.1661407796 -0.0003915199 -0.0206703887  0.0276683122 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  0.1083236236 -0.0238841423 -0.0001173622 -0.0692115310 -0.0142386418 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.1671653563 -0.0761018291  0.0187456604  0.0616581512 -0.1153957400 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#>  0.1145829206  0.0134780122 -0.0625600919 -0.0423271624  0.0702352457 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#> -0.0797154406 -0.0018482834  0.0965345480 -0.0613319830  0.0374048112 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.0149575875  0.0344594724  0.0961462490 -0.0365694035 -0.1029730480 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#>  0.0065035306  0.1385788446  0.0182523865 -0.1820552291  0.0603393451 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -0.2106000693  0.0760800884  0.0693183582 -0.0693674486 -0.0081500905 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#> -0.0730815300  0.0403708970  0.0290358106 -0.0812013521 -0.0303136142 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0055554420 -0.0537359303  0.1042351932  0.0302903010 -0.0690756959 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0194280957 -0.0243168562 -0.0619075898 -0.1050559831  0.0059784495 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#>  0.0952441747  0.0539819621 -0.0250390517 -0.1093987926 -0.1543555640 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0353877122  0.0411538274 -0.1041262870  0.0803769912 -0.0068757443 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.0029728344 -0.0525348870  0.0321418298 -0.1467054895  0.0053807472 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#> -0.0029230129 -0.0262339005  0.0385902748 -0.0483464597 -0.0441532550 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0159085620 -0.1100899423  0.0452623696 -0.0405322289  0.0403419849 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0349425293  0.0831486414  0.0351043931  0.0009343503 -0.0741418812 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.1574254524 -0.0126142051 -0.1122557831 -0.0156240523  0.0005429479 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -0.0523773238 -0.0909682728 -0.1077367012  0.0453572609 -0.0002001560 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  0.0079355045  0.1102946411  0.1136695557  0.0310813218  0.1162287285 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.0040473588 -0.0765969493 -0.0266802150  0.0376610679 -0.0807369995 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  0.0786691563 -0.0856368553 -0.0969723982  0.0098741035  0.0683308161 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#> -0.1001640605  0.0088612715 -0.0588672146 -0.0857273187 -0.0085460840 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.0526168481 -0.0029623364  0.0408009760 -0.1405721758  0.0496201599 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.0026265185  0.0141349885 -0.0248917602 -0.0443325806  0.0636146417 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#> -0.0037925361 -0.0839239161  0.1737200488  0.0708629651 -0.0104144950 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#> -0.0257180159  0.0149210861  0.0251168965  0.0382197567 -0.0613440914 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#> -0.0473663331  0.0337503638 -0.1069768360  0.0489754503 -0.0086108630 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0370440037  0.0719610234  0.0425636135  0.0350639928  0.0009045470 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#>  0.0501649371  0.0441667222 -0.0970313832 -0.0668541277  0.1162876292 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.0422091270  0.0283104533 -0.0938491919  0.1339879438  0.1107122783 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#> -0.1077737802 -0.0579936609 -0.0293990611 -0.0161365998  0.0600578140 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#> -0.0706816135 -0.0838927467 -0.0385745417 -0.1190925461 -0.0084308590 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#>  0.0646919644 -0.0857771425 -0.0895395470 -0.0151688295  0.0260917213 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#>  0.0404984239 -0.0339617398 -0.0746770892 -0.0349895362  0.0425237242 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.0028116250 -0.0214699779  0.0912486051  0.0374357789  0.0032636066 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#>  0.0011326521  0.0744438593  0.0387765007  0.0104720423 -0.0493230821 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#> -0.0961445244  0.0868761358  0.0169960171  0.0118754386  0.0132651369 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.2044344588 -0.0659244998 -0.1215733317  0.0041609408  0.1120093334 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0748089552  0.0550133714  0.0718918226  0.0295921652  0.0916902289 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#> -0.0221978784 -0.0584614589 -0.0855116639  0.0609634781  0.0617061889 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0057781190 -0.1449820627  0.0607862863  0.0060106170 -0.0077861843 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#>  0.0334469525  0.0963807781  0.0498028805  0.1517744987 -0.0717471442 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#> -0.0986614979 -0.0037596301  0.1085805285  0.0600260477  0.0177154209 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#> -0.1201741688 -0.0329468373 -0.0420383645  0.0112353535  0.0765134334 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0221778485 -0.0585826385  0.0630896356 -0.0179286346  0.1421814613 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.0580700682 -0.0480143350 -0.0200796698  0.1688278534  0.0602616621 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -0.0432712920 -0.0816981633  0.0787680549  0.0948605180  0.0086969544 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0735494917  0.0457718945  0.0638538661 -0.0224615738 -0.0236468182 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#>  0.0403263045  0.0425109339 -0.0184186792  0.1866060596  0.0563013387 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.0484218237 -0.0816632332 -0.1032960824 -0.0607605451  0.0080789415 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0744027029 -0.1059053737 -0.0223377393  0.0254049450 -0.1089703902 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0325234200  0.0242273430  0.1523607066  0.0639999409 -0.0688985235 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#>  0.0620594186  0.0359426643  0.1256079643 -0.1312565993  0.0230174279 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.1067162199  0.0776706000  0.0934586986 -0.0465855850  0.1979885804 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#>  0.1568037588 -0.1359022978 -0.0523785885  0.0823829977  0.0388834723 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -0.0416814594  0.0295586882  0.1055318930 -0.0375263456  0.0144139563 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#>  0.0745784360  0.0536881297  0.0233160399  0.0658885475 -0.0393056754 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#>  0.0545546365 -0.0893751734 -0.0444534041 -0.0156179050  0.0847951439 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#> -0.0273550294 -0.0410961080  0.0109312764  0.0770987580 -0.0600570165 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#> -0.0464820373  0.0551635070 -0.1017193232 -0.0830424007  0.0593320192 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#>  0.0259457719  0.0237226322  0.0289281692  0.0207525916  0.0143455845 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#>  0.0950994432 -0.0465794411  0.0932557212  0.1464427625 -0.0336983800 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#>  0.0489563226 -0.0134182557 -0.0393986871 -0.0469869919 -0.1251479944 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#>  0.0013852436  0.0205494935  0.0735927067  0.0388485939  0.0327085866 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#> -0.1064288755  0.0554253902  0.0107159552 -0.0125334852  0.0227522482 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  0.0549712356 -0.0229260053 -0.0389068489 -0.0350905627 -0.0375277543 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#>  0.0371570683 -0.1147475629 -0.0512330854 -0.0621255011 -0.0981916240 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.0226544422 -0.0701988030 -0.0167089590  0.0675865335 -0.0810783210 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.1287657446  0.0920488170 -0.0750810192 -0.0146912686  0.0203867307 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.1209209975  0.1085042440 -0.0167184123  0.0083134869 -0.0056444040 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -0.0148933753 -0.0258397078  0.1177237781