Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Auto Cross Covariance (ACC) for Generating Scales-Based Descriptors of the Same Length

Usage

acc(mat, lag)

Arguments

mat

A p * n matrix. Each row represents one scale (total p scales), each column represents one amino acid position (total n amino acids).

lag

The lag parameter. Must be less than the amino acids.

Value

A length lag * p^2 named vector, the element names are constructed by: the scales index (crossed scales index) and lag index.

Details

This function calculates the auto covariance and auto cross covariance for generating scale-based descriptors of the same length.

Note

To know more details about auto cross covariance, see the references.

References

Wold, S., Jonsson, J., Sj\"orstr\"om, M., Sandberg, M., & R\"annar, S. (1993). DNA and peptide sequences and chemical processes multivariately modelled by principal component analysis and partial least-squares projections to latent structures. Analytica chimica acta, 277(2), 239--253.

Sj\"ostr\"om, M., R\"annar, S., & Wieslander, A. (1995). Polypeptide sequence property relationships in Escherichia coli based on auto cross covariances. Chemometrics and intelligent laboratory systems, 29(2), 295--305.

See also

Examples

p = 8    # p is the scales number
n = 200  # n is the amino acid number
lag = 7  # the lag paramter
mat = matrix(rnorm(p * n), nrow = p, ncol = n)
acc(mat, lag)
#>     scl1.lag1     scl2.lag1     scl3.lag1     scl4.lag1     scl5.lag1 
#> -0.0732195510  0.0527299398 -0.1164922179  0.0133660991  0.0600066715 
#>     scl6.lag1     scl7.lag1     scl8.lag1     scl1.lag2     scl2.lag2 
#> -0.0315355271  0.0278193750  0.0517721536  0.0057333785  0.0457993117 
#>     scl3.lag2     scl4.lag2     scl5.lag2     scl6.lag2     scl7.lag2 
#> -0.0265456086  0.0528677901 -0.0863351897  0.0631623298  0.0704848435 
#>     scl8.lag2     scl1.lag3     scl2.lag3     scl3.lag3     scl4.lag3 
#> -0.0521407006  0.0090269225  0.0108667079  0.0502571995  0.0368025104 
#>     scl5.lag3     scl6.lag3     scl7.lag3     scl8.lag3     scl1.lag4 
#>  0.0126958955 -0.0320655803  0.0395823929  0.0344094797 -0.0496266256 
#>     scl2.lag4     scl3.lag4     scl4.lag4     scl5.lag4     scl6.lag4 
#>  0.0060881032  0.0427453447 -0.0612420115 -0.0267843273 -0.0438371053 
#>     scl7.lag4     scl8.lag4     scl1.lag5     scl2.lag5     scl3.lag5 
#> -0.0006844546  0.0562306852 -0.0322810458 -0.0955510994 -0.0627789053 
#>     scl4.lag5     scl5.lag5     scl6.lag5     scl7.lag5     scl8.lag5 
#> -0.0403254377 -0.0004117857  0.0361983627 -0.1120716630  0.0366292302 
#>     scl1.lag6     scl2.lag6     scl3.lag6     scl4.lag6     scl5.lag6 
#>  0.0461660345 -0.0321426955  0.0695206610 -0.0587821331  0.1650025816 
#>     scl6.lag6     scl7.lag6     scl8.lag6     scl1.lag7     scl2.lag7 
#>  0.1467326559 -0.0285672321  0.0781871007 -0.0101113385 -0.0868750924 
#>     scl3.lag7     scl4.lag7     scl5.lag7     scl6.lag7     scl7.lag7 
#>  0.0757292232 -0.0300068888  0.0426519840 -0.0187759768 -0.1408330426 
#>     scl8.lag7   scl1.2.lag1   scl1.3.lag1   scl1.4.lag1   scl1.5.lag1 
#>  0.0038781132  0.0433744226 -0.0630274782  0.1408936475 -0.0739628180 
#>   scl1.6.lag1   scl1.7.lag1   scl1.8.lag1   scl2.3.lag1   scl2.4.lag1 
#> -0.0608924285 -0.0346549248 -0.0550658042 -0.0307471011  0.0606732106 
#>   scl2.5.lag1   scl2.6.lag1   scl2.7.lag1   scl2.8.lag1   scl3.4.lag1 
#> -0.0523547489 -0.0718311683 -0.2346993531 -0.1192495670  0.0459680186 
#>   scl3.5.lag1   scl3.6.lag1   scl3.7.lag1   scl3.8.lag1   scl4.5.lag1 
#> -0.0408025124 -0.0800479532  0.0485921408 -0.0213420868  0.0401452728 
#>   scl4.6.lag1   scl4.7.lag1   scl4.8.lag1   scl5.6.lag1   scl5.7.lag1 
#>  0.0013761439  0.0140077190 -0.0337237300 -0.0305125342  0.0406959214 
#>   scl5.8.lag1   scl6.7.lag1   scl6.8.lag1   scl7.8.lag1   scl2.1.lag1 
#>  0.0301944927 -0.0007983560  0.0183503054  0.0642241405  0.0431828080 
#>   scl3.1.lag1   scl4.1.lag1   scl5.1.lag1   scl6.1.lag1   scl7.1.lag1 
#>  0.0238716831  0.0513892601  0.0223850640 -0.0228830073  0.0826642451 
#>   scl8.1.lag1   scl3.2.lag1   scl4.2.lag1   scl5.2.lag1   scl6.2.lag1 
#>  0.0336608464  0.0825238614  0.0192707684 -0.0452662567 -0.0368761091 
#>   scl7.2.lag1   scl8.2.lag1   scl4.3.lag1   scl5.3.lag1   scl6.3.lag1 
#>  0.0439876288  0.0571223137 -0.0328628985 -0.0335869823 -0.0240267375 
#>   scl7.3.lag1   scl8.3.lag1   scl5.4.lag1   scl6.4.lag1   scl7.4.lag1 
#>  0.0672753937  0.0131802100 -0.0320254489  0.1289761327 -0.1339084677 
#>   scl8.4.lag1   scl6.5.lag1   scl7.5.lag1   scl8.5.lag1   scl7.6.lag1 
#> -0.0545559477 -0.0254234490  0.0402344671 -0.0071543905 -0.0314935414 
#>   scl8.6.lag1   scl8.7.lag1   scl1.2.lag2   scl1.3.lag2   scl1.4.lag2 
#>  0.1081262018  0.0149540628  0.0274448414 -0.0407498115  0.0373220191 
#>   scl1.5.lag2   scl1.6.lag2   scl1.7.lag2   scl1.8.lag2   scl2.3.lag2 
#> -0.0144936791  0.0667636615  0.0298686553  0.0652860702 -0.0451448866 
#>   scl2.4.lag2   scl2.5.lag2   scl2.6.lag2   scl2.7.lag2   scl2.8.lag2 
#> -0.0342953231  0.0439220251  0.0368153539 -0.0698557439 -0.0626124244 
#>   scl3.4.lag2   scl3.5.lag2   scl3.6.lag2   scl3.7.lag2   scl3.8.lag2 
#> -0.0425427020  0.0351520631  0.0461461917 -0.1047629154 -0.0010984751 
#>   scl4.5.lag2   scl4.6.lag2   scl4.7.lag2   scl4.8.lag2   scl5.6.lag2 
#>  0.0156945494  0.0175197124  0.0362106311 -0.0256023797  0.0250347558 
#>   scl5.7.lag2   scl5.8.lag2   scl6.7.lag2   scl6.8.lag2   scl7.8.lag2 
#> -0.0077688092  0.0490620284 -0.0671954613 -0.0596042969 -0.0454984820 
#>   scl2.1.lag2   scl3.1.lag2   scl4.1.lag2   scl5.1.lag2   scl6.1.lag2 
#>  0.0502970105 -0.0294926499  0.0126245298  0.0310928838 -0.0495683183 
#>   scl7.1.lag2   scl8.1.lag2   scl3.2.lag2   scl4.2.lag2   scl5.2.lag2 
#> -0.0188310064  0.0133450202  0.0389960336  0.0706435855  0.0679429192 
#>   scl6.2.lag2   scl7.2.lag2   scl8.2.lag2   scl4.3.lag2   scl5.3.lag2 
#> -0.0208925602  0.0816783370 -0.1217505840  0.0458471716  0.0462236946 
#>   scl6.3.lag2   scl7.3.lag2   scl8.3.lag2   scl5.4.lag2   scl6.4.lag2 
#>  0.1070915285 -0.0450622892  0.0789082033  0.0518412095 -0.0356718378 
#>   scl7.4.lag2   scl8.4.lag2   scl6.5.lag2   scl7.5.lag2   scl8.5.lag2 
#>  0.0620734069 -0.0249390550 -0.0511644424  0.0520960158 -0.0947779537 
#>   scl7.6.lag2   scl8.6.lag2   scl8.7.lag2   scl1.2.lag3   scl1.3.lag3 
#>  0.0199185124  0.0096922207  0.0090086855  0.0999290682  0.0242199543 
#>   scl1.4.lag3   scl1.5.lag3   scl1.6.lag3   scl1.7.lag3   scl1.8.lag3 
#>  0.0192257562  0.0272114926 -0.0087251909 -0.0754717969 -0.0543971795 
#>   scl2.3.lag3   scl2.4.lag3   scl2.5.lag3   scl2.6.lag3   scl2.7.lag3 
#> -0.0794876180  0.1286141378  0.0435812499  0.0134323843 -0.0584706834 
#>   scl2.8.lag3   scl3.4.lag3   scl3.5.lag3   scl3.6.lag3   scl3.7.lag3 
#> -0.0018088149  0.0605587474 -0.0199672012  0.1347651517  0.0437302640 
#>   scl3.8.lag3   scl4.5.lag3   scl4.6.lag3   scl4.7.lag3   scl4.8.lag3 
#>  0.0913952348 -0.1426735336  0.0161195728 -0.1585075822  0.0639843308 
#>   scl5.6.lag3   scl5.7.lag3   scl5.8.lag3   scl6.7.lag3   scl6.8.lag3 
#>  0.0035789584 -0.0159184084 -0.0039980354 -0.1424939497 -0.0472286563 
#>   scl7.8.lag3   scl2.1.lag3   scl3.1.lag3   scl4.1.lag3   scl5.1.lag3 
#>  0.0636194614 -0.0083997994  0.0446563666 -0.0292819489 -0.1114591843 
#>   scl6.1.lag3   scl7.1.lag3   scl8.1.lag3   scl3.2.lag3   scl4.2.lag3 
#>  0.0102716507  0.0548556586  0.0682910725  0.0545287909  0.0763755437 
#>   scl5.2.lag3   scl6.2.lag3   scl7.2.lag3   scl8.2.lag3   scl4.3.lag3 
#> -0.1110628526 -0.0822295375 -0.0528595053 -0.0618060640 -0.0007276911 
#>   scl5.3.lag3   scl6.3.lag3   scl7.3.lag3   scl8.3.lag3   scl5.4.lag3 
#>  0.0122379084  0.0096473053 -0.0051414240 -0.0383162805 -0.1298833161 
#>   scl6.4.lag3   scl7.4.lag3   scl8.4.lag3   scl6.5.lag3   scl7.5.lag3 
#>  0.0237819644  0.1056310488 -0.0367980310  0.0735045357  0.0414726907 
#>   scl8.5.lag3   scl7.6.lag3   scl8.6.lag3   scl8.7.lag3   scl1.2.lag4 
#>  0.0412717734 -0.0426715701 -0.0936057801  0.0564711764 -0.0358621585 
#>   scl1.3.lag4   scl1.4.lag4   scl1.5.lag4   scl1.6.lag4   scl1.7.lag4 
#>  0.0522768490  0.0273400043 -0.0282724915  0.0154538301 -0.1594616780 
#>   scl1.8.lag4   scl2.3.lag4   scl2.4.lag4   scl2.5.lag4   scl2.6.lag4 
#> -0.0690742524 -0.0242617844 -0.0832782886 -0.0355809141 -0.0796236799 
#>   scl2.7.lag4   scl2.8.lag4   scl3.4.lag4   scl3.5.lag4   scl3.6.lag4 
#> -0.0300876767 -0.0950500278 -0.0357806695 -0.0033208310  0.0559554750 
#>   scl3.7.lag4   scl3.8.lag4   scl4.5.lag4   scl4.6.lag4   scl4.7.lag4 
#> -0.0159838317  0.1202183070 -0.1283678152 -0.0349215151  0.0160951229 
#>   scl4.8.lag4   scl5.6.lag4   scl5.7.lag4   scl5.8.lag4   scl6.7.lag4 
#>  0.0047860856 -0.1244803321  0.0077568591  0.0457804180 -0.0201686038 
#>   scl6.8.lag4   scl7.8.lag4   scl2.1.lag4   scl3.1.lag4   scl4.1.lag4 
#>  0.0013784178  0.0481268557  0.0028026213  0.0418531805 -0.0115703862 
#>   scl5.1.lag4   scl6.1.lag4   scl7.1.lag4   scl8.1.lag4   scl3.2.lag4 
#> -0.0309312131 -0.0100035052  0.0889496418  0.0583389588 -0.0543155763 
#>   scl4.2.lag4   scl5.2.lag4   scl6.2.lag4   scl7.2.lag4   scl8.2.lag4 
#> -0.0241324390 -0.1626284649 -0.0012786919 -0.0397199585 -0.0960337236 
#>   scl4.3.lag4   scl5.3.lag4   scl6.3.lag4   scl7.3.lag4   scl8.3.lag4 
#> -0.0681243999  0.0595589435  0.1550382051 -0.0649847553  0.1229770907 
#>   scl5.4.lag4   scl6.4.lag4   scl7.4.lag4   scl8.4.lag4   scl6.5.lag4 
#> -0.1125261606 -0.0669452395 -0.0171126758 -0.0152928309  0.0410280176 
#>   scl7.5.lag4   scl8.5.lag4   scl7.6.lag4   scl8.6.lag4   scl8.7.lag4 
#>  0.0094019720  0.1057110339  0.0492116088 -0.0401684107 -0.0345371074 
#>   scl1.2.lag5   scl1.3.lag5   scl1.4.lag5   scl1.5.lag5   scl1.6.lag5 
#>  0.0135419416 -0.1490910538 -0.0091966917  0.0367726221 -0.0468559247 
#>   scl1.7.lag5   scl1.8.lag5   scl2.3.lag5   scl2.4.lag5   scl2.5.lag5 
#>  0.0129659690 -0.0903226439 -0.0428196770 -0.0457530765  0.0259345930 
#>   scl2.6.lag5   scl2.7.lag5   scl2.8.lag5   scl3.4.lag5   scl3.5.lag5 
#>  0.0179162148 -0.0833842622  0.0098135798  0.0221386279  0.0152091338 
#>   scl3.6.lag5   scl3.7.lag5   scl3.8.lag5   scl4.5.lag5   scl4.6.lag5 
#>  0.0311692110  0.0327315647 -0.1634729602 -0.0338521422 -0.0071879330 
#>   scl4.7.lag5   scl4.8.lag5   scl5.6.lag5   scl5.7.lag5   scl5.8.lag5 
#> -0.0453824579  0.0109304856  0.0765432331  0.0963056271 -0.0763948470 
#>   scl6.7.lag5   scl6.8.lag5   scl7.8.lag5   scl2.1.lag5   scl3.1.lag5 
#>  0.0538069290  0.0386523746  0.0191309896 -0.0764423398  0.0060597291 
#>   scl4.1.lag5   scl5.1.lag5   scl6.1.lag5   scl7.1.lag5   scl8.1.lag5 
#>  0.0500538627 -0.0513994450 -0.0554140542 -0.0063877759  0.1296587386 
#>   scl3.2.lag5   scl4.2.lag5   scl5.2.lag5   scl6.2.lag5   scl7.2.lag5 
#>  0.0835614124 -0.0494823421 -0.0168738513  0.0389197615  0.0203282966 
#>   scl8.2.lag5   scl4.3.lag5   scl5.3.lag5   scl6.3.lag5   scl7.3.lag5 
#> -0.0021553478  0.0925001576 -0.0872921103 -0.0787784135  0.0541581162 
#>   scl8.3.lag5   scl5.4.lag5   scl6.4.lag5   scl7.4.lag5   scl8.4.lag5 
#> -0.0343845016  0.1061154107 -0.1188428015  0.1356499566  0.1423319218 
#>   scl6.5.lag5   scl7.5.lag5   scl8.5.lag5   scl7.6.lag5   scl8.6.lag5 
#> -0.0186698800  0.0321996389  0.0223629546  0.0530214637  0.1322838621 
#>   scl8.7.lag5   scl1.2.lag6   scl1.3.lag6   scl1.4.lag6   scl1.5.lag6 
#> -0.0089238326 -0.1143161681  0.0607317200  0.0036069942 -0.0320248797 
#>   scl1.6.lag6   scl1.7.lag6   scl1.8.lag6   scl2.3.lag6   scl2.4.lag6 
#>  0.0298454428 -0.1353880277  0.1322991973  0.0244713488 -0.0394746707 
#>   scl2.5.lag6   scl2.6.lag6   scl2.7.lag6   scl2.8.lag6   scl3.4.lag6 
#> -0.0495686312 -0.1889059447  0.1246939353 -0.0145643678  0.0437307580 
#>   scl3.5.lag6   scl3.6.lag6   scl3.7.lag6   scl3.8.lag6   scl4.5.lag6 
#> -0.0144085520 -0.1230342102 -0.1101016424 -0.0362513467 -0.0100620226 
#>   scl4.6.lag6   scl4.7.lag6   scl4.8.lag6   scl5.6.lag6   scl5.7.lag6 
#> -0.1409208788  0.0273936884  0.1235617596 -0.1089378871  0.0637251459 
#>   scl5.8.lag6   scl6.7.lag6   scl6.8.lag6   scl7.8.lag6   scl2.1.lag6 
#> -0.0780091362  0.0042626067  0.0784960609 -0.0048009131  0.0085502205 
#>   scl3.1.lag6   scl4.1.lag6   scl5.1.lag6   scl6.1.lag6   scl7.1.lag6 
#> -0.0744665487  0.0354271631 -0.0811127283  0.1065638679  0.0118782139 
#>   scl8.1.lag6   scl3.2.lag6   scl4.2.lag6   scl5.2.lag6   scl6.2.lag6 
#> -0.0039592227 -0.0249793145  0.0011183272 -0.0103765119  0.0720128549 
#>   scl7.2.lag6   scl8.2.lag6   scl4.3.lag6   scl5.3.lag6   scl6.3.lag6 
#> -0.0352265513  0.0215973931 -0.0787459805  0.0682978159 -0.0663933613 
#>   scl7.3.lag6   scl8.3.lag6   scl5.4.lag6   scl6.4.lag6   scl7.4.lag6 
#> -0.1538032010 -0.0233562557 -0.0663190719 -0.0911165450  0.0845964729 
#>   scl8.4.lag6   scl6.5.lag6   scl7.5.lag6   scl8.5.lag6   scl7.6.lag6 
#>  0.0684756025  0.0254971176  0.0416148362  0.0020684161  0.0580050130 
#>   scl8.6.lag6   scl8.7.lag6   scl1.2.lag7   scl1.3.lag7   scl1.4.lag7 
#>  0.0596890349 -0.0333264178  0.0507508298 -0.0143573982 -0.0940774825 
#>   scl1.5.lag7   scl1.6.lag7   scl1.7.lag7   scl1.8.lag7   scl2.3.lag7 
#> -0.0040673550 -0.0563504239 -0.0016141392 -0.0382851135  0.0486611661 
#>   scl2.4.lag7   scl2.5.lag7   scl2.6.lag7   scl2.7.lag7   scl2.8.lag7 
#> -0.0650103699 -0.0909561723 -0.1166419909  0.0878364301 -0.0596731309 
#>   scl3.4.lag7   scl3.5.lag7   scl3.6.lag7   scl3.7.lag7   scl3.8.lag7 
#> -0.0509607017  0.0552471657  0.0484345056 -0.0084992813  0.0694502475 
#>   scl4.5.lag7   scl4.6.lag7   scl4.7.lag7   scl4.8.lag7   scl5.6.lag7 
#> -0.0007129293  0.1315400786 -0.0609747924 -0.0973253987 -0.0716421988 
#>   scl5.7.lag7   scl5.8.lag7   scl6.7.lag7   scl6.8.lag7   scl7.8.lag7 
#>  0.0615805339  0.0965000990 -0.0043279423 -0.1042296052  0.0105913310 
#>   scl2.1.lag7   scl3.1.lag7   scl4.1.lag7   scl5.1.lag7   scl6.1.lag7 
#>  0.0187921575  0.0241961502 -0.0267755222  0.0369779964 -0.0122076959 
#>   scl7.1.lag7   scl8.1.lag7   scl3.2.lag7   scl4.2.lag7   scl5.2.lag7 
#> -0.0121380274 -0.0639347266  0.1503548175  0.0605255188 -0.0664488148 
#>   scl6.2.lag7   scl7.2.lag7   scl8.2.lag7   scl4.3.lag7   scl5.3.lag7 
#>  0.0351448986 -0.0824497472  0.1485459374 -0.0208420734 -0.0288202970 
#>   scl6.3.lag7   scl7.3.lag7   scl8.3.lag7   scl5.4.lag7   scl6.4.lag7 
#>  0.0488667848  0.0457308172  0.0305120221  0.0057856090 -0.0360548825 
#>   scl7.4.lag7   scl8.4.lag7   scl6.5.lag7   scl7.5.lag7   scl8.5.lag7 
#> -0.0540723173  0.0072110031 -0.0056472274 -0.0111262733  0.0676789619 
#>   scl7.6.lag7   scl8.6.lag7   scl8.7.lag7 
#> -0.0828705420  0.0783137999 -0.0035984304