Chemical File Formats Conversion

Chemical File Formats Conversion

Usage

convMolFormat(infile, outfile, from, to)

Arguments

infile

A character string. Indicating the input file location.

outfile

A character string. Indicating the output file location.

from

The format of infile. A character string supported by OpenBabel. See the note section for the supported formats.

to

The desired format of outfile. A character string supported by OpenBabel. See the note section for the supported formats.

Value

NULL

Details

This function converts between various chemical file formats via OpenBabel. The complete supported file format list could be found at http://openbabel.org/docs/dev/FileFormats/Overview.html.

Note

The supported formats include:

abinit -- ABINIT Output Format [Read-only]

acr -- ACR format [Read-only]

adf -- ADF cartesian input format [Write-only]

adfout -- ADF output format [Read-only]

alc -- Alchemy format

arc -- Accelrys/MSI Biosym/Insight II CAR format [Read-only]

axsf -- XCrySDen Structure Format [Read-only]

bgf -- MSI BGF format

box -- Dock 3.5 Box format

bs -- Ball and Stick format

c3d1 -- Chem3D Cartesian 1 format

c3d2 -- Chem3D Cartesian 2 format

cac -- CAChe MolStruct format [Write-only]

caccrt -- Cacao Cartesian format

cache -- CAChe MolStruct format [Write-only]

cacint -- Cacao Internal format [Write-only]

can -- Canonical SMILES format

car -- Accelrys/MSI Biosym/Insight II CAR format [Read-only]

castep -- CASTEP format [Read-only]

ccc -- CCC format [Read-only]

cdx -- ChemDraw binary format [Read-only]

cdxml -- ChemDraw CDXML format

cht -- Chemtool format [Write-only]

cif -- Crystallographic Information File

ck -- ChemKin format

cml -- Chemical Markup Language

cmlr -- CML Reaction format

com -- Gaussian 98/03 Input [Write-only]

CONFIG -- DL-POLY CONFIG

CONTCAR -- VASP format [Read-only]

copy -- Copy raw text [Write-only]

crk2d -- Chemical Resource Kit diagram(2D)

crk3d -- Chemical Resource Kit 3D format

csr -- Accelrys/MSI Quanta CSR format [Write-only]

cssr -- CSD CSSR format [Write-only]

ct -- ChemDraw Connection Table format

cub -- Gaussian cube format

cube -- Gaussian cube format

dat -- Generic Output file format [Read-only]

dmol -- DMol3 coordinates format

dx -- OpenDX cube format for APBS

ent -- Protein Data Bank format

fa -- FASTA format

fasta -- FASTA format

fch -- Gaussian formatted checkpoint file format [Read-only]

fchk -- Gaussian formatted checkpoint file format [Read-only]

fck -- Gaussian formatted checkpoint file format [Read-only]

feat -- Feature format

fh -- Fenske-Hall Z-Matrix format [Write-only]

fhiaims -- FHIaims XYZ format

fix -- SMILES FIX format [Write-only]

fpt -- Fingerprint format [Write-only]

fract -- Free Form Fractional format

fs -- Fastsearch format

fsa -- FASTA format

g03 -- Gaussian Output [Read-only]

g09 -- Gaussian Output [Read-only]

g92 -- Gaussian Output [Read-only]

g94 -- Gaussian Output [Read-only]

g98 -- Gaussian Output [Read-only]

gal -- Gaussian Output [Read-only]

gam -- GAMESS Output [Read-only]

gamess -- GAMESS Output [Read-only]

gamin -- GAMESS Input

gamout -- GAMESS Output [Read-only]

gau -- Gaussian 98/03 Input [Write-only]

gjc -- Gaussian 98/03 Input [Write-only]

gjf -- Gaussian 98/03 Input [Write-only]

got -- GULP format [Read-only]

gpr -- Ghemical format

gr96 -- GROMOS96 format [Write-only]

gro -- GRO format

gukin -- GAMESS-UK Input

gukout -- GAMESS-UK Output

gzmat -- Gaussian Z-Matrix Input

hin -- HyperChem HIN format

HISTORY -- DL-POLY HISTORY [Read-only]

inchi -- InChI format

inchikey -- InChIKey [Write-only]

inp -- GAMESS Input

ins -- ShelX format [Read-only]

jin -- Jaguar input format [Write-only]

jout -- Jaguar output format [Read-only]

k -- Compare molecules using InChI [Write-only]

log -- Generic Output file format [Read-only]

mcdl -- MCDL format

mcif -- Macromolecular Crystallographic Info

mdl -- MDL MOL format

ml2 -- Sybyl Mol2 format

mmcif -- Macromolecular Crystallographic Info

mmd -- MacroModel format

mmod -- MacroModel format

mna -- Multilevel Neighborhoods of Atoms (MNA) [Write-only]

mol -- MDL MOL format

mol2 -- Sybyl Mol2 format

mold -- Molden format

molden -- Molden format

molf -- Molden format

molreport -- Open Babel molecule report [Write-only]

moo -- MOPAC Output format [Read-only]

mop -- MOPAC Cartesian format

mopcrt -- MOPAC Cartesian format

mopin -- MOPAC Internal

mopout -- MOPAC Output format [Read-only]

mp -- Molpro input format [Write-only]

mpc -- MOPAC Cartesian format

mpd -- MolPrint2D format [Write-only]

mpo -- Molpro output format [Read-only]

mpqc -- MPQC output format [Read-only]

mpqcin -- MPQC simplified input format [Write-only]

mrv -- Chemical Markup Language

msi -- Accelrys/MSI Cerius II MSI format [Read-only]

msms -- M.F. Sanner's MSMS input format [Write-only]

nul -- Outputs nothing [Write-only]

nw -- NWChem input format [Write-only]

nwo -- NWChem output format [Read-only]

out -- Generic Output file format [Read-only]

outmol -- DMol3 coordinates format

output -- Generic Output file format [Read-only]

pc -- PubChem format [Read-only]

pcm -- PCModel Format

pdb -- Protein Data Bank format

pdbqt -- AutoDock PDQBT format

png -- PNG 2D depiction

POSCAR -- VASP format [Read-only]

pov -- POV-Ray input format [Write-only]

pqr -- PQR format

pqs -- Parallel Quantum Solutions format

prep -- Amber Prep format [Read-only]

pwscf -- PWscf format [Read-only]

qcin -- Q-Chem input format [Write-only]

qcout -- Q-Chem output format [Read-only]

report -- Open Babel report format [Write-only]

res -- ShelX format [Read-only]

rsmi -- Reaction SMILES format

rxn -- MDL RXN format

sd -- MDL MOL format

sdf -- MDL MOL format

smi -- SMILES format

smiles -- SMILES format

svg -- SVG 2D depiction [Write-only]

sy2 -- Sybyl Mol2 format

t41 -- ADF TAPE41 format [Read-only]

tdd -- Thermo format

text -- Read and write raw text

therm -- Thermo format

tmol -- TurboMole Coordinate format

txt -- Title format

txyz -- Tinker XYZ format

unixyz -- UniChem XYZ format

vmol -- ViewMol format

xed -- XED format [Write-only]

xml -- General XML format [Read-only]

xsf -- XCrySDen Structure Format [Read-only]

xyz -- XYZ cartesian coordinates format

yob -- YASARA.org YOB format

zin -- ZINDO input format [Write-only]

Examples

sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')
# SDF to SMILES
if (FALSE) {
convMolFormat(infile = sdf, outfile = 'aa.smi',
              from = 'sdf', to = 'smiles')}
# SMILES to MOPAC Cartesian format
if (FALSE) {
convMolFormat(infile = 'aa.smi', outfile = 'aa.mop',
              from = 'smiles', to = 'mop')}