BCUT -- Eigenvalue Based Descriptor

BCUT -- Eigenvalue Based Descriptor

Usage

extractDrugBCUT(molecules, silent = TRUE)

Arguments

molecules

Parsed molucule object.

silent

Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 6 columns:

BCUTw-1l, BCUTw-2l ... - n high lowest atom weighted BCUTS

BCUTw-1h, BCUTw-2h ... - n low highest atom weighted BCUTS

BCUTc-1l, BCUTc-2l ... - n high lowest partial charge weighted BCUTS

BCUTc-1h, BCUTc-2h ... - n low highest partial charge weighted BCUTS

BCUTp-1l, BCUTp-2l ... - n high lowest polarizability weighted BCUTS

BCUTp-1h, BCUTp-2h ... - n low highest polarizability weighted BCUTS

Details

Eigenvalue based descriptor noted for its utility in chemical diversity. Described by Pearlman et al. The descriptor is based on a weighted version of the Burden matrix which takes into account both the connectivity as well as atomic properties of a molecule. The weights are a variety of atom properties placed along the diagonal of the Burden matrix. Currently three weighting schemes are employed:

Atomic Weight

Partial Charge (Gasteiger Marsilli)

Polarizability (Kang et al.)

Note

By default, the descriptor will return the highest and lowest eigenvalues for the three classes of descriptor in a single ArrayList (in the order shown above). However it is also possible to supply a parameter list indicating how many of the highest and lowest eigenvalues (for each class of descriptor) are required. The descriptor works with the hydrogen depleted molecule.

A side effect of specifying the number of highest and lowest eigenvalues is that it is possible to get two copies of all the eigenvalues. That is, if a molecule has 5 heavy atoms, then specifying the 5 highest eigenvalues returns all of them, and specifying the 5 lowest eigenvalues returns all of them, resulting in two copies of all the eigenvalues.

Note that it is possible to specify an arbitrarily large number of eigenvalues to be returned. However if the number (i.e., nhigh or nlow) is larger than the number of heavy atoms, the remaining eignevalues will be NaN.

Given the above description, if the aim is to gt all the eigenvalues for a molecule, you should set nlow to 0 and specify the number of heavy atoms (or some large number) for nhigh (or vice versa).

References

Pearlman, R.S. and Smith, K.M., Metric Validation and the Receptor-Relevant Subspace Concept, J. Chem. Inf. Comput. Sci., 1999, 39:28-35.

Burden, F.R., Molecular identification number for substructure searches, J. Chem. Inf. Comput. Sci., 1989, 29:225-227.

Burden, F.R., Chemically Intuitive Molecular Index, Quant. Struct. -Act. Relat., 1997, 16:309-314

Kang, Y.K. and Jhon, M.S., Additivity of Atomic Static Polarizabilities and Dispersion Coefficients, Theoretica Chimica Acta, 1982, 61:41-48

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugBCUT(mol)
#> Warning: Could not calculate partial charges: Cannot invoke "java.lang.Integer.intValue()" because the return value of "org.openscience.cdk.interfaces.IAtomType.getProperty(Object)" is null
head(dat)# }
#>                                                         BCUTw.1l BCUTw.1h
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 11.99616 16.00080
#> c1(ccc(cc1)O)NC(=O)C                                    11.89000 15.99592
#> c1ccccc1NC(=O)C                                         11.89000 15.99492
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              11.89000 31.97307
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                11.90000 31.97207
#> CC(=O)NO                                                11.99900 16.00115
#>                                                           BCUTc.1l  BCUTc.1h
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O -0.3848947 0.2972769
#> c1(ccc(cc1)O)NC(=O)C                                    -0.3607188 0.2511029
#> c1ccccc1NC(=O)C                                         -0.2810520 0.2510941
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              -0.2805235 0.2643700
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                -0.2941177 0.2858059
#> CC(=O)NO                                                -0.3632885 0.2197801
#>                                                         BCUTp.1l  BCUTp.1h
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 3.726860 11.858618
#> c1(ccc(cc1)O)NC(=O)C                                    4.199143  8.415819
#> c1ccccc1NC(=O)C                                         4.191793  8.357006
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              4.464387 12.193660
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                4.592529 11.513438
#> CC(=O)NO                                                3.520334  5.198756