A Variety of Descriptors Combining Surface Area and Partial Charge Information

A Variety of Descriptors Combining Surface Area and Partial Charge Information

extractDrugCPSA(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 29 columns:

PPSA.1 - partial positive surface area -- sum of surface area on positive parts of molecule

PPSA.2 - partial positive surface area * total positive charge on the molecule

PPSA.3 - charge weighted partial positive surface area

PNSA.1 - partial negative surface area -- sum of surface area on negative parts of molecule

PNSA.2 - partial negative surface area * total negative charge on the molecule

PNSA.3 - charge weighted partial negative surface area

DPSA.1 - difference of PPSA.1 and PNSA.1

DPSA.2 - difference of FPSA.2 and PNSA.2

DPSA.3 - difference of PPSA.3 and PNSA.3

FPSA.1 - PPSA.1 / total molecular surface area

FFSA.2 - PPSA.2 / total molecular surface area

FPSA.3 - PPSA.3 / total molecular surface area

FNSA.1 - PNSA.1 / total molecular surface area

FNSA.2 - PNSA.2 / total molecular surface area

FNSA.3 - PNSA.3 / total molecular surface area

WPSA.1 - PPSA.1 * total molecular surface area / 1000

WPSA.2 - PPSA.2 * total molecular surface area /1000

WPSA.3 - PPSA.3 * total molecular surface area / 1000

WNSA.1 - PNSA.1 * total molecular surface area /1000

WNSA.2 - PNSA.2 * total molecular surface area / 1000

WNSA.3 - PNSA.3 * total molecular surface area / 1000

RPCG - relative positive charge -- most positive charge / total positive charge

RNCG - relative negative charge -- most negative charge / total negative charge

RPCS - relative positive charge surface area -- most positive surface area * RPCG

RNCS - relative negative charge surface area -- most negative surface area * RNCG

THSA - sum of solvent accessible surface areas of atoms with absolute value of partial charges less than 0.2

TPSA - sum of solvent accessible surface areas of atoms with absolute value of partial charges greater than or equal 0.2

RHSA - THSA / total molecular surface area

RPSA - TPSA / total molecular surface area

Details

Calculates 29 Charged Partial Surface Area (CPSA) descriptors. The CPSA's were developed by Stanton et al.

References

Stanton, D.T. and Jurs, P.C. , Development and Use of Charged Partial Surface Area Structural Descriptors in Computer Assissted Quantitative Structure Property Relationship Studies, Analytical Chemistry, 1990, 62:2323.2329.

Examples

sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')
# \donttest{
mol = readMolFromSDF(sdf)
dat = extractDrugCPSA(mol)
head(dat)# }

#>     PPSA.1   PPSA.2   PPSA.3    PNSA.1     PNSA.2    PNSA.3    DPSA.1   DPSA.2
#> 1 217.3909 325.7632 22.89030 157.82863 -236.50828 -37.91801  59.56231 562.2715
#> 2 256.4949 369.4516 24.08549 147.07806 -211.84912 -26.85431 109.41688 581.3007
#> 3 218.2677 230.8050 15.71699  80.56272  -85.19024 -20.90756 137.70496 315.9952
#> 4 178.1162 238.5298 16.78033 105.92785 -141.85655 -32.19461  72.18831 380.3863
#> 5 151.3214 197.5453 19.57263 114.64248 -149.66213 -36.95420  36.67895 347.2075
#> 6 279.5318 500.6395 28.63370 113.24052 -202.81294 -32.18667 166.29132 703.4524
#>     DPSA.3    FPSA.1    FPSA.2     FPSA.3    FNSA.1     FNSA.2      FNSA.3
#> 1 60.80831 0.5793699 0.8681935 0.06100508 0.4206301 -0.6303197 -0.10105552
#> 2 50.93980 0.6355602 0.9154518 0.05968064 0.3644398 -0.5249338 -0.06654139
#> 3 36.62455 0.7304066 0.7723611 0.05259502 0.2695934 -0.2850789 -0.06996463
#> 4 48.97494 0.6270724 0.8397633 0.05907652 0.3729276 -0.4994175 -0.11334374
#> 5 56.52683 0.5689547 0.7427524 0.07359131 0.4310453 -0.5627159 -0.13894443
#> 6 60.82036 0.7116892 1.2746300 0.07290150 0.2883108 -0.5163625 -0.08194738
#>      WPSA.1    WPSA.2    WPSA.3   WNSA.1    WNSA.2     WNSA.3      RPCG
#> 1  81.56934 122.23273  8.588889 59.22039 -88.74254 -14.227579 0.1843093
#> 2 103.51443 149.10070  9.720255 59.35673 -85.49659 -10.837673 0.1917478
#> 3  65.22502  68.97154  4.696714 24.07459 -25.45743  -6.247813 0.2611178
#> 4  50.59283  67.75295  4.766353 30.08817 -40.29350  -9.144687 0.2079386
#> 5  40.24603  52.53993  5.205613 30.49076 -39.80472  -9.828483 0.2130907
#> 6 109.79239 196.63735 11.246525 44.47775 -79.65932 -12.642033 0.1538340
#>        RNCG      RPCS     RNCS     THSA      TPSA      RHSA      RPSA
#> 1 0.2405045 0.4521098 8.321894 212.3968 162.82282 0.5660599 0.4339401
#> 2 0.2293170 0.7598065 6.249868 294.7051 108.86788 0.7302399 0.2697601
#> 3 0.3123638 0.7308504 6.760782 202.4131  96.41732 0.6773510 0.3226490
#> 4 0.2916922 0.5820056 9.146075 149.4775 134.56651 0.5262477 0.4737523
#> 5 0.3014902 0.7237528 8.981059 100.5617 165.40224 0.3781027 0.6218973
#> 6 0.1844904 0.5515173 4.339869 253.1220 139.65039 0.6444496 0.3555504