A Variety of Descriptors Combining Surface Area and Partial Charge Information

A Variety of Descriptors Combining Surface Area and Partial Charge Information

extractDrugCPSA(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 29 columns:

PPSA.1 - partial positive surface area -- sum of surface area on positive parts of molecule

PPSA.2 - partial positive surface area * total positive charge on the molecule

PPSA.3 - charge weighted partial positive surface area

PNSA.1 - partial negative surface area -- sum of surface area on negative parts of molecule

PNSA.2 - partial negative surface area * total negative charge on the molecule

PNSA.3 - charge weighted partial negative surface area

DPSA.1 - difference of PPSA.1 and PNSA.1

DPSA.2 - difference of FPSA.2 and PNSA.2

DPSA.3 - difference of PPSA.3 and PNSA.3

FPSA.1 - PPSA.1 / total molecular surface area

FFSA.2 - PPSA.2 / total molecular surface area

FPSA.3 - PPSA.3 / total molecular surface area

FNSA.1 - PNSA.1 / total molecular surface area

FNSA.2 - PNSA.2 / total molecular surface area

FNSA.3 - PNSA.3 / total molecular surface area

WPSA.1 - PPSA.1 * total molecular surface area / 1000

WPSA.2 - PPSA.2 * total molecular surface area /1000

WPSA.3 - PPSA.3 * total molecular surface area / 1000

WNSA.1 - PNSA.1 * total molecular surface area /1000

WNSA.2 - PNSA.2 * total molecular surface area / 1000

WNSA.3 - PNSA.3 * total molecular surface area / 1000

RPCG - relative positive charge -- most positive charge / total positive charge

RNCG - relative negative charge -- most negative charge / total negative charge

RPCS - relative positive charge surface area -- most positive surface area * RPCG

RNCS - relative negative charge surface area -- most negative surface area * RNCG

THSA - sum of solvent accessible surface areas of atoms with absolute value of partial charges less than 0.2

TPSA - sum of solvent accessible surface areas of atoms with absolute value of partial charges greater than or equal 0.2

RHSA - THSA / total molecular surface area

RPSA - TPSA / total molecular surface area

Details

Calculates 29 Charged Partial Surface Area (CPSA) descriptors. The CPSA's were developed by Stanton et al.

References

Stanton, D.T. and Jurs, P.C. , Development and Use of Charged Partial Surface Area Structural Descriptors in Computer Assissted Quantitative Structure Property Relationship Studies, Analytical Chemistry, 1990, 62:2323.2329.

Examples

sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')

mol = readMolFromSDF(sdf)
dat = extractDrugCPSA(mol)
head(dat)

#>     PPSA.1   PPSA.2   PPSA.3   PNSA.1    PNSA.2    PNSA.3   DPSA.1   DPSA.2
#> 1 327.5371 490.8187 41.15625 181.0096 -271.2453 -38.96328 146.5275 762.0639
#> 2 313.7180 451.8749 32.42346 194.4487 -280.0811 -29.18968 119.2693 731.9560
#> 3 330.5395 349.5257 31.84418 107.8858 -114.0828 -26.48350 222.6537 463.6085
#> 4 326.9037 437.7832 41.35795 159.9431 -214.1927 -43.44578 166.9606 651.9759
#> 5 329.0320 429.5409 45.43773 160.0877 -208.9894 -52.69737 168.9443 638.5302
#> 6 397.9809 712.7809 46.72667 116.9178 -209.3989 -33.16350 281.0631 922.1798
#>     DPSA.3    FPSA.1    FPSA.2     FPSA.3    FNSA.1     FNSA.2      FNSA.3
#> 1 80.11953 0.6440650 0.9651398 0.08092915 0.3559350 -0.5333733 -0.07661691
#> 2 61.61315 0.6173525 0.8892257 0.06380477 0.3826475 -0.5511599 -0.05744115
#> 3 58.32769 0.7539243 0.7972297 0.07263308 0.2460757 -0.2602103 -0.06040595
#> 4 84.80374 0.6714714 0.8992216 0.08495066 0.3285286 -0.4399592 -0.08923913
#> 5 98.13511 0.6727025 0.8781918 0.09289697 0.3272975 -0.4272766 -0.10773923
#> 6 79.89018 0.7729305 1.3843128 0.09074925 0.2270695 -0.4066798 -0.06440782
#>     WPSA.1   WPSA.2   WPSA.3   WNSA.1    WNSA.2    WNSA.3      RPCG      RNCG
#> 1 166.5679 249.6042 20.92988 92.05183 -137.9409 -19.81465 0.1843093 0.2405045
#> 2 159.4210 229.6278 16.47653 98.81237 -142.3279 -14.83323 0.1917478 0.2293170
#> 3 144.9169 153.2409 13.96130 47.29989  -50.0168 -11.61104 0.2611178 0.3123638
#> 4 159.1520 213.1333 20.13499 77.86778 -104.2790 -21.15144 0.2079386 0.2916922
#> 5 160.9360 210.0969 22.22449 78.30201 -102.2208 -25.77532 0.2130907 0.3014902
#> 6 204.9199 367.0100 24.05950 60.20082 -107.8192 -17.07585 0.1538340 0.1844904
#>       RPCS      RNCS     THSA     TPSA      RHSA      RPSA
#> 1 4.602246  8.321894 286.4968 222.0499 0.5633638 0.4366362
#> 2 2.014090  6.249868 375.6811 132.4856 0.7392871 0.2607129
#> 3 6.323088  9.643010 273.8336 164.5918 0.6245842 0.3754158
#> 4 9.691373 10.591504 237.4827 249.3640 0.4877977 0.5122023
#> 5 7.445272 14.690814 208.0820 281.0376 0.4254215 0.5745785
#> 6 3.967054  4.434442 330.4110 184.4877 0.6417010 0.3582990