A Variety of Descriptors Combining Surface Area and Partial Charge Information

A Variety of Descriptors Combining Surface Area and Partial Charge Information

extractDrugCPSA(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 29 columns:

PPSA.1 - partial positive surface area -- sum of surface area on positive parts of molecule

PPSA.2 - partial positive surface area * total positive charge on the molecule

PPSA.3 - charge weighted partial positive surface area

PNSA.1 - partial negative surface area -- sum of surface area on negative parts of molecule

PNSA.2 - partial negative surface area * total negative charge on the molecule

PNSA.3 - charge weighted partial negative surface area

DPSA.1 - difference of PPSA.1 and PNSA.1

DPSA.2 - difference of FPSA.2 and PNSA.2

DPSA.3 - difference of PPSA.3 and PNSA.3

FPSA.1 - PPSA.1 / total molecular surface area

FFSA.2 - PPSA.2 / total molecular surface area

FPSA.3 - PPSA.3 / total molecular surface area

FNSA.1 - PNSA.1 / total molecular surface area

FNSA.2 - PNSA.2 / total molecular surface area

FNSA.3 - PNSA.3 / total molecular surface area

WPSA.1 - PPSA.1 * total molecular surface area / 1000

WPSA.2 - PPSA.2 * total molecular surface area /1000

WPSA.3 - PPSA.3 * total molecular surface area / 1000

WNSA.1 - PNSA.1 * total molecular surface area /1000

WNSA.2 - PNSA.2 * total molecular surface area / 1000

WNSA.3 - PNSA.3 * total molecular surface area / 1000

RPCG - relative positive charge -- most positive charge / total positive charge

RNCG - relative negative charge -- most negative charge / total negative charge

RPCS - relative positive charge surface area -- most positive surface area * RPCG

RNCS - relative negative charge surface area -- most negative surface area * RNCG

THSA - sum of solvent accessible surface areas of atoms with absolute value of partial charges less than 0.2

TPSA - sum of solvent accessible surface areas of atoms with absolute value of partial charges greater than or equal 0.2

RHSA - THSA / total molecular surface area

RPSA - TPSA / total molecular surface area

Details

Calculates 29 Charged Partial Surface Area (CPSA) descriptors. The CPSA's were developed by Stanton et al.

References

Stanton, D.T. and Jurs, P.C. , Development and Use of Charged Partial Surface Area Structural Descriptors in Computer Assissted Quantitative Structure Property Relationship Studies, Analytical Chemistry, 1990, 62:2323.2329.

Examples

# NOT RUN {
sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')
# }# NOT RUN {
mol = readMolFromSDF(sdf)
dat = extractDrugCPSA(mol)
head(dat)
# }