Evaluates the Kier and Hall Chi cluster indices of orders 3, 4, 5 and 6

Evaluates the Kier and Hall Chi cluster indices of orders 3, 4, 5 and 6

Usage

extractDrugChiCluster(molecules, silent = TRUE)

Arguments

molecules

Parsed molucule object.

silent

Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 8 columns, the order and names of the columns returned is:

SC.3 - Simple cluster, order 3

SC.4 - Simple cluster, order 4

SC.5 - Simple cluster, order 5

SC.6 - Simple cluster, order 6

VC.3 - Valence cluster, order 3

VC.4 - Valence cluster, order 4

VC.5 - Valence cluster, order 5

VC.6 - Valence cluster, order 6

Details

Evaluates chi cluster descriptors. It utilizes the graph isomorphism code of the CDK to find fragments matching SMILES strings representing the fragments corresponding to each type of chain.

Note

These descriptors are calculated using graph isomorphism to identify the various fragments. As a result calculations may be slow. In addition, recent versions of Molconn-Z use simplified fragment definitions (i.e., rings without branches etc.) whereas these descriptors use the older more complex fragment definitions.

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugChiCluster(mol)
head(dat)# }
#>                                                              SC.3      SC.4
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 2.0530106 0.0000000
#> c1(ccc(cc1)O)NC(=O)C                                    0.9010476 0.0000000
#> c1ccccc1NC(=O)C                                         0.6123724 0.0000000
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              2.1227354 0.2886751
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                2.1876137 0.2041241
#> CC(=O)NO                                                0.4082483 0.0000000
#>                                                              SC.5      SC.6
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 0.5693952 0.0000000
#> c1(ccc(cc1)O)NC(=O)C                                    0.0000000 0.0000000
#> c1ccccc1NC(=O)C                                         0.0000000 0.0000000
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              0.4330127 0.1443376
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                0.5151284 0.1020621
#> CC(=O)NO                                                0.0000000 0.0000000
#>                                                              VC.3       VC.4
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 0.7613073 0.00000000
#> c1(ccc(cc1)O)NC(=O)C                                    0.2599310 0.00000000
#> c1ccccc1NC(=O)C                                         0.1853954 0.00000000
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              0.6607285 0.02944439
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                0.7096826 0.02549959
#> CC(=O)NO                                                0.1020621 0.00000000
#>                                                               VC.5        VC.6
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 0.12084955 0.000000000
#> c1(ccc(cc1)O)NC(=O)C                                    0.00000000 0.000000000
#> c1ccccc1NC(=O)C                                         0.00000000 0.000000000
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              0.10694097 0.016127354
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                0.09266106 0.008499862
#> CC(=O)NO                                                0.00000000 0.000000000