Calculate the Kier and Hall Chi Path Indices of Orders 0 to 7

Calculate the Kier and Hall Chi Path Indices of Orders 0 to 7

extractDrugChiPath(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 16 columns, The order and names of the columns returned is:

SP.0, SP.1, ..., SP.7 - Simple path, orders 0 to 7

VP.0, VP.1, ..., VP.7 - Valence path, orders 0 to 7

Details

Evaluates chi path descriptors. This function utilizes the graph isomorphism code of the CDK to find fragments matching SMILES strings representing the fragments corresponding to each type of chain.

Note

These descriptors are calculated using graph isomorphism to identify the various fragments. As a result calculations may be slow. In addition, recent versions of Molconn-Z use simplified fragment definitions (i.e., rings without branches etc.) whereas these descriptors use the older more complex fragment definitions.

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugChiPath(mol)
head(dat)# }

#>                                                              SP.0      SP.1
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 18.844571 12.379918
#> c1(ccc(cc1)O)NC(=O)C                                     8.267585  5.181541
#> c1ccccc1NC(=O)C                                          7.397341  4.787694
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              10.060478  5.892865
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                16.164926 10.381847
#> CC(=O)NO                                                 4.284457  2.270056
#>                                                              SP.2      SP.3
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 11.570125 9.1847024
#> c1(ccc(cc1)O)NC(=O)C                                     4.755447 2.9401663
#> c1ccccc1NC(=O)C                                          4.133585 2.5294827
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                               6.371576 3.5707105
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                 9.893336 7.1548274
#> CC(=O)NO                                                 1.802095 0.8164966
#>                                                             SP.4     SP.5
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 7.919385 6.404528
#> c1(ccc(cc1)O)NC(=O)C                                    2.231484 1.745558
#> c1ccccc1NC(=O)C                                         2.089509 1.454568
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              2.742267 1.705791
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                5.707323 3.806775
#> CC(=O)NO                                                0.000000 0.000000
#>                                                              SP.6      SP.7
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 4.6414117 3.0349134
#> c1(ccc(cc1)O)NC(=O)C                                    0.7889911 0.5326570
#> c1ccccc1NC(=O)C                                         0.5955220 0.3190356
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              1.1733040 0.6512287
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                2.4947518 1.8451355
#> CC(=O)NO                                                0.0000000 0.0000000
#>                                                              VP.0     VP.1
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 13.690573 7.920201
#> c1(ccc(cc1)O)NC(=O)C                                     6.164863 3.249098
#> c1ccccc1NC(=O)C                                          5.795000 3.114808
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                               7.533257 4.235098
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                12.667268 7.696465
#> CC(=O)NO                                                 2.855462 1.177731
#>                                                              VP.2      VP.3
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 5.9926052 4.2086673
#> c1(ccc(cc1)O)NC(=O)C                                    2.2280602 1.1876733
#> c1ccccc1NC(=O)C                                         2.0472116 1.0941577
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              3.4816990 2.1186185
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                5.9142707 4.0746883
#> CC(=O)NO                                                0.6679896 0.1574469
#>                                                              VP.4      VP.5
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 2.9753813 2.1000778
#> c1(ccc(cc1)O)NC(=O)C                                    0.7480266 0.4541008
#> c1ccccc1NC(=O)C                                         0.7247493 0.4134722
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              1.3500910 0.6781185
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                2.7399233 1.8180018
#> CC(=O)NO                                                0.0000000 0.0000000
#>                                                              VP.6       VP.7
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O 1.2066002 0.68416719
#> c1(ccc(cc1)O)NC(=O)C                                    0.1539902 0.07400724
#> c1ccccc1NC(=O)C                                         0.1276072 0.05515552
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                              0.3611446 0.15650696
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                0.8366621 0.48579813
#> CC(=O)NO                                                0.0000000 0.00000000