Descriptor Characterizing the Mass Distribution of the Molecule.

Descriptor Characterizing the Mass Distribution of the Molecule.

Usage

extractDrugGravitationalIndex(molecules, silent = TRUE)

Arguments

molecules

Parsed molucule object.

silent

Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 9 columns:

GRAV.1 - gravitational index of heavy atoms

GRAV.2 - square root of gravitational index of heavy atoms

GRAV.3 - cube root of gravitational index of heavy atoms

GRAVH.1 - gravitational index - hydrogens included

GRAVH.2 - square root of hydrogen-included gravitational index

GRAVH.3 - cube root of hydrogen-included gravitational index

GRAV.4 - grav1 for all pairs of atoms (not just bonded pairs)

GRAV.5 - grav2 for all pairs of atoms (not just bonded pairs)

GRAV.6 - grav3 for all pairs of atoms (not just bonded pairs)

Details

Descriptor characterizing the mass distribution of the molecule described by Katritzky et al. For modelling purposes the value of the descriptor is calculated both with and without H atoms. Furthermore the square and cube roots of the descriptor are also generated as described by Wessel et al.

References

Katritzky, A.R. and Mu, L. and Lobanov, V.S. and Karelson, M., Correlation of Boiling Points With Molecular Structure. 1. A Training Set of 298 Diverse Organics and a Test Set of 9 Simple Inorganics, J. Phys. Chem., 1996, 100:10400-10407.

Wessel, M.D. and Jurs, P.C. and Tolan, J.W. and Muskal, S.M. , Prediction of Human Intestinal Absorption of Drug Compounds From Molecular Structure, Journal of Chemical Information and Computer Sciences, 1998, 38:726-735.

Examples

sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')
# \donttest{
mol = readMolFromSDF(sdf)
dat = extractDrugGravitationalIndex(mol)
head(dat)# }
#>      GRAV.1   GRAV.2    GRAV.3   GRAVH.1  GRAVH.2   GRAVH.3    GRAV.4   GRAV.5
#> 1 1043.1603 32.29799 10.141846 1174.1792 34.26630 10.549814 2014.8030 44.88656
#> 2 1262.1186 35.52631 10.806873 1400.0996 37.41790 11.187155 2495.5487 49.95547
#> 3  555.1187 23.56096  8.218551  678.6475 26.05086  8.787825 1014.4004 31.84965
#> 4  587.5330 24.23908  8.375500  697.4401 26.40909  8.868203 1102.8203 33.20874
#> 5  528.8768 22.99732  8.086951  619.2162 24.88406  8.523424  916.9202 30.28069
#> 6  920.7589 30.34401  9.728562 1089.2443 33.00370 10.289046 1719.6475 41.46863
#>      GRAV.6
#> 1 12.630218
#> 2 13.564028
#> 3 10.047773
#> 4 10.331616
#> 5  9.715023
#> 6 11.980634