Calculate the Hybridization Molecular Fingerprints (in Compact Format)

Calculate the Hybridization Molecular Fingerprints (in Compact Format)

Usage

extractDrugHybridization(molecules, depth = 6, size = 1024, silent = TRUE)

Arguments

molecules

Parsed molucule object.

depth

The search depth. Default is 6.

size

The length of the fingerprint bit string. Default is 1024.

silent

Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A list, each component represents one of the molecules, each element in the component represents the index of which element in the fingerprint is 1. Each component's name is the length of the fingerprints.

Details

Calculate the hybridization molecular fingerprints. Similar to the standard type, but only consider hybridization state. This is hashed fingerprints, with a default length of 1024.

See also

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
fp  = extractDrugHybridization(mol)
head(fp)# }
#> $`1024`
#>   [1]   11   14   31   58   67   71   78   96  107  108  109  138  139  149  159
#>  [16]  161  162  170  174  196  198  206  214  221  223  234  239  247  250  254
#>  [31]  272  273  284  286  301  302  304  312  331  373  392  394  398  401  403
#>  [46]  405  410  419  427  440  444  449  467  472  491  492  493  498  499  514
#>  [61]  516  536  541  543  553  563  575  585  614  622  636  638  658  666  670
#>  [76]  671  681  686  714  717  718  734  737  742  743  744  746  749  753  756
#>  [91]  763  820  832  851  858  893  910  930  932  933  963  966  981  984  994
#> [106] 1021
#> 
#> $`1024`
#>  [1]  96 119 162 206 208 214 238 247 254 302 352 392 403 436 459 475 505 506 551
#> [20] 575 604 614 638 704 737 742 743 746 780 901 930 931 932 961 966 970 981 984
#> [39] 994
#> 
#> $`1024`
#>  [1]  96 162 214 238 247 254 302 352 403 436 459 475 505 506 551 575 604 638 704
#> [20] 742 743 780 901 931 932 961 966 970 981 984 994
#> 
#> $`1024`
#>  [1]   12   23   56   77   78   82   85   86  113  149  156  160  163  184  195
#> [16]  197  208  210  232  238  242  245  286  302  305  306  318  321  328  331
#> [31]  352  354  369  375  377  383  403  407  414  436  452  457  461  472  475
#> [46]  528  529  545  564  570  603  631  638  656  740  742  743  745  748  788
#> [61]  807  808  891  908  910  914  918  960  961  962  966  970  974  980  984
#> [76]  994 1009 1017 1018 1023
#> 
#> $`1024`
#>  [1]    3   14   17   18   31   57   58   70   78   96  136  146  149  159  162
#> [16]  169  170  172  174  187  195  209  214  216  221  223  224  238  239  248
#> [31]  250  254  272  273  289  297  305  312  331  352  353  354  375  403  411
#> [46]  414  420  433  439  455  457  461  494  499  500  518  526  528  536  543
#> [61]  557  564  575  597  638  655  742  743  745  748  753  765  781  782  827
#> [76]  831  845  885  892  904  913  917  920  954  966  984  994 1008
#> 
#> $`1024`
#>  [1] 238 403 492 638 719 742 743 932 961 966 984
#>