Descriptor that Counts the Number of Occurrences of the E-State Fragments

Descriptor that Counts the Number of Occurrences of the E-State Fragments

extractDrugKierHallSmarts(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 79 columns:

ID	Name	Pattern

0	khs.sLi	[LiD1]-*

1	khs.ssBe	[BeD2](-*)-*

2	khs.ssssBe	[BeD4](-*)(-*)(-*)-*

3	khs.ssBH	[BD2H](-*)-*

4	khs.sssB	[BD3](-*)(-*)-*

5	khs.ssssB	[BD4](-*)(-*)(-*)-*

6	khs.sCH3	[CD1H3]-*

7	khs.dCH2	[CD1H2]=*

8	khs.ssCH2	[CD2H2](-*)-*

9	khs.tCH	[CD1H]#*

10	khs.dsCH	[CD2H](=*)-*

11	khs.aaCH	[C,c;D2H](:*):*

12	khs.sssCH	[CD3H](-*)(-*)-*

13	khs.ddC	[CD2H0](=*)=*

14	khs.tsC	[CD2H0](#*)-*

15	khs.dssC	[CD3H0](=*)(-*)-*

16	khs.aasC	[C,c;D3H0](:*)(:*)-*

17	khs.aaaC	[C,c;D3H0](:*)(:*):*

18	khs.ssssC	[CD4H0](-*)(-*)(-*)-*

19	khs.sNH3	[ND1H3]-*

20	khs.sNH2	[ND1H2]-*

21	khs.ssNH2	[ND2H2](-*)-*

22	khs.dNH	[ND1H]=*

23	khs.ssNH	[ND2H](-*)-*

24	khs.aaNH	[N,nD2H](:*):*

25	khs.tN	[ND1H0]#*

26	khs.sssNH	[ND3H](-*)(-*)-*

27	khs.dsN	[ND2H0](=*)-*

28	khs.aaN	[N,nD2H0](:*):*

29	khs.sssN	[ND3H0](-*)(-*)-*

30	khs.ddsN	[ND3H0](~[OD1H0])(~[OD1H0])-,:*

31	khs.aasN	[N,nD3H0](:*)(:*)-,:*

32	khs.ssssN	[ND4H0](-*)(-*)(-*)-*

33	khs.sOH	[OD1H]-*

34	khs.dO	[OD1H0]=*

35	khs.ssO	[OD2H0](-*)-*

36	khs.aaO	[O,oD2H0](:*):*

37	khs.sF	[FD1]-*

38	khs.sSiH3	[SiD1H3]-*

39	khs.ssSiH2	[SiD2H2](-*)-*

40	khs.sssSiH	[SiD3H1](-*)(-*)-*

41	khs.ssssSi	[SiD4H0](-*)(-*)(-*)-*

42	khs.sPH2	[PD1H2]-*

43	khs.ssPH	[PD2H1](-*)-*

44	khs.sssP	[PD3H0](-*)(-*)-*

45	khs.dsssP	[PD4H0](=*)(-*)(-*)-*

46	khs.sssssP	[PD5H0](-*)(-*)(-*)(-*)-*

47	khs.sSH	[SD1H1]-*

48	khs.dS	[SD1H0]=*

49	khs.ssS	[SD2H0](-*)-*

50	khs.aaS	[S,sD2H0](:*):*

51	khs.dssS	[SD3H0](=*)(-*)-*

52	khs.ddssS	[SD4H0](~[OD1H0])(~[OD1H0])(-*)-*

53	khs.sCl	[ClD1]-*

54	khs.sGeH3	[GeD1H3](-*)

55	khs.ssGeH2	[GeD2H2](-*)-*

56	khs.sssGeH	[GeD3H1](-*)(-*)-*

57	khs.ssssGe	[GeD4H0](-*)(-*)(-*)-*

58	khs.sAsH2	[AsD1H2]-*

59	khs.ssAsH	[AsD2H1](-*)-*

60	khs.sssAs	[AsD3H0](-*)(-*)-*

61	khs.sssdAs	[AsD4H0](=*)(-*)(-*)-*

62	khs.sssssAs	[AsD5H0](-*)(-*)(-*)(-*)-*

63	khs.sSeH	[SeD1H1]-*

64	khs.dSe	[SeD1H0]=*

65	khs.ssSe	[SeD2H0](-*)-*

66	khs.aaSe	[SeD2H0](:*):*

67	khs.dssSe	[SeD3H0](=*)(-*)-*

68	khs.ddssSe	[SeD4H0](=*)(=*)(-*)-*

69	khs.sBr	[BrD1]-*

70	khs.sSnH3	[SnD1H3]-*

71	khs.ssSnH2	[SnD2H2](-*)-*

72	khs.sssSnH	[SnD3H1](-*)(-*)-*

73	khs.ssssSn	[SnD4H0](-*)(-*)(-*)-*

74	khs.sI	[ID1]-*

75	khs.sPbH3	[PbD1H3]-*

76	khs.ssPbH2	[PbD2H2](-*)-*

77	khs.sssPbH	[PbD3H1](-*)(-*)-*

78	khs.ssssPb	[PbD4H0](-*)(-*)(-*)-*

Details

A fragment count descriptor that uses e-state fragments. Traditionally the e-state descriptors identify the relevant fragments and then evaluate the actual e-state value. However it has been shown in Butina et al. that simply using the counts of the e-state fragments can lead to QSAR models that exhibit similar performance to those built using the actual e-state indices.

Atom typing and aromaticity perception should be performed prior to calling this descriptor. The atom type definitions are taken from Hall et al. The SMARTS definitions were obtained from RDKit.

References

Butina, D. , Performance of Kier-Hall E-state Descriptors in Quantitative Structure Activity Relationship (QSAR) Studies of Multifunctional Molecules, Molecules, 2004, 9:1004-1009.

Hall, L.H. and Kier, L.B. , Electrotopological State Indices for Atom Types: A Novel Combination of Electronic, Topological, and Valence State Information, Journal of Chemical Information and Computer Science, 1995, 35:1039-1045.

Author

Nan Xiao <https://nanx.me>

Examples

smi = system.file('vignettedata/FDAMDD.smi', package = 'Rcpi')
# \donttest{
mol = readMolFromSmi(smi, type = 'mol')
dat = extractDrugKierHallSmarts(mol)
head(dat)# }

#>                                                         khs.sLi khs.ssBe
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0        0
#> c1(ccc(cc1)O)NC(=O)C                                          0        0
#> c1ccccc1NC(=O)C                                               0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0        0
#> CC(=O)NO                                                      0        0
#>                                                         khs.ssssBe khs.ssBH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O          0        0
#> c1(ccc(cc1)O)NC(=O)C                                             0        0
#> c1ccccc1NC(=O)C                                                  0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                       0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                         0        0
#> CC(=O)NO                                                         0        0
#>                                                         khs.sssB khs.ssssB
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        0         0
#> c1(ccc(cc1)O)NC(=O)C                                           0         0
#> c1ccccc1NC(=O)C                                                0         0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     0         0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       0         0
#> CC(=O)NO                                                       0         0
#>                                                         khs.sCH3 khs.dCH2
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        1        0
#> c1(ccc(cc1)O)NC(=O)C                                           1        0
#> c1ccccc1NC(=O)C                                                1        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     1        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       1        0
#> CC(=O)NO                                                       1        0
#>                                                         khs.ssCH2 khs.tCH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         1       0
#> c1(ccc(cc1)O)NC(=O)C                                            0       0
#> c1ccccc1NC(=O)C                                                 0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        5       0
#> CC(=O)NO                                                        0       0
#>                                                         khs.dsCH khs.aaCH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        0        8
#> c1(ccc(cc1)O)NC(=O)C                                           0        4
#> c1ccccc1NC(=O)C                                                0        5
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       0        4
#> CC(=O)NO                                                       0        0
#>                                                         khs.sssCH khs.ddC
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         1       0
#> c1(ccc(cc1)O)NC(=O)C                                            0       0
#> c1ccccc1NC(=O)C                                                 0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        1       0
#> CC(=O)NO                                                        0       0
#>                                                         khs.tsC khs.dssC
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0        1
#> c1(ccc(cc1)O)NC(=O)C                                          0        1
#> c1ccccc1NC(=O)C                                               0        1
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0        1
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0        2
#> CC(=O)NO                                                      0        1
#>                                                         khs.aasC khs.aaaC
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        4        2
#> c1(ccc(cc1)O)NC(=O)C                                           2        0
#> c1ccccc1NC(=O)C                                                1        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     2        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       2        0
#> CC(=O)NO                                                       0        0
#>                                                         khs.ssssC khs.sNH3
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0        0
#> c1(ccc(cc1)O)NC(=O)C                                            0        0
#> c1ccccc1NC(=O)C                                                 0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        0        0
#> CC(=O)NO                                                        0        0
#>                                                         khs.sNH2 khs.ssNH2
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        0         0
#> c1(ccc(cc1)O)NC(=O)C                                           0         0
#> c1ccccc1NC(=O)C                                                0         0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     1         0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       0         0
#> CC(=O)NO                                                       0         0
#>                                                         khs.dNH khs.ssNH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0        0
#> c1(ccc(cc1)O)NC(=O)C                                          0        1
#> c1ccccc1NC(=O)C                                               0        1
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0        1
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0        2
#> CC(=O)NO                                                      0        1
#>                                                         khs.aaNH khs.tN
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        0      0
#> c1(ccc(cc1)O)NC(=O)C                                           0      0
#> c1ccccc1NC(=O)C                                                0      0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     0      0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       0      0
#> CC(=O)NO                                                       0      0
#>                                                         khs.sssNH khs.dsN
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0       0
#> c1(ccc(cc1)O)NC(=O)C                                            0       0
#> c1ccccc1NC(=O)C                                                 0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        0       0
#> CC(=O)NO                                                        0       0
#>                                                         khs.aaN khs.sssN
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0        0
#> c1(ccc(cc1)O)NC(=O)C                                          0        0
#> c1ccccc1NC(=O)C                                               0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    2        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0        0
#> CC(=O)NO                                                      0        0
#>                                                         khs.ddsN khs.aasN
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        1        0
#> c1(ccc(cc1)O)NC(=O)C                                           0        0
#> c1ccccc1NC(=O)C                                                0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       0        0
#> CC(=O)NO                                                       0        0
#>                                                         khs.ssssN khs.sOH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0       1
#> c1(ccc(cc1)O)NC(=O)C                                            0       1
#> c1ccccc1NC(=O)C                                                 0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        0       0
#> CC(=O)NO                                                        0       1
#>                                                         khs.dO khs.ssO khs.aaO
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O      3       0       1
#> c1(ccc(cc1)O)NC(=O)C                                         1       0       0
#> c1ccccc1NC(=O)C                                              1       0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                   3       0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                     4       0       0
#> CC(=O)NO                                                     1       0       0
#>                                                         khs.sF khs.sSiH3
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O      0         0
#> c1(ccc(cc1)O)NC(=O)C                                         0         0
#> c1ccccc1NC(=O)C                                              0         0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                   0         0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                     0         0
#> CC(=O)NO                                                     0         0
#>                                                         khs.ssSiH2 khs.sssSiH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O          0          0
#> c1(ccc(cc1)O)NC(=O)C                                             0          0
#> c1ccccc1NC(=O)C                                                  0          0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                       0          0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                         0          0
#> CC(=O)NO                                                         0          0
#>                                                         khs.ssssSi khs.sPH2
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O          0        0
#> c1(ccc(cc1)O)NC(=O)C                                             0        0
#> c1ccccc1NC(=O)C                                                  0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                       0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                         0        0
#> CC(=O)NO                                                         0        0
#>                                                         khs.ssPH khs.sssP
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        0        0
#> c1(ccc(cc1)O)NC(=O)C                                           0        0
#> c1ccccc1NC(=O)C                                                0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       0        0
#> CC(=O)NO                                                       0        0
#>                                                         khs.dsssP khs.sssssP
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0          0
#> c1(ccc(cc1)O)NC(=O)C                                            0          0
#> c1ccccc1NC(=O)C                                                 0          0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0          0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        0          0
#> CC(=O)NO                                                        0          0
#>                                                         khs.sSH khs.dS khs.ssS
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0      0       0
#> c1(ccc(cc1)O)NC(=O)C                                          0      0       0
#> c1ccccc1NC(=O)C                                               0      0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0      0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0      0       0
#> CC(=O)NO                                                      0      0       0
#>                                                         khs.aaS khs.dssS
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0        0
#> c1(ccc(cc1)O)NC(=O)C                                          0        0
#> c1ccccc1NC(=O)C                                               0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    1        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0        0
#> CC(=O)NO                                                      0        0
#>                                                         khs.ddssS khs.sCl
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0       0
#> c1(ccc(cc1)O)NC(=O)C                                            0       0
#> c1ccccc1NC(=O)C                                                 0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      1       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        1       0
#> CC(=O)NO                                                        0       0
#>                                                         khs.sGeH3 khs.ssGeH2
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0          0
#> c1(ccc(cc1)O)NC(=O)C                                            0          0
#> c1ccccc1NC(=O)C                                                 0          0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0          0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        0          0
#> CC(=O)NO                                                        0          0
#>                                                         khs.sssGeH khs.ssssGe
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O          0          0
#> c1(ccc(cc1)O)NC(=O)C                                             0          0
#> c1ccccc1NC(=O)C                                                  0          0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                       0          0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                         0          0
#> CC(=O)NO                                                         0          0
#>                                                         khs.sAsH2 khs.ssAsH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0         0
#> c1(ccc(cc1)O)NC(=O)C                                            0         0
#> c1ccccc1NC(=O)C                                                 0         0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0         0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        0         0
#> CC(=O)NO                                                        0         0
#>                                                         khs.sssAs khs.sssdAs
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0          0
#> c1(ccc(cc1)O)NC(=O)C                                            0          0
#> c1ccccc1NC(=O)C                                                 0          0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0          0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        0          0
#> CC(=O)NO                                                        0          0
#>                                                         khs.sssssAs khs.sSeH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O           0        0
#> c1(ccc(cc1)O)NC(=O)C                                              0        0
#> c1ccccc1NC(=O)C                                                   0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                        0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                          0        0
#> CC(=O)NO                                                          0        0
#>                                                         khs.dSe khs.ssSe
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O       0        0
#> c1(ccc(cc1)O)NC(=O)C                                          0        0
#> c1ccccc1NC(=O)C                                               0        0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                    0        0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                      0        0
#> CC(=O)NO                                                      0        0
#>                                                         khs.aaSe khs.dssSe
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O        0         0
#> c1(ccc(cc1)O)NC(=O)C                                           0         0
#> c1ccccc1NC(=O)C                                                0         0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                     0         0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                       0         0
#> CC(=O)NO                                                       0         0
#>                                                         khs.ddssSe khs.sBr
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O          0       0
#> c1(ccc(cc1)O)NC(=O)C                                             0       0
#> c1ccccc1NC(=O)C                                                  0       0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                       0       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                         0       0
#> CC(=O)NO                                                         0       0
#>                                                         khs.sSnH3 khs.ssSnH2
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O         0          0
#> c1(ccc(cc1)O)NC(=O)C                                            0          0
#> c1ccccc1NC(=O)C                                                 0          0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                      0          0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                        0          0
#> CC(=O)NO                                                        0          0
#>                                                         khs.sssSnH khs.ssssSn
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O          0          0
#> c1(ccc(cc1)O)NC(=O)C                                             0          0
#> c1ccccc1NC(=O)C                                                  0          0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                       0          0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                         0          0
#> CC(=O)NO                                                         0          0
#>                                                         khs.sI khs.sPbH3
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O      0         0
#> c1(ccc(cc1)O)NC(=O)C                                         0         0
#> c1ccccc1NC(=O)C                                              0         0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                   0         0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                     0         0
#> CC(=O)NO                                                     0         0
#>                                                         khs.ssPbH2 khs.sssPbH
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O          0          0
#> c1(ccc(cc1)O)NC(=O)C                                             0          0
#> c1ccccc1NC(=O)C                                                  0          0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                       0          0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                         0          0
#> CC(=O)NO                                                         0          0
#>                                                         khs.ssssPb
#> c1ccc2c(c1)oc(=O)c(C(CC(=O)C)c1ccc([N+](=O)[O-])cc1)c2O          0
#> c1(ccc(cc1)O)NC(=O)C                                             0
#> c1ccccc1NC(=O)C                                                  0
#> n1nc(S(=O)(=O)N)sc1NC(=O)C                                       0
#> O=S(=O)(c1ccc(cc1)C(=O)C)NC(=O)NC1CCCCC1                         0
#> CC(=O)NO                                                         0