Descriptor that Calculates the Principal Moments of Inertia and Ratios of the Principal Moments

Descriptor that Calculates the Principal Moments of Inertia and Ratios of the Principal Moments

extractDrugMomentOfInertia(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 7 columns named MOMI.X, MOMI.Y, MOMI.Z, MOMI.XY, MOMI.XZ, MOMI.YZ, MOMI.R:

MOMI.X - MI along X axis

MOMI.Y - MI along Y axis

MOMI.Z - MI along Z axis

MOMI.XY - X/Y

MOMI.XZ - X/Z

MOMI.YZ - Y/Z

MOMI.R - Radius of gyration

One important aspect of the algorithm is that if the eigenvalues of the MI tensor are below 1e-3, then the ratio's are set to a default of 1000.

Details

A descriptor that calculates the moment of inertia and radius of gyration. Moment of inertia (MI) values characterize the mass distribution of a molecule. Related to the MI values, ratios of the MI values along the three principal axes are also well know modeling variables. This descriptor calculates the MI values along the X, Y and Z axes as well as the ratio's X/Y, X/Z and Y/Z. Finally it also calculates the radius of gyration of the molecule.

Author

Nan Xiao <https://nanx.me>

Examples

sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')
# \donttest{
mol = readMolFromSDF(sdf)
dat = extractDrugMomentOfInertia(mol)
head(dat)# }

#>      MOMI.X    MOMI.Y   MOMI.Z  MOMI.XY  MOMI.XZ  MOMI.YZ   MOMI.R
#> 1 1273.6497 1171.0788 306.7954 1.087587 4.151463 3.817133 5.171376
#> 2 1551.0291 1303.5889 450.6547 1.189815 3.441724 2.892656 5.463571
#> 3  365.1319  290.3866 211.5703 1.257399 1.725818 1.372530 3.890696
#> 4  372.8907  277.4748 194.3705 1.343872 1.918453 1.427556 3.788606
#> 5  313.0617  217.0054 158.2076 1.442645 1.978803 1.371649 3.633500
#> 6 1516.4741 1341.8793 282.4184 1.130112 5.369601 4.751387 5.477884