Calculate Holistic Descriptors Described by Todeschini et al.

Calculate Holistic Descriptors Described by Todeschini et al.

Usage

extractDrugWHIM(molecules, silent = TRUE)

Arguments

molecules

Parsed molucule object.

silent

Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 17 columns:

Wlambda1

Wlambda2

wlambda3

Wnu1

Wnu2

Wgamma1

Wgamma2

Wgamma3

Weta1

Weta2

Weta3

WT

WA

WV

WK

WG

WD

Each name will have a suffix of the form .X where X indicates the weighting scheme used. Possible values of X are

unity

mass

volume

eneg

polar

Details

Holistic descriptors described by Todeschini et al, the descriptors are based on a number of atom weightings. There are six different possible weightings:

unit weights

atomic masses

van der Waals volumes

Mulliken atomic electronegativites

atomic polarizabilities

E-state values described by Kier and Hall

Currently weighting schemes 1, 2, 3, 4 and 5 are implemented. The weight values are taken from Todeschini et al. and as a result 19 elements are considered. For each weighting scheme we can obtain

11 directional WHIM descriptors (lambda1 .. 3, nu1 .. 2, gamma1 .. 3, eta1 .. 3)

6 non-directional WHIM descriptors (T, A, V, K, G, D)

Though Todeschini et al. mentions that for planar molecules only 8 directional WHIM descriptors are required the current code will return all 11.

References

Todeschini, R. and Gramatica, P., New 3D Molecular Descriptors: The WHIM theory and QAR Applications, Persepectives in Drug Discovery and Design, 1998, ?:355-380.

Examples

sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')
# \donttest{
mol = readMolFromSDF(sdf)
#> Error in loadMolecules(normalizePath(sdffile)): The package "rcdk" is required to load molecular structures
dat = extractDrugWHIM(mol)
#> Error in evaluateDescriptor(molecules, type = "WHIMDescriptor", silent = silent): The package "rcdk" is required to compute molecular descriptors
head(dat)# }
#> Error in eval(expr, envir, enclos): object 'dat' not found