Calculate Holistic Descriptors Described by Todeschini et al.

Calculate Holistic Descriptors Described by Todeschini et al.

extractDrugWHIM(molecules, silent = TRUE)

Arguments

molecules
Parsed molucule object.

silent
Logical. Whether the calculating process should be shown or not, default is TRUE.

Value

A data frame, each row represents one of the molecules, each column represents one feature. This function returns 17 columns:

Wlambda1

Wlambda2

wlambda3

Wnu1

Wnu2

Wgamma1

Wgamma2

Wgamma3

Weta1

Weta2

Weta3

WT

WA

WV

WK

WG

WD

Each name will have a suffix of the form .X where X indicates the weighting scheme used. Possible values of X are

unity

mass

volume

eneg

polar

Details

Holistic descriptors described by Todeschini et al, the descriptors are based on a number of atom weightings. There are six different possible weightings:

unit weights

atomic masses

van der Waals volumes

Mulliken atomic electronegativites

atomic polarizabilities

E-state values described by Kier and Hall

Currently weighting schemes 1, 2, 3, 4 and 5 are implemented. The weight values are taken from Todeschini et al. and as a result 19 elements are considered. For each weighting scheme we can obtain

11 directional WHIM descriptors (lambda1 .. 3, nu1 .. 2, gamma1 .. 3, eta1 .. 3)

6 non-directional WHIM descriptors (T, A, V, K, G, D)

Though Todeschini et al. mentions that for planar molecules only 8 directional WHIM descriptors are required the current code will return all 11.

References

Todeschini, R. and Gramatica, P., New 3D Molecular Descriptors: The WHIM theory and QAR Applications, Persepectives in Drug Discovery and Design, 1998, ?:355-380.

Examples

sdf = system.file('sysdata/OptAA3d.sdf', package = 'Rcpi')
# \donttest{
mol = readMolFromSDF(sdf)
dat = extractDrugWHIM(mol)
head(dat)# }

#>   Wlambda1.unity Wlambda2.unity Wlambda3.unity Wnu1.unity Wnu2.unity
#> 1      0.9148748       1.701534       6.056852 0.10548221  0.1961816
#> 2      0.8052440       2.366736       7.105803 0.07834802  0.2302769
#> 3      1.0418358       1.743401       2.260112 0.20649429  0.3455462
#> 4      0.7630751       1.778282       2.037981 0.16663435  0.3883273
#> 5      0.5726677       1.586719       1.948499 0.13940694  0.3862616
#> 6      0.5693994       1.861933       7.803989 0.05563083  0.1819125
#>   Wgamma1.unity Wgamma2.unity Wgamma3.unity Weta1.unity Weta2.unity Weta3.unity
#> 1           NaN           NaN           NaN   0.4181859   0.4843001   0.5091855
#> 2           NaN           NaN           NaN   0.3537120   0.4569316   0.5467652
#> 3           NaN           NaN           NaN   0.5090874   0.4764000   0.6107750
#> 4           NaN           NaN           NaN   0.5168376   0.4269940   0.5609107
#> 5           NaN           NaN           NaN   0.3795667   0.4857948   0.5923140
#> 6           NaN           NaN           NaN   0.3763019   0.4371447   0.5757821
#>    WT.unity  WA.unity WV.unity   WK.unity WG.unity WD.unity
#> 1  8.673261 17.403893 35.50580 0.27375215      NaN 1.411672
#> 2 10.277784 24.445269 48.26530 0.26853129      NaN 1.357409
#> 3  5.045349  8.111285 17.26176 0.08596964      NaN 1.596262
#> 4  4.579339  6.536201 13.88100 0.08377878      NaN 1.504742
#> 5  4.107886  5.116225 10.99464 0.10574856      NaN 1.457675
#> 6 10.235321 20.034277 38.54326 0.32184247      NaN 1.389229