Retrieve Drug Molecules in MOL Format from the DrugBank Database

Retrieve Drug Molecules in MOL Format from the DrugBank Database

getMolFromDrugBank(id, parallel = 5)

Arguments

id
A character vector, as the DrugBank drug ID.

parallel
An integer, the parallel parameter, indicates how many process the user would like to use for retrieving the data (using RCurl), default is 5. For regular cases, we recommend a number less than 20.

Value

A length of id character vector, each element containing the corresponding drug molecule.

Details

This function retrieves drug molecules in MOL format from the DrugBank database.

See also

Examples

id = 'DB00859'  # Penicillamine

getMolFromDrugBank(id)

#> [1] "\n  Mrv0541 08131209382D          \n\n 10  9  0  0  0  0            999 V2000\n    0.8250   -0.8250    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    1.6500   -0.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    0.8250    0.8250    0.0000 C   0  0  1  0  0  0  0  0  0  0  0  0\n    1.5395    1.2375    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    0.8250    0.0000    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n    0.1105    1.2375    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n    1.5395    2.0625    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    2.2539    0.8250    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    0.0000    0.0000    0.0000 S   0  0  0  0  0  0  0  0  0  0  0  0\n    1.5395    0.4125    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n  4  3  1  0  0  0  0\n  5  1  1  0  0  0  0\n  5  2  1  0  0  0  0\n  5  3  1  0  0  0  0\n  3  6  1  6  0  0  0\n  7  4  2  0  0  0  0\n  8  4  1  0  0  0  0\n  9  5  1  0  0  0  0\n  3 10  1  6  0  0  0\nM  END\n> <DATABASE_ID>\nDB00859\n\n> <DATABASE_NAME>\ndrugbank\n\n> <SMILES>\n[H][C@](N)(C(O)=O)C(C)(C)S\n\n> <INCHI_IDENTIFIER>\nInChI=1S/C5H11NO2S/c1-5(2,9)3(6)4(7)8/h3,9H,6H2,1-2H3,(H,7,8)/t3-/m0/s1\n\n> <INCHI_KEY>\nVVNCNSJFMMFHPL-VKHMYHEASA-N\n\n> <FORMULA>\nC5H11NO2S\n\n> <MOLECULAR_WEIGHT>\n149.211\n\n> <EXACT_MASS>\n149.051049291\n\n> <JCHEM_ACCEPTOR_COUNT>\n3\n\n> <JCHEM_AVERAGE_POLARIZABILITY>\n14.761607659503875\n\n> <JCHEM_BIOAVAILABILITY>\n1\n\n> <JCHEM_DONOR_COUNT>\n3\n\n> <JCHEM_FORMAL_CHARGE>\n0\n\n> <JCHEM_GHOSE_FILTER>\n0\n\n> <JCHEM_IUPAC>\n(2S)-2-amino-3-methyl-3-sulfanylbutanoic acid\n\n> <ALOGPS_LOGP>\n-1.70\n\n> <JCHEM_LOGP>\n-2.146518265659755\n\n> <ALOGPS_LOGS>\n-1.51\n\n> <JCHEM_MDDR_LIKE_RULE>\n0\n\n> <JCHEM_NUMBER_OF_RINGS>\n0\n\n> <JCHEM_PHYSIOLOGICAL_CHARGE>\n0\n\n> <JCHEM_PKA>\n10.076558230275829\n\n> <JCHEM_PKA_STRONGEST_ACIDIC>\n2.555552931139977\n\n> <JCHEM_PKA_STRONGEST_BASIC>\n9.093679244005854\n\n> <JCHEM_POLAR_SURFACE_AREA>\n63.32000000000001\n\n> <JCHEM_REFRACTIVITY>\n37.22500000000001\n\n> <JCHEM_ROTATABLE_BOND_COUNT>\n2\n\n> <JCHEM_RULE_OF_FIVE>\n1\n\n> <ALOGPS_SOLUBILITY>\n4.65e+00 g/l\n\n> <JCHEM_TRADITIONAL_IUPAC>\ndepen\n\n> <JCHEM_VEBER_RULE>\n0\n\n> <DRUGBANK_ID>\nDB00859\n\n> <SECONDARY_ACCESSION_NUMBERS>\nAPRD01171\n\n> <DRUG_GROUPS>\napproved\n\n> <GENERIC_NAME>\nPenicillamine\n\n> <SYNONYMS>\n(−)-penicillamine; (S)-2-amino-3-mercapto-3-methylbutanoic acid; (S)-3,3-dimethylcysteine; 3-mercapto-D-valine; D-(−)-penicillamine; D-penicillamine; D-β,β-dimethylcysteine; penicilamina; Penicillamine\n\n> <PRODUCTS>\nCuprimine; Depen; Depen Tab 250mg; Penicillamine\n\n> <INTERNATIONAL_BRANDS>\nAtamir\n\n$$$$\n"