Retrieve Drug Molecules in MOL Format from the PubChem Database

Retrieve Drug Molecules in MOL Format from the PubChem Database

Usage

getMolFromPubChem(id, parallel = 5)

Arguments

id

A character vector, as the PubChem drug ID.

parallel

An integer, the parallel parameter, indicates how many processes the user would like to use for retrieving the data (using RCurl), default is 5. For regular cases, we recommend a number less than 20.

Value

A length of id character vector, each element containing the corresponding drug molecule.

Details

This function retrieves drug molecules in MOL format from the PubChem database.

See also

Examples

id = c('7847562', '7847563')  # Penicillamine
# \donttest{
getMolFromPubChem(id)# }
#> [1] "7847562\n  -OEChem-06272300272D\n\n  9  8  0     1  0  0  0  0  0999 V2000\n   14.1653  -22.6051    0.0000 S   0  0  0  0  0  0  0  0  0  0  0  0\n   10.5700  -20.5100    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n    9.3576  -22.6100    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   11.7825  -24.0098    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n   12.9779  -21.9196    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   11.7824  -22.6100    0.0000 C   0  0  3  0  0  0  0  0  0  0  0  0\n   10.5700  -21.9100    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   12.9780  -20.5103    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   14.1904  -21.2196    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n  1  5  1  0  0  0  0\n  2  7  2  0  0  0  0\n  3  7  1  0  0  0  0\n  6  4  1  1  0  0  0\n  5  6  1  0  0  0  0\n  5  8  1  0  0  0  0\n  5  9  1  0  0  0  0\n  6  7  1  0  0  0  0\nM  END\n> <PUBCHEM_COMPOUND_ID_TYPE>\n0\n\n> <PUBCHEM_TOTAL_CHARGE>\n0\n\n> <PUBCHEM_SUBSTANCE_ID>\n7847562\n\n> <PUBCHEM_SUBSTANCE_VERSION>\n11\n\n> <PUBCHEM_EXT_DATASOURCE_NAME>\nKEGG\n\n> <PUBCHEM_EXT_DATASOURCE_REGID>\nD00496\n\n> <PUBCHEM_SUBSTANCE_COMMENT>\nSame as: <a href=\"http://pubchem.ncbi.nlm.nih.gov/summary/summary.cgi?sid=9622\">C07418</a>\nATC code: M01CC01\n\n> <PUBCHEM_SUBSTANCE_SYNONYM>\n52-67-5\nCuprimine (TN)\nD00496\nDepen (TN)\nPenicillamine (JAN/USP/INN)\n\n> <PUBCHEM_GENERIC_REGISTRY_NAME>\n52-67-5\n\n> <PUBCHEM_XREF_EXT_ID>\nD00496\n\n> <PUBCHEM_EXT_DATASOURCE_URL>\nhttp://www.genome.jp/kegg/\n\n> <PUBCHEM_EXT_SUBSTANCE_URL>\nhttp://www.genome.jp/dbget-bin/www_bget?dr:D00496\n\n> <PUBCHEM_CID_ASSOCIATIONS>\n5852  1\n\n> <PUBCHEM_COORDINATE_TYPE>\n1\n3\n\n> <PUBCHEM_BONDANNOTATIONS>\n6  4  5\n\n$$$$\n"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> [2] "7847563\n  -OEChem-06272300272D\n\n 26 27  0     1  0  0  0  0  0999 V2000\n   34.3547  -18.4020    0.0000 K   0  3  0  0  0  0  0  0  0  0  0  0\n   30.9856  -14.7946    0.0000 S   0  0  0  0  0  0  0  0  0  0  0  0\n   27.0539  -17.3222    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   32.8811  -18.3753    0.0000 O   0  5  0  0  0  0  0  0  0  0  0  0\n   30.6344  -19.4986    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   25.8603  -16.6202    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   23.4495  -15.2137    0.0000 O   0  0  0  0  0  0  0  0  0  0  0  0\n   29.6517  -16.6202    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n   27.0539  -14.5138    0.0000 N   0  0  0  0  0  0  0  0  0  0  0  0\n   29.6517  -15.2160    0.0000 C   0  0  3  0  0  0  0  0  0  0  0  0\n   28.2474  -15.2160    0.0000 C   0  0  3  0  0  0  0  0  0  0  0  0\n   30.9856  -17.0414    0.0000 C   0  0  3  0  0  0  0  0  0  0  0  0\n   28.2474  -16.6202    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   31.8280  -15.9180    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   31.4771  -18.3753    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   32.8111  -16.9009    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   32.8111  -14.9350    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   25.8603  -15.2160    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   24.6414  -14.5187    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   22.2365  -14.5133    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   22.2362  -13.0845    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   21.0247  -15.2132    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   21.0242  -12.3850    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   19.8127  -14.5138    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   19.8125  -13.0849    0.0000 C   0  0  0  0  0  0  0  0  0  0  0  0\n   29.6517  -13.8117    0.0000 H   0  0  0  0  0  0  0  0  0  0  0  0\n  2 10  1  0  0  0  0\n  2 14  1  0  0  0  0\n  3 13  2  0  0  0  0\n  4 15  1  0  0  0  0\n  5 15  2  0  0  0  0\n  6 18  2  0  0  0  0\n  7 19  1  0  0  0  0\n  7 20  1  0  0  0  0\n  8 10  1  0  0  0  0\n  8 12  1  0  0  0  0\n  8 13  1  0  0  0  0\n 11  9  1  1  0  0  0\n  9 18  1  0  0  0  0\n 10 11  1  0  0  0  0\n 10 26  1  6  0  0  0\n 11 13  1  0  0  0  0\n 12 14  1  0  0  0  0\n 12 15  1  6  0  0  0\n 14 16  1  0  0  0  0\n 14 17  1  0  0  0  0\n 18 19  1  0  0  0  0\n 20 21  2  0  0  0  0\n 20 22  1  0  0  0  0\n 21 23  1  0  0  0  0\n 22 24  2  0  0  0  0\n 23 25  2  0  0  0  0\n 24 25  1  0  0  0  0\nM  CHG  2   1   1   4  -1\nM  END\n> <PUBCHEM_COMPOUND_ID_TYPE>\n0\n\n> <PUBCHEM_TOTAL_CHARGE>\n0\n\n> <PUBCHEM_SUBSTANCE_ID>\n7847563\n\n> <PUBCHEM_SUBSTANCE_VERSION>\n9\n\n> <PUBCHEM_EXT_DATASOURCE_NAME>\nKEGG\n\n> <PUBCHEM_EXT_DATASOURCE_REGID>\nD00497\n\n> <PUBCHEM_SUBSTANCE_COMMENT>\nTarget: penicillin binding protein\nActivity: Antibacterial, Cell wall biosynthesis inhibitor\nATC code: J01CE02\n\n> <PUBCHEM_SUBSTANCE_SYNONYM>\n132-98-9\nD00497\nPenicillin V potassium (USP)\nPhenoxymethylpenicillin potassium (JAN)\nVeetids (TN)\n\n> <PUBCHEM_GENERIC_REGISTRY_NAME>\n132-98-9\n\n> <PUBCHEM_XREF_EXT_ID>\nD00497\n\n> <PUBCHEM_EXT_DATASOURCE_URL>\nhttp://www.genome.jp/kegg/\n\n> <PUBCHEM_EXT_SUBSTANCE_URL>\nhttp://www.genome.jp/dbget-bin/www_bget?dr:D00497\n\n> <PUBCHEM_CID_ASSOCIATIONS>\n23676814  1\n6869  2\n5462222  2\n\n> <PUBCHEM_COORDINATE_TYPE>\n1\n3\n\n> <PUBCHEM_BONDANNOTATIONS>\n10  26  6\n12  15  6\n11  9  5\n\n$$$$\n"