Parallelized Drug Molecule Similarity Search by Molecular Fingerprints Similarity or Maximum Common Substructure Search

Parallelized Drug Molecule Similarity Search by Molecular Fingerprints Similarity or Maximum Common Substructure Search

searchDrug(mol, moldb, cores = 2, method = c("fp", "mcs"),
  fptype = c("standard", "extended", "graph", "hybrid", "maccs",
  "estate", "pubchem", "kr", "shortestpath", "fp2", "fp3", "fp4",
  "obmaccs"), fpsim = c("tanimoto", "euclidean", "cosine", "dice",
  "hamming"), mcssim = c("tanimoto", "overlap"), ...)

Arguments

mol
The query molecule. The location of a sdf file containing one molecule.

moldb
The molecule database. The location of a sdf file containing all the molecules to be searched with.

cores
Integer. The number of CPU cores to use for parallel search, default is 2. Users could use the detectCores() function in the parallel package to see how many cores they could use.

method
'fp' or 'mcs'. Search by molecular fingerprints or by maximum common substructure searching.

fptype
The fingerprint type, only available when method = 'fp'. Rcpi supports 13 types of fingerprints, including 'standard', 'extended', 'graph', 'hybrid', 'maccs', 'estate', 'pubchem', 'kr', 'shortestpath', 'fp2', 'fp3', 'fp4', 'obmaccs'.

fpsim
Similarity measure type for fingerprint, only available when method = 'fp'. Including 'tanimoto', 'euclidean', 'cosine', 'dice' and 'hamming'. See calcDrugFPSim for details.

mcssim
Similarity measure type for maximum common substructure search, only available when method = 'mcs'. Including 'tanimoto' and 'overlap'.

...
Other possible parameter for maximum common substructure search, see calcDrugMCSSim for available options.

Value

Named numerical vector. With the decreasing similarity value of the molecules in the database.

Details

This function does compound similarity search derived by various molecular fingerprints with various similarity measures or derived by maximum common substructure search. This function runs for a query compound against a set of molecules.

Examples

mol = system.file('compseq/DB00530.sdf', package = 'Rcpi')
# DrugBank ID DB00530: Erlotinib
moldb = system.file('compseq/tyrphostin.sdf', package = 'Rcpi')
# Database composed by searching 'tyrphostin' in PubChem and filtered by Lipinski's Rule of Five

searchDrug(mol, moldb, cores = 4, method = 'fp', fptype = 'maccs', fpsim = 'hamming')

#>        66         9        14        13        73         2        12        26 
#> 0.3313253 0.3253012 0.3253012 0.3192771 0.3192771 0.3072289 0.3012048 0.2951807 
#>        67         5         8        15        32        78       103        25 
#> 0.2951807 0.2891566 0.2891566 0.2891566 0.2891566 0.2891566 0.2891566 0.2831325 
#>        72        51        68        88       105        27        85        86 
#> 0.2831325 0.2771084 0.2771084 0.2771084 0.2771084 0.2710843 0.2710843 0.2710843 
#>       104        34        50        65        70        81         6        17 
#> 0.2710843 0.2650602 0.2650602 0.2650602 0.2650602 0.2650602 0.2590361 0.2590361 
#>        52        55        64        75        21        30        31        40 
#> 0.2530120 0.2530120 0.2530120 0.2530120 0.2469880 0.2469880 0.2469880 0.2469880 
#>        41        48        59        60        69        89        91        93 
#> 0.2469880 0.2469880 0.2469880 0.2469880 0.2469880 0.2469880 0.2469880 0.2469880 
#>        19        33        42        61        62        87        97        11 
#> 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639 0.2409639 0.2349398 
#>        20        22        46        63        74        84        96        98 
#> 0.2349398 0.2349398 0.2349398 0.2349398 0.2349398 0.2349398 0.2349398 0.2349398 
#>        99         4         7        10        23        29        37        43 
#> 0.2349398 0.2289157 0.2289157 0.2289157 0.2289157 0.2289157 0.2289157 0.2289157 
#>        44        47        49        53        54        57        58        71 
#> 0.2289157 0.2289157 0.2289157 0.2289157 0.2289157 0.2289157 0.2289157 0.2289157 
#>        76        77        79        82         1         3        28        56 
#> 0.2289157 0.2289157 0.2289157 0.2289157 0.2228916 0.2228916 0.2228916 0.2228916 
#>        90        95        24        36        45        35        38        80 
#> 0.2228916 0.2228916 0.2168675 0.2168675 0.2168675 0.2108434 0.2108434 0.2108434 
#>        39       102        94        16        18        83       101        92 
#> 0.2048193 0.2048193 0.1927711 0.1686747 0.1686747 0.1686747 0.1385542 0.1204819 
#>       100 
#> 0.1204819 

searchDrug(mol, moldb, cores = 4, method = 'fp', fptype = 'fp2', fpsim = 'tanimoto')

#> Error in loadNamespace(name): there is no package called ‘ChemmineOB’

searchDrug(mol, moldb, cores = 4, method = 'mcs', mcssim = 'tanimoto')

#>        92       100        23        39        94        64        60        59 
#> 0.7000000 0.7000000 0.4000000 0.4000000 0.4000000 0.3783784 0.3636364 0.3529412 
#>        45        37        42        10        35        44        47        71 
#> 0.3513514 0.3421053 0.3421053 0.3333333 0.3333333 0.3333333 0.3333333 0.3333333 
#>        76        77        83       101        38        58        73         4 
#> 0.3333333 0.3333333 0.3255814 0.3243243 0.3170732 0.3055556 0.3055556 0.3000000 
#>        43        81        16        18        62        40        95        99 
#> 0.3000000 0.2941176 0.2894737 0.2894737 0.2777778 0.2666667 0.2647059 0.2647059 
#>         6        33        48        69        93        97        61        63 
#> 0.2571429 0.2571429 0.2571429 0.2571429 0.2571429 0.2571429 0.2500000 0.2500000 
#>        79        86        88        84        90        19        50         7 
#> 0.2500000 0.2500000 0.2500000 0.2352941 0.2352941 0.2325581 0.2307692 0.2250000 
#>        29        30        53        80        91        65        82        52 
#> 0.2250000 0.2250000 0.2250000 0.2250000 0.2250000 0.2222222 0.2222222 0.2195122 
#>        54        55        75        57       103         3        56         1 
#> 0.2195122 0.2195122 0.2195122 0.2162162 0.2162162 0.2142857 0.2142857 0.2127660 
#>        67        70        72        24        49        51        87         2 
#> 0.2121212 0.2121212 0.2105263 0.2093023 0.2093023 0.2093023 0.2058824 0.2051282 
#>        31       102       104        22        46        74        98        21 
#> 0.2051282 0.2051282 0.2051282 0.2045455 0.2045455 0.2045455 0.2045455 0.2000000 
#>        41        36        68       105        78        85        11        20 
#> 0.2000000 0.1956522 0.1951220 0.1951220 0.1944444 0.1944444 0.1914894 0.1914894 
#>        26        96         9        14        89         5         8        15 
#> 0.1914894 0.1914894 0.1904762 0.1904762 0.1904762 0.1891892 0.1891892 0.1891892 
#>        32        66        27        34        25        17        12        28 
#> 0.1891892 0.1891892 0.1851852 0.1794872 0.1764706 0.1750000 0.1666667 0.1666667 
#>        13 
#> 0.1333333